% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADDT.plot.FracFail.v.Time.R
\name{ADDT.plot.FracFail.v.Time}
\alias{ADDT.plot.FracFail.v.Time}
\title{Title}
\usage{
ADDT.plot.FracFail.v.Time(
  results.object,
  x.of.interest = NULL,
  FailLevel = NULL,
  plan.values.string = NULL,
  plan.string = NULL,
  quantile.range = c(0.01, 0.99),
  ylim = c(NA, NA),
  xlim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  my.title = NULL,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  grids = F,
  numplotsim = 50,
  nxpoints = 50,
  cex = 1,
  plot.distribution = "lognormal",
  how.time.range = "modified test plan"
)
}
\arguments{
\item{results.object}{}

\item{x.of.interest}{}

\item{FailLevel}{}

\item{plan.values.string}{}

\item{plan.string}{}

\item{quantile.range}{}

\item{ylim}{}

\item{xlim}{}

\item{xlab}{}

\item{ylab}{}

\item{my.title}{}

\item{title.option}{}

\item{grids}{}

\item{numplotsim}{}

\item{nxpoints}{}

\item{cex}{}

\item{plot.distribution}{}

\item{how.time.range}{}
}
\description{
Title
}
\examples{
\dontrun{

InsulationBrkdwn.ADDTplan <- get.allocation.matrix(list(DegreesC = c(180,225,250,275)),
                                                   times = c(1,2,4,8,16,32,48,64),
                                                   time.units = "Weeks",
                                                   reps = 4)

plot(InsulationBrkdwn.ADDTplan)

InsulationBrkdwn.ADDTpv <- get.ADDT.plan.values(distribution = "normal",
                                                transformation.x = "Arrhenius", 
                                                transformation.Response = "log", 
                                                transformation.time = "linear",
                                                beta0 = 2.58850162033243,
                                                beta1 = -476873415881.376,
                                                beta2 = 1.41806367703643,
                                                sigma = 0.172609,
                                                time.units = "Weeks",
                                                response.units = "Volts", 
                                                FailLevel = 10, 
                                                use.condition = 100)

print(InsulationBrkdwn.ADDTpv)

InsulationBrkdwn.vADDTplan <- hframe.to.vframe(InsulationBrkdwn.ADDTplan)
sum(allocation(InsulationBrkdwn.vADDTplan))

names(InsulationBrkdwn.ADDTpv)

InsulationBrkdwn.plan.sim.out <- sim.ADDT.test.plan(ADDT.test.plan = InsulationBrkdwn.ADDTplan, 
                                                    ADDT.plan.values = InsulationBrkdwn.ADDTpv, 
                                                    number.sim = 5)

ADDT.plot.time.v.x(InsulationBrkdwn.plan.sim.out)

ADDT.plot.Deg.v.Time(InsulationBrkdwn.plan.sim.out)
ADDT.plot.FracFail.v.Time(InsulationBrkdwn.plan.sim.out)

ADDT.vcv(ADDT.plan.values = InsulationBrkdwn.ADDTpv,
         ADDT.test.plan = hframe.to.vframe(InsulationBrkdwn.ADDTplan))


}
}
