% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF.R
\name{AF}
\alias{AF}
\title{Acceleration factors}
\usage{
AF(stress, stress0, coef, relationship, power)
}
\arguments{
\item{stress}{}

\item{stress0}{}

\item{coef}{}

\item{relationship}{}

\item{power}{}
}
\description{
Acceleration factors
}
\examples{
\dontrun{

AF(160,80,.8,"arrhenius")

AFplot(160,80,.8,"arrhenius")

AFplot(160,80,c(.7,.8,.9),"arrhenius")

AF(130,110,-2,"Inverse Power Rule")

AFplot(130,110,-2,"Inverse Power Rule")
}
}
