% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFplot.R
\name{AFplot}
\alias{AFplot}
\title{Title}
\usage{
AFplot(stress, stress0, coef, relationship, power, number.points = 100,
  grid = T, my.title = NULL, title.option = "full")
}
\arguments{
\item{stress}{}

\item{stress0}{}

\item{coef}{}

\item{relationship}{}

\item{power}{}

\item{number.points}{}

\item{grid}{}

\item{my.title}{}

\item{title.option}{}
}
\description{
Title
}
\examples{
\dontrun{

AF(160,80,.8,"arrhenius")

AFplot(160,80,.8,"arrhenius")

AFplot(160,80,c(.7,.8,.9),"arrhenius")

AF(130,110,-2,"Inverse Power Rule")

AFplot(130,110,-2,"Inverse Power Rule")
}
}
