% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alloya}
\alias{alloya}
\title{Fatigue crack growth in an unknown metallic alloy.}
\format{A \code{data.frame} with 262 rows and 3 variables
\tabular{rlll}{
  [, 1] \tab inches \tab Crack length observed at \code{megacycles} \tab \bold{Numeric}\cr
  [, 2] \tab specimen \tab Test specimen designator \tab \bold{Categoric}\cr 
  [, 3] \tab megacycles \tab Number of fatigue cycles (in millions) when \code{length} was inspected  \tab \bold{Numeric}
  }}
\source{
Hudak, S.J., Saxena, A., Bucci, R. J., and Malcom, R .C. (1978), 
        Development of standard methods of testing and analyzing fatigure crack growth rate data, 
        Technical Report AFML-TR-78-40, Westinghouse R & D Center, Westinghouse Electric Corporation, Pittsburgh, PA.

Bogdanoff, J. L. and Kozin, F. (1985), Probablistic Models of Cumulative Damage, New York, NY; John Wiley & Sons.
}
\description{
Twenty-one specimens of an unknown alloy were subjected 
             to the same cyclic load profile.  Prior to testing, an initial crack 
             (length = 0.9 in) was cut into each specimen to serve as a crack 
             nucleation site.  After each increment of 10,000 fatigue cycles the length
             of the crack was recorded.  A specimen was considered to have failed when
             the crack length exceeded 1.6 inches, otherwise the test concluded after 1.2 million 
             cycles.
}
