% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alloyc}
\alias{alloyc}
\title{Metal alloy tensile strength test}
\format{A \code{data.frame} with 9 rows and 4 variables:
\tabular{rlll}{
  [, 1] \tab ksil   \tab Lower limit of the stress interval \tab \bold{Numeric}\cr
  [, 2] \tab ksiu   \tab Upper limit of the stress interval \tab \bold{Numeric}\cr
  [, 3] \tab event  \tab Event observed in the interval (interval-censored)\tab \bold{Categoric}\cr
  [, 4] \tab count  \tab Number of events observed in the interval\tab \bold{Numeric}
  }}
\source{
Meeker W.Q. and Escobar L.A., Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons. \bold{276}
}
\description{
These data resulted from a test that was conducted to obtain information on the strength of an alloy produced by 
             an alternate process.  The designation for the alloy was not provided and is only referred to as "Alloy C".  Specimens
             of the alloy were subjected to various levels of constant stress, measured in ksi (1 ksi = 1000 psi).  The dataset 
             includes the number of failed specimens observed in each interval of applied stress.
}
