% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altplan.R
\name{altplan}
\alias{altplan}
\title{Title}
\usage{
altplan(alt.plan.values, plan.type, use.condition, highest.condition,
  censor.time, quantile = 0.1, xihold = 0.5, pifix = 0.2,
  pmlim = 0.2, kprint = 0, method = NULL, sample.size = 100)
}
\arguments{
\item{alt.plan.values}{}

\item{plan.type}{}

\item{use.condition}{}

\item{highest.condition}{}

\item{censor.time}{}

\item{quantile}{}

\item{xihold}{}

\item{pifix}{}

\item{pmlim}{}

\item{kprint}{}

\item{method}{}

\item{sample.size}{}
}
\description{
Title
}
\examples{
\dontrun{

AdhesiveBond.Weibull.altpv <-
  get.alt.plan.values.from.slope.and.point(distribution = "Weibull",
                                           relationship = "Arrhenius",
                                           accelvar.units = c("DegreesC"),
                                           time.units = "Days", 
                                           censor.time = 183,
                                           probs = c(.001), 
                                           accelvar = c(50),
                                           slope = 0.726, 
                                           beta = 1.667)

print(AdhesiveBond.Weibull.altpv)

evaluate(altplan(AdhesiveBond.Weibull.altpv,
                 "Optimum",
                 use.condition = 50, 
                 highest.condition = 120,
                 censor.time = 183, 
                 quantile = 0.1,
                 sample.size = 300), 
         AdhesiveBond.Weibull.altpv,
         use.condition = 50, 
         quantile.of.interest = 0.1)


AdhesiveBond.Weibull.optc.altplan <-
  altplan(AdhesiveBond.Weibull.altpv,
          "Optimized compromise",
          use.condition = 50,
          highest.condition = 120,
          censor.time = 250,
          quantile = 0.1)

evaluate(altplan(AdhesiveBond.Weibull.altpv,
                 "Equal expected",
                 use.condition = 50,
                 highest.condition = 120,
                 censor.time = 250,
                 quantile = 0.1),
         AdhesiveBond.Weibull.altpv,
         use.condition = 50,
         quantile.of.interest = 0.1)

AdhesiveBond.Weibull.421.altplan <-
  altplan(AdhesiveBond.Weibull.altpv,
          "421",
          use.condition = 50,
          highest.condition = 120,
          censor.time = 250,
          quantile = 0.1)

evaluate(altplan(AdhesiveBond.Weibull.altpv,
                 "Traditional", 
                 use.condition = 50,
                 highest.condition = 120,
                 censor.time = 250,
                 quantile = 0.1), 
         AdhesiveBond.Weibull.altpv,
         use.condition = 50, 
         quantile.of.interest = 0.1)

}
}
