% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altsim.R
\name{altsim}
\alias{altsim}
\title{Simulate Accelerated Life Test Data}
\usage{
altsim(accel.var.mat, nsamsz, centim, theta, distribution, number.sim,
  kctype = 1, escale = 10000, e = rep(1e-04, number.parameters),
  parameter.fixed = rep(F, number.parameters), intercept = T,
  kprint = 0, maxit = 500, debug1 = F, randomize = T)
}
\arguments{
\item{accel.var.mat}{}

\item{nsamsz}{}

\item{centim}{}

\item{theta}{}

\item{distribution}{}

\item{number.sim}{}

\item{kctype}{}

\item{escale}{}

\item{e}{}

\item{parameter.fixed}{}

\item{intercept}{}

\item{kprint}{}

\item{maxit}{}

\item{debug1}{}

\item{randomize}{}
}
\value{
List of plan values and simulated data
}
\description{
Simulates data from an accelerated life test
}
