% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ALTSIM}
\alias{ALTSIM}
\title{Parametric accelerated life test simulation}
\usage{
ALTSIM(x, y, cen, wt, nrow, nter, ny, nty, ty, tcodes, kdist, gamthr, lfix,
  nparm, intcpt, escale, e, maxit, kprint, dscrat, iscrat, devian, thetah,
  fsder, vcv, r, res, fv, theta, xnew, ynew, centim, acvar, nsubex, nacvar,
  nsamsz, krfail, kctype, retmat, numret, numsim, iersim)
}
\arguments{
\item{thetah}{(nparm) space for thetahat}

\item{theta}{(nparm) true value of theta}

\item{xnew}{hold space for x}

\item{ynew}{hold space for y}

\item{centim}{(nsubex) vector of censor times for alt}

\item{acvar}{(nsubex,nacvar) matrix of accelerating vars}

\item{nsubex}{number of subexperiments}

\item{nacvar}{number of accelerating variables}

\item{nsamsz}{(nsubex) vector giving number of units at each comb}

\item{kctype}{= 1 for type 1 at centim, = 2 for type 2 with centim failures}

\item{retmat}{(numret,numsim) matrix to return results, one result per col}

\item{numsim}{number of simulations to be run}

\item{iersim}{> 0 if data space too small}

\item{Accelerating}{Variable - X1, X1, Comb1, Comb2, Comb3}

\item{nrowr}{number of rows being returned in retmat}
}
\description{
Parametric accelerated life test simulation (spec censoring)
}
\details{
Parameter meanings same as in wqm_mlboth except those listed here
}
\note{
Modified 19 December 1997 to account for multiple accelerating variables
}
