% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altsimReturnFrame.R
\name{altsimReturnFrame}
\alias{altsimReturnFrame}
\title{Title}
\usage{
altsimReturnFrame(
  accel.var.mat,
  nsamsz,
  centim,
  theta,
  distribution,
  relationship,
  time.units = "Time",
  kctype = 1,
  escale = 10000,
  e = rep(1e-04, number.parameters),
  parameter.fixed = rep(F, number.parameters),
  intercept = T,
  kprint = 0,
  maxit = 500,
  debug1 = F,
  randomize = T,
  my.title = NULL
)
}
\arguments{
\item{accel.var.mat}{}

\item{nsamsz}{}

\item{centim}{}

\item{theta}{}

\item{distribution}{}

\item{relationship}{}

\item{time.units}{}

\item{kctype}{}

\item{escale}{}

\item{e}{}

\item{parameter.fixed}{}

\item{intercept}{}

\item{kprint}{}

\item{maxit}{}

\item{debug1}{}

\item{randomize}{}

\item{my.title}{}
}
\description{
Title
}
\examples{
\dontrun{

test.matrix <- cbind(V1 = c(1,1,2,2),V2 = c(1,2,1,2))

altsim.out <- altsim(accel.var.mat = test.matrix, 
                     nsamsz = c(4,4,4,4), 
                     centim = c(40000,40000,40000,40000), 
                     theta = c(1,-24,5,1), 
                     distribution = "normal",
                     number.sim = 10, 
                     debug = F)

altsim.out <- altsim(accel.var.mat = as.matrix(c(1,2,3,4)), 
                     nsamsz = c(4,4,4,4), 
                     centim = c(200,200,200,200), 
                     theta = c(1,-2,3), 
                     distribution = "normal", 
                     number.sim = 5, 
                     debug = F, 
                     kprint = 0)

altsimReturnFrame(accel.var.mat = test.matrix, 
                  nsamsz = c(4,4,4,4),
                  centim = c(100,100,100,100),
                  theta = c(1,1,1,2),
                  distribution =" lognormal", 
                  relationship = c("linear","log"))

## Simulate a single ALT data set

altsimReturnFrame(accel.var.mat = test.matrix, 
                  nsamsz = c(4,4,4,4),
                  centim = c(100,100,100,100), 
                  theta = c(1,1,1,2), 
                  distribution = "lognormal", 
                  relationship = c("linear","log"))

}
}
