% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1.plot.R
\name{ar1.plot}
\alias{ar1.plot}
\title{Title}
\usage{
ar1.plot(data.rdu, my.title = NULL, xlab = "Interarrival Times",
  ylab = "Lagged Interarrival Times", nacf = 5, plot.acf = F, ...)
}
\arguments{
\item{data.rdu}{}

\item{my.title}{}

\item{xlab}{}

\item{ylab}{}

\item{nacf}{}

\item{plot.acf}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")
ar1.plot(halfbeak.rdu,
         xlab = "Lagged Thousands of Hours Between Maintenance Actions",
         ylab = "Thousands of Hours Between Maintenance Actions")

ar1.plot(halfbeak.rdu,
         xlab = "Lagged Thousands of Hours Between Maintenance Actions",
         ylab = "Thousands of Hours Between Maintenance Actions",
         plot.acf = T)



}
}
