% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym.sample.size.R
\name{asym.sample.size}
\alias{asym.sample.size}
\title{Title}
\usage{
asym.sample.size(plan.values, quantile.of.interest, Rvalue = NULL,
  HalfWidth = NULL, censor.time, fraction.failing, conf.level = 0.95)
}
\arguments{
\item{plan.values}{}

\item{quantile.of.interest}{}

\item{Rvalue}{}

\item{HalfWidth}{}

\item{censor.time}{}

\item{fraction.failing}{}

\item{conf.level}{}
}
\description{
Title
}
\examples{
\dontrun{

plan.values3 <- get.plan.values("Weibull",
                                prob = c(.2,.12),
                                time = c(1000,500), 
                                time.units = "Hours")

asym.sample.size(plan.values3,
                 censor.time = 500,
                 Rvalue = 1.5,
                 quantile.of.interest = 0.1)

asym.sample.size(plan.values3,
                 fraction.failing = 0.1,
                 Rvalue = 1.5,
                 quantile.of.interest = 0.1)

asym.sample.size(bulb.plan.values1,
                 fraction.failing = 0.1,
                 HalfWidth = 50,
                 quantile.of.interest = 0.1)

}
}
