% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym.test.plan.properties.R
\name{asym.test.plan.properties}
\alias{asym.test.plan.properties}
\title{Title}
\usage{
asym.test.plan.properties(
  plan.values,
  n,
  proportion.failing,
  quantiles = c(0.1, 0.3, 0.5, 0.632)
)
}
\arguments{
\item{plan.values}{}

\item{n}{}

\item{proportion.failing}{}

\item{quantiles}{}
}
\description{
Title
}
\examples{
\dontrun{

 plan.values3 <- get.plan.values("Weibull",
                                 prob = c(.2,.12),
                                 time = c(1000,500), 
                                 time.units = "Hours")
                               
 asym.test.plan.properties(plan.values3, 
                           n = 50, 
                           proportion.failing = 0.1)

}
}
