% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{at7987}
\alias{at7987}
\title{Fatigue life test (alloy T7987)}
\format{
A \code{data.frame} with 68 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab kilocycles \tab Accumulated cycles at \code{event} (in thousands) \tab \bold{Numeric}\cr
  [, 2] \tab event \tab Event observed at \code{kilocycles} (failed/right-censored) \tab \bold{Categoric}\cr 
  [, 3] \tab count \tab Number of events observed at \code{kilocycles} \tab \bold{Numeric}
  }
}
\source{
Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons.
}
\description{
A test was conducted to assess the fatigue life of an alloy, 
             designated as T-7987. In this test, 72 specimens were subjected 
             to an unreported load spectrum. The number of cycles at which a 
             specimen failed was recorded (rounded to the nearest thousand cycles).
             Of the 72 units tested, 67 units failures were observed and 5 units were right 
             censored at 300 kilocycles.
}
