% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.gmleprobplot.R
\name{basic.gmleprobplot}
\alias{basic.gmleprobplot}
\title{Title}
\usage{
basic.gmleprobplot(
  data.ld,
  distribution,
  plot.dist,
  xlab = get.time.units(data.ld),
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  time.range = c(NA, NA),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  interactive = T,
  original.par = F,
  mle.quantiles = T,
  my.title = NULL,
  type = "y",
  cex = 1,
  cex.points = 1.2,
  cexlab = 1,
  sub.title = "",
  grids = F,
  linear.axes = F,
  slope.axis = F,
  print.table = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  trunc.correct = T,
  add = F,
  plot.censored.ticks = F,
  pch = 16,
  lty = c(3, 4, 5, 6, 7),
  lwd = c(2, 2, 2, 2, 2),
  length.time.vec = 100,
  compare.dists = NULL,
  col.ci = 4,
  lwd.ci = 2,
  xxx.mle.out = NULL,
  ciMethod = "normal.approx",
  extrapolate.ci = F,
  band.type = "p",
  ...
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{plot.dist}{}

\item{xlab}{}

\item{xlim}{}

\item{ylim}{}

\item{time.range}{}

\item{conf.level}{}

\item{interactive}{}

\item{original.par}{}

\item{mle.quantiles}{}

\item{my.title}{}

\item{type}{}

\item{cex}{}

\item{cex.points}{}

\item{cexlab}{}

\item{sub.title}{}

\item{grids}{}

\item{linear.axes}{}

\item{slope.axis}{}

\item{print.table}{}

\item{title.option}{}

\item{ylab}{}

\item{trunc.correct}{}

\item{add}{}

\item{plot.censored.ticks}{}

\item{pch}{}

\item{lty}{}

\item{lwd}{}

\item{length.time.vec}{}

\item{compare.dists}{}

\item{col.ci}{}

\item{lwd.ci}{}

\item{xxx.mle.out}{}

\item{ciMethod}{}

\item{extrapolate.ci}{}

\item{band.type}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

AlloyC.ld <- frame.to.ld(alloyc,
                         response.column = c(1,2),
                         censor.column = 3,
                         case.weight.column = 4,
                         data.title = "Alloy C",
                         time.units = "ksi")

AlloyC.sevgets.gmle.out <- gets.mle(AlloyC.ld, distribution = "sev")

AlloyC.norgets.gmle.out <- gets.mle(AlloyC.ld, distribution = "normal")

basic.gmleprobplot(AlloyC.ld, distribution = "sevgets",
                   xxx.mle.out = AlloyC.sevgets.gmle.out,
                   my.title = "",
                   cexlab = 1.5,
                   plot.dist = "sev")

AlloyC.gets.nor.gmle.out <- gets.mle(AlloyC.ld, distribution = "normal")

basic.gmleprobplot(AlloyC.ld,
                   distribution = "normalgets",
                   xxx.mle.out = AlloyC.norgets.gmle.out, 
                   my.title = "", 
                   cexlab = 1.5,
                   plot.dist = "sev")

}
}
