% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bearingcage}
\alias{bearingcage}
\title{Large system bearing cage fracture test}
\format{A \code{data.frame} with 25 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab hours \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 2] \tab event \tab Event observed at \code{hours} (failure/right-censored/left-censored/interval-censored)  \tab \bold{Categoric}\cr
  [, 3] \tab count \tab Number of events observed at \code{hours} \tab \bold{Numeric}
   }}
\source{
Abernethy, R. B., Breneman, J. E., Medlin, C. H., and Reinman, G. L. (1983) 
        Weibull Analysis Handbook, Air Force Wright Aeronautical Laboratories Technical Report AFWAL-TR-83-2079
}
\description{
Ball bearing assembly failure data

The service life requirement for a ball bearing assembly was specified 
         such that \eqn{t_{0.1}}{t[0.1]} (aka the B10 life) be greater that 8000 hours. 
         Analysts were concerned that the design of the bearing cage in this assembly 
         was inadequate and could lead to premature failures during service.
         Service times were collected for 1703 assemblies that were introduced into 
         service over time.  The analysts wanted to use the service life data to 
         determine if a redesign was needed to ensure that the units could meet the 
         service life requirement.  Management was also interested in determining the 
         number of additional failures that could be expected over the next year
         for the population of assemblies already in service.
}
\details{
Bearing cages are used in ball bearing assemblies to ensure that the 
         ball bearings do not drift out of position relative to the other ball 
         bearings during use.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{berkson20}},
  \code{\link{bleed}}, \code{\link{ceramicbearing}},
  \code{\link{cirpack6}}, \code{\link{comptime}},
  \code{\link{computerlab}}, \code{\link{cylinder}},
  \code{\link{devicea}}, \code{\link{deviceb}},
  \code{\link{devicec}}, \code{\link{deviceg}},
  \code{\link{fan}}, \code{\link{gaaslaser}},
  \code{\link{grampus}}, \code{\link{grids1}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
