% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-bisa.R
\name{dbisa}
\alias{dbisa}
\alias{pbisa}
\alias{qbisa}
\title{The Birnbaum Sanders Distribution}
\usage{
dbisa(x, shape, scale = 1)

pbisa(q, shape, scale = 1)

qbisa(p, shape, scale = 1)
}
\description{
The Birnbaum Sanders Distribution
}
