% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisa.mle.R
\name{bisa.mle}
\alias{bisa.mle}
\title{Title}
\usage{
bisa.mle(data.ld, debug1 = F, theta.start = NULL)
}
\arguments{
\item{data.ld}{}

\item{debug1}{}

\item{theta.start}{}
}
\description{
Title
}
\examples{
\dontrun{

at7987.ld <- frame.to.ld(at7987, 
                         response.column = 1,
                         censor.column = 2, 
                         case.weight.column = 3,
                         time.units = "Kilocycles")

summary(at7987.ld)

at7987.bisa.gmle.out <- bisa.mle(at7987.ld)
}
}
