% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bkfatigue10}
\alias{bkfatigue10}
\title{Yokobori fatigue-fracture test}
\format{
A \code{data.frame} with 63 rows and 1 variable:
\tabular{rlll}{
  [, 1] \tab kilocycles \tab Accumulated cycles at failure (in thousands) \tab \bold{Numeric}
   }
}
\source{
Bogdanoff, J.,L. and Kozin, F. (1985) Probablistic Models of Cumulative Damage, pp. 224-225, 
        New York, NY; John Wiley & Sons.

Yokobori, T. (1951) Fatigue fracture in steel, Journal of the Physical Society of Japan, \bold{6}, 81-86.
}
\description{
Yokobori describes a fatigue-fracture test on 0.41\% carbon steel cylindrical specimens, tested at 
             \eqn{\pm 37.1 kg/mm^2} stress amplitude.
}
