% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bleed}
\alias{bleed}
\title{Bleed system reliability data}
\format{A \code{data.frame} with 60 rows and 4 variables:
\tabular{rlll}{
  [, 1] \tab hours \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 2] \tab event \tab Event observed at \code{hours} (failure/right-censored/left-censored/interval-censored)  \tab \bold{Categoric}\cr
  [, 3] \tab count \tab Number of events observed at \code{hours} \tab \bold{Numeric}\cr
  [, 4] \tab base  \tab Base where observation was made \tab \bold{Categoric}
   }}
\source{
Abernethy, R. B., Breneman, J. E., Medlin, C. H., and Reinman, G. L., (1983) 
        Weibull Analysis Handbook, Air Force Wright Aeronautical Laboratories Technical Report AFWAL-TR-83-2079
}
\description{
Abernethy, Breneman, Medlin, and Reinman give failure and 
             running time for 2256 bleed systems operating at several 
             geographically separated bases. They observed a change in 
             the probability plot before and after 600 of operation.  
             Further examination showed that 9 of the 19 failures occurred 
             at Base 'D'.  Separate analyses of the Base D data and the data 
             from the other bases indicated different life distributions. 
             The large slope for Base D indicated strong wearout behavior. 
             The relatively small slope for the other bases suggested infant 
             mortality or accidental failures. After investigation it was 
             determined that the early-failure problem at base D was caused by 
             salt air (Base D was near the ocean). A change in maintenance 
             procedures there solved the dominant bleed system reliability problem.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{ceramicbearing}},
  \code{\link{cirpack6}}, \code{\link{comptime}},
  \code{\link{computerlab}}, \code{\link{cylinder}},
  \code{\link{devicea}}, \code{\link{deviceb}},
  \code{\link{devicec}}, \code{\link{deviceg}},
  \code{\link{fan}}, \code{\link{gaaslaser}},
  \code{\link{grampus}}, \code{\link{grids1}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
