% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.pivot.R
\name{boot.pivot}
\alias{boot.pivot}
\title{Bootstrap based prediction intervals}
\usage{
boot.pivot(B = 100, N = 20, mu = 5.13, sigma = 0.161, t_c = NULL,
  fails = NULL, alpha = 0.05)
}
\arguments{
\item{B}{The number of bootstrap cycles}

\item{N}{The size of the original data sample}

\item{mu}{ML estimate for the mean of the original data sample}

\item{sigma}{ML estimate for the standard deviation of the original data sample}

\item{t_c}{Censoring time for Type-I censored tests}

\item{fails}{Number of failures for Type-II censored tests}

\item{alpha}{Desired significance level}
}
\description{
Bootstrap based prediction intervals
}
