% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ceramicbearing}
\alias{ceramicbearing}
\title{Rolling contact fatigue of ceramic ball bearings}
\format{A \code{data.frame} with 69 rows and 5 variables:
\tabular{rlll}{
  [, 1] \tab kilocycles \tab Accumulated cycles at \code{event} (in thousands) \tab \bold{Numeric}\cr
  [, 2] \tab mode \tab Failure mode observed at \code{kilocycles} \tab \bold{Categoric}\cr
  [, 3] \tab event \tab Event observed at \code{kilocycles} (failure/right-censored/left-censored/interval-censored)  \tab \bold{Categoric}\cr
  [, 4] \tab count \tab Number of events observed at \code{kilocycles} \tab \bold{Numeric}\cr
  [, 5] \tab kilonewtons \tab Applied stress (in thousands) \tab \bold{Numeric}
   }}
\source{
McCool, J. I. (1980), 
        Confidence limits for Weibull regression with censored data, 
        IEEE Transactions on Reliability, \bold{R-29}, 145-150.
}
\description{
McCool gives the results of a life test 
             on rolling contact fatigue of ceramic 
             ball bearings. Ten specimens were tested
             at each of 4 levels of stress.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{cirpack6}}, \code{\link{comptime}},
  \code{\link{computerlab}}, \code{\link{cylinder}},
  \code{\link{devicea}}, \code{\link{deviceb}},
  \code{\link{devicec}}, \code{\link{deviceg}},
  \code{\link{fan}}, \code{\link{gaaslaser}},
  \code{\link{grampus}}, \code{\link{grids1}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
