% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cirpack6}
\alias{cirpack6}
\title{Circuit pack field tracking study}
\format{A \code{data.frame} with 27 rows and 7 variables:
\tabular{rlll}{
  [, 1] \tab lower    \tab Start of an observation interval \tab \bold{Numeric}\cr
  [, 2] \tab upper    \tab End of an observation interval  \tab \bold{Numeric}\cr
  [, 3] \tab event    \tab Event observed in the interval (failure/right-censored/left-censored/interval-censored) \tab \bold{Categoric}\cr
  [, 4] \tab truntype \tab Truncation type (left-truncated/right-truncated/interval-truncated) \tab \bold{Categoric}\cr
  [, 5] \tab truntime \tab Truncation time \tab \bold{Numeric}\cr
  [, 6] \tab count    \tab Number of events observed in the interval \tab \bold{Numeric}\cr
  [, 7] \tab vendor   \tab Producing vendor \tab \bold{Categoric}
  }}
\source{
Meeker, W. Q. and Escobar, L. A. (1998), Statistical Methods for Reliability Data, New York, NY; Wiley-Interscience
}
\description{
Circuit packs were manufactured to the same design 
             specification, but by two different vendors. The 
             trial ran for 10,000 hours to determine which vendor's
             circuit packs were more reliable.  The 4993 circuit packs 
             from Vendor 1 came straight from production.  
             The 4993 circuit packs from Vendor 2 had already seen 1000 
             hours of burn-in testing at the manufacturing plant under 
             operating conditions similar to those in the field trial. 
             The circuit packs manufactured by Vendor 2 were sold at a 
             higher price because field reliability was supposed to have 
             been improved by the burn-in screening of circuit packs 
             containing defective components. Failures during the first 
             1000 hours of burn-in were not recorded. This is the reason 
             for the unknown entries in the table and for having information 
             out to 11,000 hours for Vendor 2.  These data are for the first 
             failure in a position; information on circuit packs replaced 
             after initial failure in a position was not part of the study.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{comptime}},
  \code{\link{computerlab}}, \code{\link{cylinder}},
  \code{\link{devicea}}, \code{\link{deviceb}},
  \code{\link{devicec}}, \code{\link{deviceg}},
  \code{\link{fan}}, \code{\link{gaaslaser}},
  \code{\link{grampus}}, \code{\link{grids1}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
