% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.many.exponential.profiles.R
\name{compare.many.exponential.profiles}
\alias{compare.many.exponential.profiles}
\title{Title}
\usage{
compare.many.exponential.profiles(theta = 5, sample.size = 10,
  number.simulation = 10, profile.title = NULL)
}
\arguments{
\item{theta}{}

\item{sample.size}{}

\item{number.simulation}{}

\item{profile.title}{}
}
\description{
Title
}
\examples{
\dontrun{

compare.many.exponential.profiles(theta = 5,
                                  sample.size = 3,
                                  number.simulation = 10)

compare.many.exponential.profiles(theta =5,
                                  sample.size = 1000,
                                  number.simulation = 10)
}
}
