% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.mleprobplot.R
\name{compare.mleprobplot}
\alias{compare.mleprobplot}
\title{Title}
\usage{
\method{compare}{mleprobplot}(data.ld, main.distribution,
  compare.distribution, gamthr = 0, xlim = c(NA, NA), ylim = c(NA,
  NA), time.range = c(NA, NA), my.title = NULL, sub.title = NULL,
  grids = F, linear.axes = F, slope.axis = F, band.type = "none",
  xlab = get.time.units(data.ld),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100, cex = 1.2,
  title.option = GetSMRDDefault("SMRD.TitleOption"), trunc.correct = T,
  do.legend = T, plot.censored.ticks = F)
}
\arguments{
\item{data.ld}{}

\item{main.distribution}{}

\item{compare.distribution}{}

\item{gamthr}{}

\item{xlim}{}

\item{ylim}{}

\item{time.range}{}

\item{my.title}{}

\item{sub.title}{}

\item{grids}{}

\item{linear.axes}{}

\item{slope.axis}{}

\item{band.type}{}

\item{xlab}{}

\item{conf.level}{}

\item{cex}{}

\item{title.option}{}

\item{trunc.correct}{}

\item{do.legend}{}

\item{plot.censored.ticks}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1, 
                                censor.column = 3,
                                time.units = "Kilometers")

mleprobplot(ShockAbsorber.ld, distribution = "Weibull")
mleprobplot(ShockAbsorber.ld, distribution = "loglogistic")
mleprobplot(ShockAbsorber.ld, distribution = "lognormal")
mleprobplot(ShockAbsorber.ld, distribution = "frechet")


#compare lognormal and Weibull distributions

compare.mleprobplot(ShockAbsorber.ld, 
                    main.distribution = "Lognormal",
                    compare.distribution ="Weibull")
                    
compare.mleprobplot(ShockAbsorber.ld, main.distribution= "Lognormal",compare.distribution=c("Weibull","Loglogistic"))

compare.mleprobplot(ShockAbsorber.ld, 
                    main.distribution = "Lognormal",
                    xlim = c(100,100000),
                    ylim = c(.005,.9),
                    compare.distribution = c("Weibull","Loglogistic"))

compare.mleprobplot(ShockAbsorber.ld, 
                    main.distribution = "Lognormal",
                    xlim = c(100,50001),
                    ylim = c(.005,.9),
                    compare.distribution = c("Weibull","Loglogistic","Exponential"),
                    band.type = "chull")

compare.mleprobplot(ShockAbsorber.ld, 
                    main.distribution = "Lognormal",
                    xlim = c(100,100000),
                    ylim = c(.005,.9),
                    compare.distribution = c("Weibull","Exponential"),
                    band.type = "chull")

}
}
