% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.effect.R
\name{component.effect}
\alias{component.effect}
\title{Title}
\usage{
component.effect(
  parallel = TRUE,
  dependent = FALSE,
  mu = logb(10),
  xlim = c(0, 1),
  ylim = c(0, 1),
  sd = 1,
  rho = c(0, 0.4, 0.7, 0.9, 1)
)
}
\arguments{
\item{parallel}{}

\item{dependent}{}

\item{mu}{}

\item{xlim}{}

\item{ylim}{}

\item{sd}{}

\item{rho}{}
}
\description{
Title
}
\examples{
\dontrun{
component.effect(parallel = TRUE)

parallel.effect()

parallel.dep.effect()

series.dep.effect()

}
}
