% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondProbInterval.R
\name{CondProbInterval}
\alias{CondProbInterval}
\title{Title}
\usage{
CondProbInterval(mu, sigma, distribution, age, tL, tU)
}
\arguments{
\item{mu}{}

\item{sigma}{}

\item{distribution}{}

\item{age}{}

\item{tL}{}

\item{tU}{}
}
\description{
Title
}
\examples{
\dontrun{

BearingCage.ld <- frame.to.ld(bearingcage,
                              response.column = 1, 
                              censor.column = 2,
                              case.weight.column = 3)

BearingCage.mlest.weib <- mlest(BearingCage.ld,
                                dist = "Weibull")

CondProbInterval2(BearingCage.mlest.weib,
                  age = 50,
                  tL = 50,
                  tU = 350)

CondProbInterval2(BearingCage.mlest.weib,
                  age = 150,
                  tL = 150,
                  tU = 450)

# Table 12.1 from Meeker and Escobar (1998)

PredictTable(BearingCage.mlest.weib,
             FtimeStart = 0, 
             FtimeEnd = 300)

CondProbInterval(mu = 3.7393, 
                 sigma = 0.7639,
                 distribution = "lognormal",
                 age = 1,
                 tL = 1,
                 tU = 36)

}
}
