% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondProbInterval2.R
\name{CondProbInterval2}
\alias{CondProbInterval2}
\title{Title}
\usage{
CondProbInterval2(mlest.out, age, tL, tU)
}
\arguments{
\item{mlest.out}{}

\item{age}{}

\item{tL}{}

\item{tU}{}
}
\description{
Title
}
\examples{
\dontrun{

DeviceN.ld <- frame.to.ld(devicen,
                          response.column = "months", 
                          censor.column = "event",
                          case.weight.column = "counts")
DeviceN.mlest.lnorm <- mlest(DeviceN.ld,
                             dist = "Lognormal")

CondProbInterval2(DeviceN.mlest.lnorm,
                  age = 1,
                  tL = 1,
                  tU = 36)
CondProbInterval2(DeviceN.mlest.lnorm,
                  age = 2,
                  tL = 2,
                  tU = 36)
CondProbInterval2(DeviceN.mlest.lnorm,
                  age = 1,
                  tL = 1,
                  tU = 2)
CondProbInterval2(DeviceN.mlest.lnorm,
                  age = 2,
                  tL = 2,
                  tU = 3)
}
}
