% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{connectionstrength}
\alias{connectionstrength}
\title{#####}
\format{A \code{data.frame} with 14 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab strength \tab Strength (in ???) of the connection \tab \bold{Numeric}\cr
  [, 2] \tab mode \tab Failure mode observed at \code{strength} \tab \bold{Categoric}\cr
  [, 3] \tab count \tab Number of failures observed at \code{strength}\tab \bold{Numeric}
   }}
\source{
Nelson W. (1984), Applied Life Data Analysis, New York: John Wiley & Sons, pg. 111.
}
\description{
####
}
\seealso{
Other data-notdone: \code{\link{adhesivebondc}},
  \code{\link{adhesivestrength}},
  \code{\link{amsaaexactfail}}, \code{\link{appliancea}},
  \code{\link{applianceb}}, \code{\link{aprel72}},
  \code{\link{atrazinejune}}, \code{\link{bearinga}},
  \code{\link{censortruncationtest}},
  \code{\link{chemicalprocess}}, \code{\link{cirpack4}},
  \code{\link{cirpack5}}, \code{\link{componentd}},
  \code{\link{concrete}}, \code{\link{customerlife}},
  \code{\link{devicedconnected}}, \code{\link{deviceh}},
  \code{\link{devicen}}, \code{\link{diskber}},
  \code{\link{doatrun}}, \code{\link{electromech}},
  \code{\link{engineemissions}},
  \code{\link{epoxyweathering}}, \code{\link{filament}},
  \code{\link{hcfdata}}, \code{\link{heater}},
  \code{\link{hpcrepairs}}, \code{\link{icdevice1}},
  \code{\link{inconel}}, \code{\link{insulation}},
  \code{\link{laminatepanel}}, \code{\link{largeball}},
  \code{\link{locomotivecontrol}},
  \code{\link{luminosity}}, \code{\link{machineh}},
  \code{\link{mechanicalswitch}},
  \code{\link{multiplefail}}, \code{\link{newspring}},
  \code{\link{parta}}, \code{\link{piccioto}},
  \code{\link{pipelinethickness}}, \code{\link{pmax}},
  \code{\link{prob3_5}}, \code{\link{riverchem}},
  \code{\link{shelflifea}}, \code{\link{testdadtplan}},
  \code{\link{titanium2}}, \code{\link{titanium}},
  \code{\link{tractorbreaks}}, \code{\link{turbinedevice}},
  \code{\link{vehiclemotor}},
  \code{\link{voltageendurance}}, \code{\link{workstation}}
}
\concept{data-notdone}
