% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumulativePredictTable.R
\name{CumulativePredictTable}
\alias{CumulativePredictTable}
\title{Title}
\usage{
CumulativePredictTable(mlest.out, time.increment, number.time.units.ahead,
  warranty.time)
}
\arguments{
\item{mlest.out}{}

\item{time.increment}{}

\item{number.time.units.ahead}{}

\item{warranty.time}{}
}
\description{
Title
}
\examples{
\dontrun{

# No warranty enforcement

DeviceN.NoEnforce.weib <-
  CumulativePredictTable(DeviceN.mlest.weib,
                         time.increment = 1,
                         number.time.units.ahead = 50,
                         warranty.time = 1000)

PlotCumulativePredictTable(DeviceN.NoEnforce.weib,
                           plot.what = "density")

PlotCumulativePredictTable(DeviceN.NoEnforce.weib,
                           plot.what = "cumulative")

# Strict warranty enforcement

DeviceN.Enforced.weib <-
  CumulativePredictTable(DeviceN.mlest.weib,
                         time.increment = 1,
                         number.time.units.ahead = 50,
                         warranty.time = 36)

PlotCumulativePredictTable(DeviceN.Enforced.weib,
                           plot.what = "density")

PlotCumulativePredictTable(DeviceN.Enforced.weib,
                           plot.what = "cumulative")

}
}
