% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.plot.R
\name{design.plot}
\alias{design.plot}
\title{Title}
\usage{
design.plot(data.ld, x.cols = c(1, 2),
  xlab = get.x.columns(data.ld)[x.cols[1]],
  ylab = get.x.columns(data.ld)[x.cols[2]], pch = 16, cex = 1.5)
}
\arguments{
\item{data.ld}{}

\item{x.cols}{}

\item{xlab}{}

\item{ylab}{}

\item{pch}{}

\item{cex}{}
}
\description{
Title
}
\examples{
\dontrun{

Tantalum.ld <- frame.to.ld(tantalum, 
                           response.column = 1,
                           censor.column = 2, 
                           case.weight.column = 3, 
                           x.column = c(4, 5), 
                           data.title = "Tantalum Capacitor Data", 
                           time.units = "Hours", 
                           xlabel = c("Volts","DegreesC"))

summary(Tantalum.ld)

design.plot(Tantalum.ld)

}
}
