% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dest.degrad.mle.R
\name{dest.degrad.mle}
\alias{dest.degrad.mle}
\title{Title}
\usage{
dest.degrad.mle(data.ddd, distribution, transformation.response,
  transformation.x, transformation.time,
  group.var = 1:ncol(xmat(data.ddd)), debug1 = map.SMRDDebugLevel() >=
  4, theta.start = NULL, power = NULL)
}
\arguments{
\item{data.ddd}{}

\item{distribution}{}

\item{transformation.response}{}

\item{transformation.x}{}

\item{transformation.time}{}

\item{group.var}{}

\item{debug1}{}

\item{theta.start}{}

\item{power}{}
}
\description{
Title
}
\examples{
\dontrun{

AdhesiveStrength.ddd <- frame.to.ddd(adhesivestrength, 
                                     response.column = "pounds",
                                     time.column = "days",
                                     x.columns = "celsius",
                                     data.title = "AdhesiveStrength Strength Data",
                                     time.units = "Days")

AdhesiveStrength.gmle <- dest.degrad.mle(AdhesiveStrength.ddd,
                                         distribution = "normal",
                                         transformation.response = "log", 
                                         transformation.x = "arrhenius", 
                                         transformation.time = "linear")

get.sub.model.dest.degrad.mle.out(AdhesiveStrength.gmle,c(50,60,70))

}
}
