% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deviceb}
\alias{deviceb}
\title{Power output degradation in integrated circuit devices}
\format{
A \code{data.frame} with 570 rows and 4 variables:
\tabular{rlll}{
  [, 1] \tab powerdrop \tab Degradation (measured in decibels dB) of a device under test \tab \bold{Numeric}\cr
  [, 2] \tab device    \tab Label for the device under test \tab \bold{Categoric}\cr
  [, 3] \tab hours     \tab Time at which the degradation measure was observed \tab \bold{Numeric}\cr
  [, 4] \tab celsius   \tab Temperature applied \tab \bold{Numeric}
   }
}
\source{
Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York: John Wiley & Sons.
}
\description{
Samples of integrated circuit devices called "Device B" were tested at each of three levels of junction temperature. 
             The purpose of the test was to provide design engineers with an assessment of the proportion of these devices that 
             would "fail" before 15 years (about 130 thousand hours) of operation at 80 degrees celsius. Failure for an individual 
             device was defined as when the measured power output dropped more than 0.5 decibels (dB) below initial output. At 
             standard the operating temperature (80 degrees celsius), the devices will degrade too slowly to provide useful 
             information in 6 months.  Because units at low temperature degrade more slowly, they had to be run for longer periods 
             of time to accumulate appreciable degradation.  Because of severe limitations in the number of test positions, fewer 
             units were run at lower temperatures. 
             
             Note: The original data from this experiment are proprietary. The observations in this dataset were actually simulated 
             from a model suggested by physical theory and limited real data that was available at the time when the more complete 
             experiment was being planned.
}
