% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{devicec}
\alias{devicec}
\title{Integrated circuit accelerated life test data}
\format{A \code{data.frame} with 26 rows and 4 variables:
\tabular{rlll}{
  [, 1] \tab celsius   \tab Temperature to which a device was exposed during the test \tab \bold{Numeric}\cr
  [, 2] \tab kilohours \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 3] \tab event     \tab Event observed at \code{hours} (failure/right-censored)  \tab \bold{Categoric}\cr
  [, 4] \tab count     \tab Number of events observed at \code{hours} \tab \bold{Numeric}
   }}
\source{
Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons.
}
\description{
Observations from an accelerated life test performed on an integrated circuit device called "Device C". Failures 
             were caused by a chemical reaction inside the circuit package. Reliability engineers tested 10 circuits at five
             different temperatures over a period of 3000 hours. The purpose of the experiment was to estimate the activation 
             energy of the failure-causing reaction and to obtain an estimate of the integrated circuit life distribution at 
             an 80 degrees celsius junction temperature.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{deviceg}},
  \code{\link{fan}}, \code{\link{gaaslaser}},
  \code{\link{grampus}}, \code{\link{grids1}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
