% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{devicedconnected}
\alias{devicedconnected}
\alias{devicednotconnected}
\title{#####}
\format{
A \code{data.frame} with 986 rows and 9 variables:
\tabular{rlll}{
  [, 1] \tab weeks \tab Failure time \tab \bold{Numeric}\cr
  [, 2] \tab cycles \tab Total accumulated cycles at failure \tab \bold{Numeric}\cr
  [, 3] \tab causeoffailure.fm1 \tab Type of event observed (failure, right-censored, left-censored, or interval-censored)  \tab \bold{Categoric}\cr
  [, 4] \tab causeoffailure.fm2 \tab Start time (in hours) of an inspection interval \tab \bold{Numeric}\cr
  [, 5] \tab causeoffailure.fm3 \tab Time (in hours) at which an interval ends \tab \bold{Numeric}\cr
  [, 6] \tab causeoffailure.fmother \tab Type of event observed (failure, right-censored, left-censored, or interval-censored)  \tab \bold{Categoric}\cr
  [, 7] \tab perweek \tab Average number of cycles (across all users) \tab \bold{Numeric}\cr
  [, 8] \tab weeksinserted \tab Time inserted into service \tab \bold{Numeric}\cr
  [, 9] \tab weeksreturned \tab Time return to manufacturer \tab \bold{Numeric}
   }

A \code{data.frame} with 986 rows and 9 variables:
\tabular{rlll}{
  [, 1] \tab weeks \tab Failure time \tab \bold{Numeric}\cr
  [, 2] \tab cycles \tab Total accumulated cycles at failure \tab \bold{Numeric}\cr
  [, 3] \tab causeoffailure.fm1 \tab Type of event observed (failure, right-censored, left-censored, or interval-censored)  \tab \bold{Categoric}\cr
  [, 4] \tab causeoffailure.fm2 \tab Start time (in hours) of an inspection interval \tab \bold{Numeric}\cr
  [, 5] \tab causeoffailure.fm3 \tab Time (in hours) at which an interval ends \tab \bold{Numeric}\cr
  [, 6] \tab causeoffailure.fmother \tab Type of event observed (failure, right-censored, left-censored, or interval-censored)  \tab \bold{Categoric}\cr
  [, 7] \tab perweek \tab Average number of cycles (across all users) \tab \bold{Numeric}\cr
  [, 8] \tab weeksinserted \tab Time inserted into service \tab \bold{Numeric}\cr
  [, 9] \tab weeksreturned \tab Time return to manufacturer \tab \bold{Numeric}
   }
}
\source{
####

####
}
\description{
####

####
}
\seealso{
Other data-notdone: 
\code{\link{adhesivebondc}},
\code{\link{adhesivestrength}},
\code{\link{amsaaexactfail}},
\code{\link{appliancea}},
\code{\link{applianceb}},
\code{\link{aprel72}},
\code{\link{atrazinejune}},
\code{\link{bearinga}},
\code{\link{censortruncationtest}},
\code{\link{chemicalprocess}},
\code{\link{cirpack4}},
\code{\link{cirpack5}},
\code{\link{componentd}},
\code{\link{concrete}},
\code{\link{connectionstrength}},
\code{\link{customerlife}},
\code{\link{deviceh}},
\code{\link{devicen}},
\code{\link{diskber}},
\code{\link{doatrun}},
\code{\link{electromech}},
\code{\link{engineemissions}},
\code{\link{epoxyweathering}},
\code{\link{filament}},
\code{\link{hcfdata}},
\code{\link{heater}},
\code{\link{hpcrepairs}},
\code{\link{icdevice1}},
\code{\link{inconel}},
\code{\link{insulation}},
\code{\link{laminatepanel}},
\code{\link{largeball}},
\code{\link{locomotivecontrol}},
\code{\link{luminosity}},
\code{\link{machineh}},
\code{\link{mechanicalswitch}},
\code{\link{multiplefail}},
\code{\link{newspring}},
\code{\link{parta}},
\code{\link{piccioto}},
\code{\link{pipelinethickness}},
\code{\link{pmax}()},
\code{\link{prob3_5}},
\code{\link{riverchem}},
\code{\link{shelflifea}},
\code{\link{testdadtplan}},
\code{\link{titanium2}},
\code{\link{titanium}},
\code{\link{tractorbreaks}},
\code{\link{turbinedevice}},
\code{\link{vehiclemotor}},
\code{\link{voltageendurance}},
\code{\link{workstation}}

Other data-notdone: 
\code{\link{adhesivebondc}},
\code{\link{adhesivestrength}},
\code{\link{amsaaexactfail}},
\code{\link{appliancea}},
\code{\link{applianceb}},
\code{\link{aprel72}},
\code{\link{atrazinejune}},
\code{\link{bearinga}},
\code{\link{censortruncationtest}},
\code{\link{chemicalprocess}},
\code{\link{cirpack4}},
\code{\link{cirpack5}},
\code{\link{componentd}},
\code{\link{concrete}},
\code{\link{connectionstrength}},
\code{\link{customerlife}},
\code{\link{deviceh}},
\code{\link{devicen}},
\code{\link{diskber}},
\code{\link{doatrun}},
\code{\link{electromech}},
\code{\link{engineemissions}},
\code{\link{epoxyweathering}},
\code{\link{filament}},
\code{\link{hcfdata}},
\code{\link{heater}},
\code{\link{hpcrepairs}},
\code{\link{icdevice1}},
\code{\link{inconel}},
\code{\link{insulation}},
\code{\link{laminatepanel}},
\code{\link{largeball}},
\code{\link{locomotivecontrol}},
\code{\link{luminosity}},
\code{\link{machineh}},
\code{\link{mechanicalswitch}},
\code{\link{multiplefail}},
\code{\link{newspring}},
\code{\link{parta}},
\code{\link{piccioto}},
\code{\link{pipelinethickness}},
\code{\link{pmax}()},
\code{\link{prob3_5}},
\code{\link{riverchem}},
\code{\link{shelflifea}},
\code{\link{testdadtplan}},
\code{\link{titanium2}},
\code{\link{titanium}},
\code{\link{tractorbreaks}},
\code{\link{turbinedevice}},
\code{\link{vehiclemotor}},
\code{\link{voltageendurance}},
\code{\link{workstation}}
}
\concept{data-notdone}
