% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.plot2.R
\name{distribution.plot}
\alias{distribution.plot}
\title{Create a four-panel distribution plot}
\usage{
distribution.plot(distribution, shape, prob.range = c(0.01, 0.99),
  number.points = 1000, scale = rep(1, length(shape)),
  location = rep(0, length(shape)), shape2 = 1, exponential2 = T,
  cex = 16, lwd = 1.5, ...)
}
\arguments{
\item{distribution}{Generates an array of ggplot2 grobs showing
the cdf, pdf, survival function and hazard 
function for a distribution}

\item{shape}{A vector of shape parameters}

\item{prob.range}{A vector (length 2) of values between 0 and 1 
providing the lower and upper limits of the 
probability range}

\item{number.points}{The number points to be used in the plot}

\item{scale}{A numeric value of the scale parameter (see Details)}

\item{location}{A numeric value of the location parameter (see Details)}

\item{shape2}{A numeric value of the second shape parameter (see Details)}

\item{exponential2}{Is this 2-param exponential}

\item{cex}{A positive numeric value giving the amount by which the
plot text should be magnified relative to the default.}

\item{lwd}{A positive numeric value giving the amount by which the
line widths should be magnified relative to the default.}

\item{...}{Currently not used}
}
\description{
Create a four-panel distribution plot
}
\examples{
\dontrun{
distribution.plot("Exponential",
                  shape = c(.5,1))

distribution.plot("Lognormal",
                  shape = c(.3, .8))

distribution.plot("Normal",
                  shape = c( .30, .5,.8),
                  location = 5)

distribution.plot("Weibull",
                  shape = c(.8,1,1.5))

distribution.plot("Smallest Extreme Value",
                  shape = c(5,6,7),
                  location = 50)

distribution.plot("Largest Extreme Value",
                  shape = c(5, 6, 7),
                  location = 10)

distribution.plot("Logistic",
                  shape = c(1, 2, 3),
                  location = 15)

distribution.plot("Loglogistic",
                  shape = c(.2,.4,.6), 
                  prob.range = c(0.001, 0.95))
}
}
