% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.profile.plot.R
\name{do.profile.plot}
\alias{do.profile.plot}
\title{Title}
\usage{
do.profile.plot(
  gmle.out,
  parameter,
  original.parameter,
  quantile,
  failure.probability.at
)
}
\arguments{
\item{gmle.out}{}

\item{parameter}{}

\item{original.parameter}{}

\item{quantile}{}

\item{failure.probability.at}{}
}
\description{
Title
}
\examples{
\dontrun{

BearingCage.ld <- frame.to.ld(bearingcage,
                              response.column = 1, 
                              censor.column = 2, 
                              case.weight.column = 3,
                              time.units = "Hours")

summary(BearingCage.ld)

BearingCage.weibull.gmle.out <- ls.mle(BearingCage.ld,
                                       distribution = "Weibull")

names(BearingCage.weibull.gmle.out)

do.profile.plot(BearingCage.weibull.gmle.out,
                parameter = "log(scale)")

do.profile.plot(BearingCage.weibull.gmle.out,
                original.parameter = "scale")

do.profile.plot(BearingCage.weibull.gmle.out,
                quantile = 0.1)

do.profile.plot(BearingCage.weibull.gmle.out,
                failure.probability = 5000)

}
}
