% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echapter.R
\name{echapter}
\alias{echapter}
\title{Render an enchapter help document}
\usage{
echapter(chapter)
}
\arguments{
\item{chapter}{\code{numeric} Which chapter to run. Should be a value between 1 and 24}
}
\description{
Render an enchapter help document
}
