% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expon.mle.R
\name{expon.mle}
\alias{expon.mle}
\title{Title}
\usage{
expon.mle(data.ld, debug1 = F, theta.start = NULL)
}
\arguments{
\item{data.ld}{}

\item{debug1}{}

\item{theta.start}{}
}
\description{
Title
}
\examples{
\dontrun{

berkson200.ld <- frame.to.ld(berkson200,
                             response.column = c(1,2),
                             censor.column = 3,
                             case.weight.column = 4,
                             time.units = "1/5000 Seconds")

summary(berkson200.ld)

plot(berkson200.ld)
plot(berkson200.ld, dist = "Exponential")

expon.mle(berkson200.ld)
}
}
