% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-egeng.R
\name{degengl}
\alias{degengl}
\alias{pegeng}
\alias{qegengl}
\title{The Extended Generalized Gamma Distribution}
\usage{
degengl(x, mu, sigma, delta)

pegeng(q, mu, sigma, delta, distribution = NULL)

qegengl(p, mu, sigma, delta, distribution = NULL, smalldelta = 1e-04)
}
\description{
The Extended Generalized Gamma Distribution
}
