% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.ADDT.stableparam.R
\name{f.ADDT.stableparam}
\alias{f.ADDT.stableparam}
\title{Title}
\usage{
f.ADDT.stableparam(beta, model)
}
\arguments{
\item{beta}{}

\item{model}{}
}
\description{
Title
}
\examples{
\dontrun{

AdhesiveBondC.ADDTpv <-
  get.ADDT.plan.values(distribution = "sev",
                       transformation.x = c("Arrhenius","Humidity"),
                       transformation.response = "log",
                       transformation.time = "Square root",
                       beta0 = 2.168,
                       the.slope = -0.00709030595,
                       slope.at = c(30,50),
                       beta2 = c(0.6666,.2),
                       sigma = 0.1807,
                       time.units = "Weeks",
                       response.units = "Pounds",
                       FailLevel = 2,
                       use.condition = "30;50")

AdhesiveBondC.ADDTplan <-
  get.allocation.matrix(list(DegreesC = c(40,50,60),RH = c(20,80)),
                        times = c(1,2,5,10,20,50),
                        time.units = "Weeks",
                        reps = 6)

tmp.pmodel <- 
  pseudo.model((ADDT.plan.values = AdhesiveBondC.ADDTpv,
               ADDT.test.plan = hframe.to.vframe(AdhesiveBondC.ADDTplan))

f.ADDT.stableparam(AdhesiveBondC.ADDTpv$theta.vec, 
                   tmp.pmodel)

f.ADDT.origparam(f.ADDT.stableparam(AdhesiveBondC.ADDTpv$theta.vec, tmp.pmodel), tmp.pmodel)

}
}
