% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filament}
\alias{filament}
\title{Wire filament test design data}
\format{
A \code{data.frame} with 6 rows and 7 variables:
\tabular{rlll}{
  [, 1] \tab celsius \tab Temperature applied \tab \bold{Numeric}\cr
  [, 2] \tab volts \tab Electrical power applied \tab \bold{Numeric}\cr
  [, 3] \tab weeks1 \tab Number of units tested for one week \tab \bold{Categoric}\cr
  [, 4] \tab weeks2 \tab Number of units tested for two weeks \tab \bold{Categoric}\cr
  [, 5] \tab weeks5 \tab Number of units tested for five weeks \tab \bold{Categoric}\cr
  [, 6] \tab weeks10 \tab Number of units tested for ten weeks \tab \bold{Categoric}\cr
  [, 7] \tab weeks20 \tab Number of units tested for twenty week \tab \bold{Categoric}
   }
}
\source{
####
}
\description{
####
}
\seealso{
Other data-notdone: 
\code{\link{adhesivebondc}},
\code{\link{adhesivestrength}},
\code{\link{amsaaexactfail}},
\code{\link{appliancea}},
\code{\link{applianceb}},
\code{\link{aprel72}},
\code{\link{atrazinejune}},
\code{\link{bearinga}},
\code{\link{censortruncationtest}},
\code{\link{chemicalprocess}},
\code{\link{cirpack4}},
\code{\link{cirpack5}},
\code{\link{componentd}},
\code{\link{concrete}},
\code{\link{connectionstrength}},
\code{\link{customerlife}},
\code{\link{devicedconnected}},
\code{\link{deviceh}},
\code{\link{devicen}},
\code{\link{diskber}},
\code{\link{doatrun}},
\code{\link{electromech}},
\code{\link{engineemissions}},
\code{\link{epoxyweathering}},
\code{\link{hcfdata}},
\code{\link{heater}},
\code{\link{hpcrepairs}},
\code{\link{icdevice1}},
\code{\link{inconel}},
\code{\link{insulation}},
\code{\link{laminatepanel}},
\code{\link{largeball}},
\code{\link{locomotivecontrol}},
\code{\link{luminosity}},
\code{\link{machineh}},
\code{\link{mechanicalswitch}},
\code{\link{multiplefail}},
\code{\link{newspring}},
\code{\link{parta}},
\code{\link{piccioto}},
\code{\link{pipelinethickness}},
\code{\link{pmax}()},
\code{\link{prob3_5}},
\code{\link{riverchem}},
\code{\link{shelflifea}},
\code{\link{testdadtplan}},
\code{\link{titanium2}},
\code{\link{titanium}},
\code{\link{tractorbreaks}},
\code{\link{turbinedevice}},
\code{\link{vehiclemotor}},
\code{\link{voltageendurance}},
\code{\link{workstation}}
}
\concept{data-notdone}
