% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.power.and.loglin.process.R
\name{fit.power.and.loglin.process}
\alias{fit.power.and.loglin.process}
\title{Title}
\usage{
fit.power.and.loglin.process(
  data.rdu,
  start.gamma1 = 0.001,
  xlab = paste("Age in", get.time.units(data.rdu)),
  ylab = "Mean Cumulative Failures",
  plot.seg = T
)
}
\arguments{
\item{data.rdu}{}

\item{start.gamma1}{}

\item{xlab}{}

\item{ylab}{}

\item{plot.seg}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")

fit.power.and.loglin.process(halfbeak.rdu,
                             xlab = "Thousands of Hours of Operation",
                             ylab = "Cumulative Number of Maintenance Actions")


}
}
