% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.power.process.R
\name{fit.power.process}
\alias{fit.power.process}
\title{Title}
\usage{
fit.power.process(data.ld)
}
\arguments{
\item{data.ld}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")

summary(halfbeak.rdu)
event.plot(halfbeak.rdu)
print(mcf(halfbeak.rdu))
mcf.plot(halfbeak.rdu)
interarrival.times(halfbeak.rdu)
mcf.plot(halfbeak.rdu,
         xlab = "Thousands of Hours of Operation",
         ylab = "Cumulative Number of Maintenance Actions")

fit.power.process(halfbeak.rdu)
fit.loglin.process(halfbeak.rdu)
fit.power.and.loglin.process(halfbeak.rdu)
                             

}
}
