% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four.groupi.mleprobplot.R
\name{four.groupi.mleprobplot}
\alias{four.groupi.mleprobplot}
\title{Title}
\usage{
four.groupi.mleprobplot(
  data.ld,
  stresses = get.x.markers(data.ld, group.var = group.var),
  group.var = 1,
  xlab = get.time.units(data.ld),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  time.vector = c(NA, NA),
  dump = 1,
  grids = F,
  my.title = NULL,
  cex = 1.1,
  linear.axes = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  pch = (1:(length(stresses) + 1))[-2],
  lty = 1:length(stresses),
  ci.list = NULL,
  lwd = rep(1, length(stresses)),
  plot.censored.ticks = F,
  trunc.correct = T,
  col.fhat.vec = 1:length(stresses),
  col.ci = 6,
  shape = NULL,
  landscape = F,
  ...
)
}
\arguments{
\item{data.ld}{}

\item{stresses}{}

\item{group.var}{}

\item{xlab}{}

\item{ylab}{}

\item{xlim}{}

\item{ylim}{}

\item{time.vector}{}

\item{dump}{}

\item{grids}{}

\item{my.title}{}

\item{cex}{}

\item{linear.axes}{}

\item{title.option}{}

\item{pch}{}

\item{lty}{}

\item{ci.list}{}

\item{lwd}{}

\item{plot.censored.ticks}{}

\item{trunc.correct}{}

\item{col.fhat.vec}{}

\item{col.ci}{}

\item{shape}{}

\item{landscape}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{
DeviceA.ld <- frame.to.ld(devicea, 
                          data.title = "Device-A ALT Results",
                          response.column = 1,
                          time.units = "Hours",
                          censor.column = 2,
                          case.weight.column = 3,
                          x.columns = 4, 
                          xlab = "Degrees C")

print(DeviceA.ld)
summary(DeviceA.ld)

censored.data.plot(DeviceA.ld)

censored.data.plot(DeviceA.ld, 
                   y.axis ="log", 
                   x.axis = "Arrhenius")

groupi.mleprobplot(DeviceA.ld, 
                   distribution = "Weibull")

four.groupi.mleprobplot(DeviceA.ld)

DeviceA.weib.groupi <- groupi.mleprobplot(DeviceA.ld,
                                          distribution = "Weibull")

print(DeviceA.weib.groupi)
summary(DeviceA.weib.groupi)

}
}
