% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four.groupm.mleprobplot.R
\name{four.groupm.mleprobplot}
\alias{four.groupm.mleprobplot}
\title{Title}
\usage{
four.groupm.mleprobplot(
  data.ld,
  formula = NULL,
  stresses = get.x.markers(data.ld, group.var = group.var),
  group.var = 1,
  xlab = get.time.units(data.ld),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  relationship = NULL,
  power = NULL,
  my.title = NULL,
  dump = 0,
  mle.intervals = F,
  cex = 1.1,
  grids = F,
  linear.axes = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  lwd = 2,
  plot.censored.ticks = F,
  time.range = c(NA, NA),
  shape = NULL,
  ci.list = NULL,
  col.ci = 6,
  printem = F,
  trunc.correct = T,
  new.data = NULL,
  plotem = T,
  stresses.limit = 12,
  landscape = F,
  ...
)
}
\arguments{
\item{data.ld}{}

\item{formula}{}

\item{stresses}{}

\item{group.var}{}

\item{xlab}{}

\item{ylab}{}

\item{conf.level}{}

\item{xlim}{}

\item{ylim}{}

\item{relationship}{}

\item{power}{}

\item{my.title}{}

\item{dump}{}

\item{mle.intervals}{}

\item{cex}{}

\item{grids}{}

\item{linear.axes}{}

\item{title.option}{}

\item{lwd}{}

\item{plot.censored.ticks}{}

\item{time.range}{}

\item{shape}{}

\item{ci.list}{}

\item{col.ci}{}

\item{printem}{}

\item{trunc.correct}{}

\item{new.data}{}

\item{plotem}{}

\item{stresses.limit}{}

\item{landscape}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{
DeviceA.ld <- frame.to.ld(devicea, 
                          data.title = "Device-A ALT Results",
                          response.column = 1,
                          time.units = "Hours",
                          censor.column = 2,
                          case.weight.column = 3,
                          x.columns = 4, 
                          xlab = "Degrees C")

print(DeviceA.ld)
summary(DeviceA.ld)

censored.data.plot(DeviceA.ld)

censored.data.plot(DeviceA.ld, 
                   y.axis ="log", 
                   x.axis = "Arrhenius")

groupi.mleprobplot(DeviceA.ld, 
                   distribution = "Weibull")

four.groupi.mleprobplot(DeviceA.ld)

DeviceA.weib.groupi <- groupi.mleprobplot(DeviceA.ld,
                                          distribution = "Weibull")

print(DeviceA.weib.groupi)
summary(DeviceA.weib.groupi)


DeviceA.lognor.groupi <- groupi.mleprobplot(DeviceA.ld, 
                                            distribution = "Lognormal")

summary(DeviceA.lognor.groupi)

failure.probabilities(DeviceA.lognor.groupi)
quantiles(DeviceA.lognor.groupi)

four.groupm.mleprobplot(DeviceA.ld, 
                        relationship = "Arrhenius")

}
}
