% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fr.conf.R
\name{Fr.conf}
\alias{Fr.conf}
\title{Title}
\usage{
Fr.conf(fcn, fcn.arg2, gmle.out, extrapolate = F, ptwise.if.simult = F,
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100, ...)
}
\arguments{
\item{fcn}{}

\item{fcn.arg2}{}

\item{gmle.out}{}

\item{extrapolate}{}

\item{ptwise.if.simult}{}

\item{conf.level}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

Fan.ld <- frame.to.ld(fan,
                      response.column = 1, 
                      censor.column = 2, 
                      case.weight.column = 3,
                      time.units = "Hours")

Fan.egeng.gmle.out <- FillRegion(Fan.egeng.gmle.out,
                                 nbound = 10,
                                 iter = 500)

summary(Fan.egeng.gmle.out.jcr)

names(Fan.egeng.gmle.out.jcr)

basic.gmleprobplot(Fan.ld,distribution = "egeng",
                   xlim = c(200,99999),
                   ylim = c(.0011,.69),
                   xxx.mle.out = Fan.egeng.gmle.out,
                   my.title = "",
                   cexlab = 1.5,
                   conlev = .95,
                   ciMethod = "lr.approx",
                   length.time.vec = 2)

Fan.weibull.gmle.out <- ls2.mle(Fan.ld, distribution = "weibull")


Fan.weibull.gmle.out <-  FillRegion(Fan.weibull.gmle.out,
                                    nbound = 4,
                                    iter = 50,
                                    cull = 2 )

fcn = function(theta,time,distribution){
  
      f.phibf((log(time)-theta[1]) / exp(theta[2]),
                      distribution = "Weibull")
  
}

Fr.conf(fcn, 
        fcn.arg2 = log.seq(200,2000,length=10),
        gmle.out = Fan.weibull.gmle.out,
        ptwise = T)

Fr.conf(fcn,
        fcn.arg2 = log.seq(200,2000,length = 10),
        gmle.out = Fan.weibull.gmle.out,  
        ptwise = T, 
        extrapolate = T)

}
}
