% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fr.profile.R
\name{Fr.profile}
\alias{Fr.profile}
\title{Title}
\usage{
Fr.profile(
  fcn,
  gmle.out,
  nlevels = 10,
  lowest.conf.level = 0.1,
  xlab = as.character(substitute(fcn)),
  distribution = gmle.out$model$distribution,
  subtitle = " ",
  ...
)
}
\arguments{
\item{fcn}{}

\item{gmle.out}{}

\item{nlevels}{}

\item{lowest.conf.level}{}

\item{xlab}{}

\item{distribution}{}

\item{subtitle}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

BearingCage.ld <- frame.to.ld(bearingcage,
                              response.column = 1, 
                              censor.column = 2, 
                              case.weight.column = 3,
                              time.units = "Hours")

summary(BearingCage.ld)

BearingCage.weibull.gmle.out <- ls.mle(BearingCage.ld,
                                       distribution = "Weibull")

names(BearingCage.weibull.gmle.out)

BearingCage.weibull.gmle.out.jcr <- FillRegion(BearingCage.weibull.gmle.out,
                                               nbound = 4,
                                               iter = 500,
                                               cull = 2)

summary(BearingCage.weibull.gmle.out.jcr)

mleprobplot(BearingCage.ld,
            distribution = "Weibull",
            xlim = c(200,10000),
            ylim = c(.00031,.19),
            time.vec = log.seq(200,10000,20))

basic.gmleprobplot(BearingCage.ld,
                   distribution = "Weibull",
                   xxx.mle.out = BearingCage.weibull.gmle.out,
                   my.title = "",
                   cexlab = 1.5,
                   conf.level = 0.95,
                   length.time.vec = 20,
                   xlim = c(200,10000),
                   ylim = c(.00031,.19), 
                   time.vec = log.seq(200,10000,20),
                   ciMethod = "lr.approx")

basic.gmleprobplot(BearingCage.ld,
                   distribution = "Weibull",
                   xxx.mle.out = BearingCage.weibull.gmle.out,
                   my.title = "",
                   cexlab = 1.5,
                   conf.level = 0.95,
                   length.time.vec = 20, 
                   xlim = c(200,10000),
                   ylim = c(.00031,.19), 
                   time.vec = log.seq(200,10000,20))

profile.plot(Fr.profile(fcn <- function(x){x[1]},
                        BearingCage.weibull.gmle.out))

tmppro <- Fr.profile(fcn <- function(x){x[1]}, 
                     BearingCage.weibull.gmle.out)


}
}
