% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.to.ld.R
\name{frame.to.ld}
\alias{frame.to.ld}
\title{Create a \code{life.data} object from a \code{data.frame}}
\usage{
frame.to.ld(frame, response.column, censor.column, case.weight.column,
  failure.mode.column, truncation.type.column, truncation.response.column,
  time.units = names.the.frame[response.column[1]], x.columns = NULL,
  xlabel = NULL, data.title = deparse(substitute(frame)),
  data.note = "", func.call = match.call(), residual.rmd = NULL)
}
\arguments{
\item{frame}{A \code{data.frame} class object.}

\item{response.column}{The numeric index or name (as a character string) of the column(s) in \code{frame} containing the responses.

For reliability data, responses are often the amount of usage (measured in time, distance, cycles) that are recorded when an event occurred.

Events for which responses should be recorded include when one or more units fail or are censored. If the responses are recorded as an interval (i.e. an event occurs between a and b), \code{response.column} will accept vector arguments.}

\item{censor.column}{The numeric index or name (as a character string) of the column in \code{frame} that contains the type(s) of censoring (exact failures, right-censoring, left-censoring, interval censoring) occurring at each event.

In practice, several different labels are used to refer to differnt types of censoring.  \code{frame.to.ld} can accept many of these different labels.  To view the list of see \code{SMRDOptionsDefaults}.}

\item{case.weight.column}{The numeric index or name (as a character string) of the column in \code{frame} containing the number of events occurring at each entry in the \code{response.column}.
\bold{(optional)} This column may be left undefined, in which case it will be assumed that only one unit was observed to have failed/censored for each entry in \code{response.column}.}

\item{failure.mode.column}{The numeric index or name (as a character string) of the column in \code{frame} containing the type(s) of failure modes observed.
\bold{(optional)} This column may be left undefined, in which case it will be assumed that only one failure mode was observed}

\item{truncation.type.column}{}

\item{truncation.response.column}{The time at which an obsevation is trucated.}

\item{time.units}{A character string denoting the unit of measure used to quantify system lifetime.
\code{time.units} should be provided for each \code{life.data} object as this string will be automatically substituted in many SMRD plots.}

\item{x.columns}{}

\item{xlabel}{}

\item{data.note}{}

\item{func.call}{}

\item{residual.rmd}{}
}
\description{
Create a \code{life.data} object from a \code{data.frame}
}
\seealso{
\code{\link{frame.to.rmd}}, \code{\link{frame.to.ddd}}, \code{\link{surv.to.ld}}, \code{\link{SMRDOptionsDefaults}}
}
\author{
William Q. Meeker, PhD
}
