% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.to.rdu.R
\name{frame.to.rdu}
\alias{frame.to.rdu}
\title{Title}
\usage{
frame.to.rdu(frame, time.column, ID.column, cost.count.column,
  event.column, time.units = names.the.frame[time.column[1]],
  failure.censor.names = GetSMRDDefault("SMRD.FailName"),
  right.censor.names = GetSMRDDefault("SMRD.RcName"),
  left.censor.names = GetSMRDDefault("SMRD.LcName"),
  interval.censor.names = GetSMRDDefault("SMRD.IcName"), data.title,
  data.note = "", skip = 0)
}
\arguments{
\item{frame}{}

\item{time.column}{}

\item{ID.column}{}

\item{cost.count.column}{}

\item{event.column}{}

\item{time.units}{}

\item{failure.censor.names}{}

\item{right.censor.names}{}

\item{left.censor.names}{}

\item{interval.censor.names}{}

\item{data.title}{}

\item{data.note}{}

\item{skip}{}
}
\description{
Title
}
\examples{
\dontrun{

WorkStation.rdu <- frame.to.rdu(workstation,
                                ID.column = "station",
                                time.column = "days",
                                event.column = "event")

#attr(WorkStation.rdu,"WindowInfo")

WorkStation.mcf <- mcf(WorkStation.rdu)

#sqrt(WorkStation.mcf$Var)

event.plot(WorkStation.rdu)

plot(WorkStation.mcf)
}
}
