% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaaslaser}
\alias{gaaslaser}
\title{Gallium-Arsenic Laser Degradation Test}
\format{A \code{data.frame} with 255 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab increase \tab Percent increase in observed operating current \tab \bold{Numeric}\cr
  [, 2] \tab unit \tab Unit designator code \tab \bold{Categoric}\cr
  [, 3] \tab hours \tab Accumulated time when \code{increase} was measured \tab \bold{Numeric}
   }}
\source{
Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons.
}
\description{
Over the life of some lasers devices, degradation causes
             a decrease in light output.  In this test, a feedback mechanism was used 
             to maintain a constant light output by increasing the operating current.
             Fifteen GaAs laser devices were tested at an elevated temperature of 
             80^{o} C to accelerate the degradation process.  The operating current
             was measured in 250-hour intervals and the data were recorded as the
             percent increase in current compared to the initial operating current
             measured when the test began.  A device was considered to have failed
             if the percent increase reached 10\%.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{grampus}}, \code{\link{grids1}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
