% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_ld.R
\name{gadget_ld}
\alias{gadget_ld}
\title{SMRD Analysis Shiny Gadget}
\usage{
gadget_ld(data, theme = "flatly", width = "100\%", height = "600px",
  css = NULL)
}
\arguments{
\item{data}{A data set}

\item{theme}{\code{character} A bootswatch theme provided to \code{shinythemes::shinytheme}}

\item{width}{\code{character} Width of the gadget (in valid css units)}

\item{height}{\code{character} Height of the gadget (in valid css units)}

\item{css}{\code{character} Path to a custom css file}
}
\value{
A \code{list} of length 4
  \item{obj}{A \code{life.data} object}
  \item{dist}{The selected distribution}
  \item{cdfest}{A cdfest object}
  \item{mlest}{A mlest object}
}
\description{
Shiny gadget used to visually inspect column values in a data set 
and subset rows by specifying column values
}
\examples{
\dontrun{gadget_ld(data = shockabsorber)}

}
