% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-geng.R
\name{dgeng}
\alias{dgeng}
\alias{pgeng}
\alias{qgeng}
\alias{rgeng}
\title{The Generalized Gamma Distribution}
\usage{
dgeng(x, theta, delta, varrho)

pgeng(q, theta, delta, varrho)

qgeng(p, theta, delta, varrho, smalldelta = 1e-05)

rgeng(n, theta, delta, varrho)
}
\description{
The Generalized Gamma Distribution
}
