% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genmax}
\alias{genmax}
\title{New version that sends down arrays instead of pointers!!}
\usage{
genmax(kmod, kdist, theta, thetas, kodet, ifix, nparm, npard, y, ncoly,
  nrownw, x, ncolx, codes, weight, ty, ncolty, tcodes, kprint, kparv,
  nrvar, mrelat, nrelat, mnrvar, xlogl, yhat, resid, vcvs, vcv, r, start,
  lstar, conlev, ilabp, ilabd, ier, nxd, intd, ipxcd, irelad, fstder,
  nregr, kcentr, kpoint, ifit, kgtall, llog, kmodp, maxit, pest, epsx,
  npardm, nnum, kparm, iup, nterd, maxpd, pfail, kmccde, nstart, maxmsd,
  tol, lsd, pchmax)
}
\description{
#  kmod           (I)  model code (usual list 1-5)
 #  -------------------------------------------------------------
 #    0  distribution kdist with one or more parameters related to
 #       explanatory variables
 #    1  same as 0 with an lfp parameter
 #    2  same as 0 with a doa parameter
 #    3  steady-state model with distribution kdist plus exponential
 #       competing risk
 #    4  proportional hazards model for distribution kdist plus
 #       explanatory variables for the power (or other) parameters
 #
 #  kdist         (I)   distribution number (usual list 1-12)
 #                   >100 is a user specified distribution
 #
 #  theta(nparm)   (O)   double precision vector returning ml estimates
 #                      of model parameters and values of any
 #                      fixed parameters
 #
 #  thetas(nparm) (O)   double precision vector returning ml
 #                      scaled estimates
 #                   of model parameters and (scaled) values of
 #                   any fixed parameters
 #                   data scaled according to centered and scaled x
 #
 #  kodet(nparm)  (O)   integer vector of parameter codes indicating range
 #                        0 fixed
 #                        1 unrestricted
 #                        2 positive
 #                        3 0-1
 #
 #
 #  ifix(nparm)   (I)   an integer vector of 0's and 1's to indicate which
 #                   parameters are to be fixed (1 for fixed)
 #
 #  nparm       (O)     number of model parameters
 #                   nparm is the sum of the parameters in all of the
 #                   regression relationships plus any model parameters
 #                   that do not depend on explanatory variables
 #                   If this number is needed above before calling
 #                   genmax (to allocate parameter space) it can be
 #                   gotten from a call to gensiz
 #                   the model stuff (kmod,kdist,kparv,nrvar,mrelat,nrelat)
 #                   has been set.
 #
 #  npard       (O)     number of model parameters
 #                   npard is the  number of parameters in the distributional
 #                   model (i.e., for a particular combination of the
 #                   explanatory variables
 #                   If this number is needed above before calling
 #                   genmax (to allocate parameter space) it can be
 #                   gotten from a call to gensiz
 #                   the model stuff (kmod,kdist,kparv,nrvar,mrelat,nrelat)
 #                   has been set.
 #
 #  y            (I)  real matrix of response times .
 #                      transformed times may be returned
 #
 #  ncoly        (I)    number of columns in the observation matrix (1 or 2)
 #
 #  nrownw        (I)     number of rows in the data matrix
 #
 #  x         (I)     real matrix x of explanatory variables
 #                      which  must have a
 #                      column of ones in the first column (even if it
 #                      is not used).
 #
 #               *****  note that upon return, x may not be in its original
 #                      form due to scaling or other standardization.
 #
 #  ncolx         (I)   number of columns in x, not including the required col
 #                      of ones (because we think of the ones as column zero)
 #
 #  codes         (I)  real vector of censor codes
 #
 #  weight        (I)   real vector of observation weights
 #
 #  ty          (I)   real matrix of truncation times
 #
 #  ncolty       (I)    number of columns in the matrix of truncation times
 #                      (0, 1, or 2)
 #
 #  tc           (I)  real vector of truncation codes
 #
 #  kprint       (I)    print code 0-none  1-minimal 2-usual 3-light debug
 #                   4 and higher provides more debug output
 #
 #
 #  kparv        (I) integer vector giving parameter
 #                   numbers having regression
 #                   relationships
 #
 #  nrvar      (I)   integer vector. nrvar(i) gives the number of columns
 #                   of x that are in the relationship for the
 #                   parameter specified in kparv(i)
 #
 #  mrelat(mnrvar,nrelat)  (I) integer matrix
 #                   in which entry i,j gives the col of x
 #                   for the ith term of the jth relationship.  i=1,nrvar(j),
 #                   j=1,nrelat
 #
 #  nrelat     (I) number of regression relationships
 #
 #  mnrvar     (I)   integer vector. nrvar(i) gives the number of columns
 #
 #  xlogl        (O) double precision scaler returning value of the
 #                   log likelihood
 #
 #  yhat       (O) real vector (matrix) of fitted values (same size as y)
 #
 #  resid       (O) real vector  (matrix) of residuals (same size as y)
 #
 #  vcv(nparm,nparm) (O) double precision matrix returning estimated matrix of
 #                   variances and covariances of the ml estimates
 #
 #  vcvs(nparm,nparm)(O)  double precision matrix returning estimated
 #                        matrix of variances and covariances of the
 #                        ml scaled estimates
 #
 #  r(nparm,nparm)   (O)  double precision matrix returning estimated
 #                        correlation  matrix from the vcv matrix
 #
 # start           (I--opt)  optional vector of starting parameter values
 #                       (ignored if lstar=0)
 #
 # lstar            (I)    =1 is start values are in start
 #                         =0 if automatic start values are
 #                             to be computed velow
 #
 # ilabp           (O)  integer vector of parameter labels (length 8*nparm)
 #                      in this new version we copy over to stack and pass
 #                      pointers below
 #
 # ilabd           (O)  integer vector of distribution parmameters
 #                      (length 8*nparm)
 #                      in this new version we copy over to stack and pass
 #                      pointers below
 #
 #  ier            code       meaning
 #     -------------------------------------------------------------------
 #     third digit:  0 no optimization errors detected
 #                   1 likelihood shape caused problems with the powell alg
 #                   2 convergence criterion not met after maximum number
 #                     or iterations
 #
 #   second digit:   0 first derivatives of likelihood small
 #                   1 first derivatives of the loglikelihood too large
 #
 #    first digit:   0 estimated fisher info matrix inverted successfully
 #                   1 estimated fisher info matrix appears to be singular
 #
 #note: all pointers sent down in the argument list are pointing to the
 #     rs(.) stack in labeled common.
 #
 #      fortran typing conventions are generally followed for typing
 #      except that double precision is used except when data is to
 #      be taken from the worksheet.
}
