% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GENSIZ}
\alias{GENSIZ}
\title{Compute the sizes of parameter vectors}
\usage{
GENSIZ(
  kmod,
  kdist,
  kparv,
  nrvar,
  mrelat,
  nrelat,
  mnrvar,
  ncolx,
  kprint,
  nparm,
  npard,
  ier,
  nxd,
  intd,
  ipxcd,
  irelad,
  ilabp,
  ilabd,
  nregr,
  kgtall,
  llog,
  kmodp,
  npardm,
  nnum,
  kparm,
  iup,
  nterd,
  maxpd
)
}
\arguments{
\item{kmod}{Integer model code (see Details)}

\item{kdist}{Integer code representing distribution number 
usually in 1-12. \code{kdist > 100}> signifies
a user specified distribution.}

\item{kparv}{Integer vector giving parameter numbers 
that have regression relationships.}

\item{nrvar}{Integer vector. \code{nrvar[i]} gives the number of 
columns in the x-matrix that are in the relationship 
for the parameter specified in \code{kparv[i]}.}

\item{mrelat}{Integer matrix with dimensions [mnrvar,nrelat]
in which \code{mrelat[i,j]} gives the col of 
the x-matrix for the ith term of the jth relationship.  
\code{i = 1,nrvar[j], j = 1,nrelat} (see Details).}

\item{nrelat}{Integer giving the number of regression relationships}

\item{mnrvar}{Integer giving the number of columns in the x-matrix
\code{mnrvar = max(nrvar)}.}

\item{ncolx}{Integer giving the number of columns in the x-matrix. This does not include
the required column of ones (we think of the ones as column zero).}

\item{kprint}{Integer giving the print code for debugging output (see Details).}

\item{nparm}{Integer giving the sum of the parameters in all of the regression 
relationships plus any model parameters that do not depend on 
explanatory variables.}

\item{npard}{Integer giving the number of parameters in the distributional
model (i.e., for a particular combination of the explanatory 
variables.}

\item{ier}{Three-digit integer giving error codes upon return (see Details)}

\item{nxd}{Integer vector (length = 5) giving the number of dimensions in the x-matrix for each relationship}

\item{intd}{Integer vector (length = 5) giving the number of relationships with an intercept term}

\item{ipxcd}{Integer vector (length = 5) giving the columns in the x-matrix that are involved with each relationship}

\item{irelad}{Integer vector (length = 5) giving the number of things (don't know yet).}

\item{ilabp}{Integer vector of parameter labels (length 8*nparm) in this new 
version we copy over to stack and pass pointers below.}

\item{ilabd}{Integer vector of distribution parmameters (length 8*nparm)
in this new version we copy over to stack and pass 
pointers below.}
}
\description{
Given model information compute nparm and 
             npard giving the sizes of parameter vectors.
}
\details{
See Rcpp function \code{genmax} for more information 
         on parameters than what is shown here.
         
         \code{kmod} values:
         \describe{
          \item{0}{Distribution \code{kdist} with one or more parameters related to
                   explanatory variables.}
          \item{1}{Same as \code{kmod = 0} with an lfp parameter.}
          \item{2}{Same as \code{kmod = 0} with a doa parameter.}
          \item{3}{Steady-state model with distribution \code{kdist} plus
                   exponential competing risk.}
          \item{4}{Proportional hazards model for distribution \code{kdist} 
                   plus explanatory variables for the power (or other) parameters.}
         }
         
         Columns in the x-matrix may have a functional relationship
         with the location parameter, the scale parameter, or another model
         parameter. Each column in the matrix \code{mrelat} gives which columns
         in the x-matrix are related to a specific parameter.  If all columns
         in the x-matrix are related to the location parameter only one column
         in \code{mrelat} will contain non-zero elements.
         
         \code{kprint} values:
         \describe{
          \item{0}{no printing at all}
          \item{1}{minimal printing}
          \item{2}{usual printing - no debugging}
          \item{3}{light debugging}
          \item{4+}{heavy debugging}
         }
         
         \code{ier} error code values:
         \itemize{
          \item{third digit:
            \describe{
             \item{0}{No optimization errors detected}
             \item{1}{Likelihood shape caused problems with the powell alg}
             \item{2}{Convergence criterion not met after maximum number of iterations}
            }
            }
            \item{second digit:
             \describe{
              \item{0}{first derivatives of likelihood small}
              \item{1}{first derivatives of the loglikelihood too large}
             }
             }
             \item{
              \describe{
              \item{0}{estimated fisher info matrix inverted successfully}
              \item{1}{estimated fisher info matrix appears to be singular}
              }
             }
         }
}
