% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ADDT.plan.values.R
\name{get.ADDT.plan.values}
\alias{get.ADDT.plan.values}
\title{Title}
\usage{
get.ADDT.plan.values(distribution, transformation.x,
  transformation.response, transformation.time, beta0, beta1, beta2,
  slope.at, sigma, beta, the.slope, time.units = "Time", response.units,
  accelvar.units, beta0.cr = NULL, beta1.cr = NULL, beta2.cr = NULL,
  slope.at.cr = NULL, sigma.cr = NULL, beta.cr = NULL,
  the.slope.cr = NULL, power = NULL, FailLevel = NULL,
  use.condition = NULL)
}
\arguments{
\item{distribution}{}

\item{transformation.x}{}

\item{transformation.response}{}

\item{transformation.time}{}

\item{beta0}{}

\item{beta1}{}

\item{beta2}{}

\item{slope.at}{}

\item{sigma}{}

\item{beta}{}

\item{the.slope}{}

\item{time.units}{}

\item{response.units}{}

\item{accelvar.units}{}

\item{beta0.cr}{}

\item{beta1.cr}{}

\item{beta2.cr}{}

\item{slope.at.cr}{}

\item{sigma.cr}{}

\item{beta.cr}{}

\item{the.slope.cr}{}

\item{power}{}

\item{FailLevel}{}

\item{use.condition}{}
}
\description{
Title
}
\examples{
\dontrun{

InsulationBrkdwn.ADDTplan <- get.allocation.matrix(list(DegreesC = c(180,225,250,275)),
                                                   times = c(1,2,4,8,16,32,48,64),
                                                   time.units = "Weeks",
                                                   reps = 4)

plot(InsulationBrkdwn.ADDTplan)

InsulationBrkdwn.ADDTpv <- get.ADDT.plan.values(distribution = "normal",
                                                transformation.x = "Arrhenius", 
                                                transformation.Response = "log", 
                                                transformation.time = "linear",
                                                beta0 = 2.58850162033243,
                                                beta1 = -476873415881.376,
                                                beta2 = 1.41806367703643,
                                                sigma = 0.172609,
                                                time.units = "Weeks",
                                                response.units = "Volts", 
                                                FailLevel = 10, 
                                                use.condition = 100)

print(InsulationBrkdwn.ADDTpv)

InsulationBrkdwn.vADDTplan <- hframe.to.vframe(InsulationBrkdwn.ADDTplan)
sum(allocation(InsulationBrkdwn.vADDTplan))

names(InsulationBrkdwn.ADDTpv)

InsulationBrkdwn.plan.sim.out <- sim.ADDT.test.plan(ADDT.test.plan = InsulationBrkdwn.ADDTplan, 
                                                    ADDT.plan.values = InsulationBrkdwn.ADDTpv, 
                                                    number.sim = 5)

ADDT.plot.time.v.x(InsulationBrkdwn.plan.sim.out)

ADDT.plot.Deg.v.Time(InsulationBrkdwn.plan.sim.out)
ADDT.plot.FracFail.v.Time(InsulationBrkdwn.plan.sim.out)

ADDT.vcv(ADDT.plan.values = InsulationBrkdwn.ADDTpv,
         ADDT.test.plan = hframe.to.vframe(InsulationBrkdwn.ADDTplan))


}
}
