% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.alt.plan.values.from.slope.and.point.R
\name{get.alt.plan.values.from.slope.and.point}
\alias{get.alt.plan.values.from.slope.and.point}
\title{Title}
\usage{
get.alt.plan.values.from.slope.and.point(distribution, relationship, slope,
  probs, accelvar, censor.time, beta, sigma, time.units = "Time",
  accelvar.units, power = NULL, use.conditions = NULL)
}
\arguments{
\item{distribution}{}

\item{relationship}{}

\item{slope}{}

\item{probs}{}

\item{accelvar}{}

\item{censor.time}{}

\item{beta}{}

\item{sigma}{}

\item{time.units}{}

\item{accelvar.units}{}

\item{power}{}

\item{use.conditions}{}
}
\description{
Title
}
\examples{
\dontrun{

AdhesiveBond.Weibull.altpv <- get.alt.plan.values.from.slope.and.point(distribution = "Weibull",
                                                                       relationship = "Arrhenius", 
                                                                       accelvar.units = c("DegreesC"),
                                                                       time.units = "Days", 
                                                                       censor.time = 183,
                                                                       probs = c(.001), 
                                                                       accelvar = c(50),
                                                                       slope = 0.726, 
                                                                       beta = 1.667)

print(AdhesiveBond.Weibull.altpv)

AdhesiveBond1.altplan <- get.alt.test.plan.direct(accel.variable.levels = c(78,98,120),
                                                  number.of.units = c(155,60,84),
                                                  censor.times = c(183,183,183))

AdhesiveBond2.altplan <- get.alt.test.plan.direct(accel.variable.levels = c(80,100,120),
                                                  number.of.units = c(212,88,0),
                                                  censor.times = c(243,243,243))

AdhesiveBond3.altplan <- get.alt.test.plan.direct(accel.variable.levels = c(80,100,120),
                                                  number.of.units = c(150,60,90),
                                                  censor.times = c(243,243,243))

plot(AdhesiveBond1.altplan,
     ALT.plan.values = AdhesiveBond.Weibull.altpv,
     use.condition = 50)

plot(AdhesiveBond2.altplan,
     ALT.plan.values = AdhesiveBond.Weibull.altpv,
     use.condition = 50)

plot(AdhesiveBond3.altplan,
     ALT.plan.values = AdhesiveBond.Weibull.altpv,
     use.condition = 50)

}
}
