% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.alt.test.plan.direct.R
\name{get.alt.test.plan.direct}
\alias{get.alt.test.plan.direct}
\title{Title}
\usage{
get.alt.test.plan.direct(accel.variable.levels, number.of.units,
  censor.times = NULL, accelvar.names = "accel.variable",
  describe.string = "")
}
\arguments{
\item{accel.variable.levels}{}

\item{number.of.units}{}

\item{censor.times}{}

\item{accelvar.names}{}

\item{describe.string}{}
}
\description{
Title
}
\examples{
\dontrun{

advbond.model1 <- 
  get.alt.plan.values.from.two.points(
    distribution = "Weibull",
    relationship = "Inverse Power Rule", 
    time.units = "hours", 
    censor.time = 1000,
    probs = c(.001,.9), 
    accelvar.units = "volts", 
    accelvar = c(110,150), 
    beta = 1.667)

advbond.test.plan1 <- 
  get.alt.test.plan.direct(accel.variable.levels = c(130,140,150),
                           number.of.units = c(100,100,100))

plot(advbond.test.plan1,
     ALT.plan.values = advbond.model1, 
     my.title = "",
     use.conditions = 120)

advbond.model2 <- 
  get.alt.plan.values.from.two.points(distribution = "Lognormal",	
                                      relationship = "Arrhenius", 
                                      time.units = "days",
                                      censor.time = 183,
                                      probs = c(.001,.9),
                                      accelvar = c(50,120),
                                      sigma = .5)


advbond.test.plan2 <- get.alt.test.plan.direct(accel.variable.levels = c(80,100,120),
                         number.of.units = c(100,100,100))

plot(advbond.test.plan2, 
     ALT.plan.values = advbond.model2, 
     my.title = "",
     use.condition = 50)
}
}
