% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.alt.test.plan.frame.R
\name{get.alt.test.plan.frame}
\alias{get.alt.test.plan.frame}
\title{Title}
\usage{
get.alt.test.plan.frame(
  frame,
  levels.columns,
  censor.column,
  allocation.column,
  describe.string = ""
)
}
\arguments{
\item{frame}{}

\item{levels.columns}{}

\item{censor.column}{}

\item{allocation.column}{}

\item{describe.string}{}
}
\description{
Title
}
\examples{
\dontrun{

AdhesiveBond2.frame <- data.frame(DegreesC = c(80,100,120),
                                  SampleSize = c(150,60,90),
                                  CensorTimes = c(243,243,243))

AdhesiveBond2.altplan <- 
  get.alt.test.plan.frame(AdhesiveBond2.frame,
                          levels.columns = "DegreesC", 
                          censor.column = "CensorTimes",
                          allocation.column = "SampleSize",
                          describe.string = "AdhesiveBond Test Plan")

print(AdhesiveBond2.altplan)

}
}
