% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.plan.values.R
\name{get.plan.values}
\alias{get.plan.values}
\title{Get planning value for a reliability test}
\usage{
get.plan.values(distribution, beta, sigma, prob, time, time.units = "Time")
}
\arguments{
\item{distribution}{}

\item{beta}{}

\item{sigma}{}

\item{prob}{}

\item{time}{}

\item{time.units}{}
}
\value{
A list of reliability test plan values
}
\description{
Get planning value for a reliability test
}
\examples{
\dontrun{

 plan.values1 <- get.plan.values("Weibull", 
                                 beta = 2, 
                                 prob = .1, 
                                 time = 100, 
                                 time.units = "Hours")
                                 
life.test.simulation(plan.values1, 
                     n = 50,
                     censor.time = 120, 
                     number.detail = 5, 
                     quantile.mark = 0.2) 
                     
plan.values2 <- get.plan.values("Lognormal", 
                                sigma = 0.5,
                                prob = 0.1, 
                                time = 100, 
                                time.units = "Hours")

summary(plan.values2)
plot(plan.values2)

plot(plan.values2, censor.time=1000, grids=F)

life.test.simulation(plan.values2, 
                     n = 50,
                     censor.time = 1000, 
                     quantile.mark = .1)

plan.values3 <- get.plan.values("Weibull",
                                prob = c(.2,.12),
                                time = c(1000,500), 
                                time.units = "Hours")

plan.values4 <- get.plan.values("Weibull",
                                prob = c(.05,.15),
                                time = c(40000,100000),
                                time.units = "Hours")

summary(plan.values3)
plot(plan.values3)


life.test.simulation(plan.values3, 
                     n = 50, 
                     censor.time = 1000, 
                     quantile.mark = 0.1)

#compare the simulated value with the large-sample approx below

asd.quant(plan.values3, 
          n=50, 
          censor.time = 1000, 
          quantile.mark = 0.1)

#compare:

asd.quant(plan.values3,
          n=50, 
          censor.time = 1000, 
          quantile.mark = 0.1)*sqrt(50)

asd.quant(plan.values3,
          n=500, 
          censor.time = 1000, 
          quantile.mark = 0.1)*sqrt(500)

asd.quant(plan.values3,
          n=5000, 
          censor.time = 1000, 
          quantile.mark = 0.1)*sqrt(5000)

}
}
\seealso{
code{link{life.test.simulation}}
         code{link{asd.quant}}
}
