% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sub.model.dest.degrad.mle.out.R
\name{get.sub.model.dest.degrad.mle.out}
\alias{get.sub.model.dest.degrad.mle.out}
\title{Title}
\usage{
\method{get.sub.model}{dest.degrad.mle.out}(x, stresses, ...)
}
\arguments{
\item{x}{}

\item{stresses}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

AdhesiveStrength.ddd <- frame.to.ddd(adhesivestrength, 
                                     response.column = "pounds",
                                     time.column = "days",
                                     x.columns = "celsius",
                                     data.title = "AdhesiveStrength Strength Data",
                                     time.units = "Days")

AdhesiveStrength.gmle <- dest.degrad.mle(AdhesiveStrength.ddd,
                                         distribution = "normal",
                                         transformation.response = "log", 
                                         transformation.x = "arrhenius", 
                                         transformation.time = "linear")

get.sub.model.dest.degrad.mle.out(AdhesiveStrength.gmle,c(50,60,70))


# Or more simply
get.sub.model(AdhesiveStrength.gmle,c(50,60,70))

}
}
