% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.pdf.plot.R
\name{gets.pdf.plot}
\alias{gets.pdf.plot}
\title{Title}
\usage{
gets.pdf.plot(
  distvec = c("sev", "normal", "lev"),
  sigmavec = c(-0.75, 0, 0.75),
  zetavec = c(0.5, 1, 2),
  alphavec = 0,
  len = 200,
  lwd = 1.5,
  cex = 16,
  plotem = T,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

gets.pdf.plot()

}
}
