% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmleprobplot.R
\name{gmleprobplot}
\alias{gmleprobplot}
\title{Title}
\usage{
gmleprobplot(data.ld, distribution, plot.dist = "lognormal",
  xlab = get.time.units(data.ld), xlim = c(NA, NA), ylim = c(NA, NA),
  time.range = c(NA, NA),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100, interactive = T,
  original.par = F, mle.quantiles = T, my.title = NULL, type = "y",
  cex = 1, cexlab = 1, sub.title = "", grids = F,
  linear.axes = F, slope.axis = F, print.table = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"), trunc.correct = T,
  add = F, plot.censored.ticks = F, pch = 16, lty = c(3, 4, 5, 6,
  7), lwd = c(2, 2, 2, 2, 2), length.time.vec = 100,
  compare.dists = NULL, col.ci = 4, lwd.ci = 2, xxx.mle.out = NULL,
  ciMethod = "normal.approx", extrapolate.ci = F, band.type = "p",
  ...)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{plot.dist}{}

\item{xlab}{}

\item{xlim}{}

\item{ylim}{}

\item{time.range}{}

\item{conf.level}{}

\item{interactive}{}

\item{original.par}{}

\item{mle.quantiles}{}

\item{my.title}{}

\item{type}{}

\item{cex}{}

\item{cexlab}{}

\item{sub.title}{}

\item{grids}{}

\item{linear.axes}{}

\item{slope.axis}{}

\item{print.table}{}

\item{title.option}{}

\item{ylab}{}

\item{trunc.correct}{}

\item{add}{}

\item{plot.censored.ticks}{}

\item{pch}{}

\item{lty}{}

\item{lwd}{}

\item{length.time.vec}{}

\item{compare.dists}{}

\item{col.ci}{}

\item{lwd.ci}{}

\item{xxx.mle.out}{}

\item{ciMethod}{}

\item{extrapolate.ci}{}

\item{band.type}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{
lzbearing.ld <- frame.to.ld(lzbearing, response.column = 1)

gmleprobplot(lzbearing.ld,
             distribution="bisa",
             compare=c("Lognormal"))

}
}
