% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmlest.R
\name{gmlest}
\alias{gmlest}
\title{Generalized maximum likelihood}
\usage{
gmlest(data.ld, distribution, theta.start = NULL, explan.vars = NULL,
  mu.relat = NULL, sigma.relat = NULL, prob.relat = NULL,
  gamthr = 0, escale = 10000, e = rep(1e-04, nparm),
  parameter.fixed = rep(F, nparm), intercept = T, model = 0,
  kprint = 0, conlev = 0.95, maxit = 500, debug1 = F)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{theta.start}{}

\item{explan.vars}{}

\item{mu.relat}{}

\item{sigma.relat}{}

\item{prob.relat}{}

\item{gamthr}{}

\item{escale}{}

\item{e}{}

\item{parameter.fixed}{}

\item{intercept}{}

\item{model}{}

\item{kprint}{}

\item{conlev}{}

\item{maxit}{}

\item{debug1}{}
}
\value{
A big thing
}
\description{
Generalized maximum likelihood
}
\examples{
\dontrun{

 ConnectionStrength.ld <- 
 frame.to.ld(connectionstrength,
              response.column = 1,
              failure.mode.column = 2,
              case.weight.column = 3)

 mfm.to.ld(ConnectionStrength.ld)
 
 gmlest(ConnectionStrength.Bond.ld,
        distribution = "normal")
}
}
