% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-goma.R
\name{dgoma}
\alias{dgoma}
\alias{pgoma}
\alias{qgoma}
\title{The Gompertz Makeham Distribution}
\usage{
dgoma(x, shape, shape2, scale = 1)

pgoma(q, shape, shape2, scale = 1)

qgoma(p, shape = stop("no shape arg"), shape2 = stop("no shape2 arg"),
  scale = 1)
}
\description{
The Gompertz Makeham Distribution
}
