% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grampus}
\alias{grampus}
\title{Unscheduled maintenance actions for the U.S.S. Grampus}
\format{A \code{data.frame} with 57 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab unit   \tab Unit type \tab \bold{Categoric}\cr
  [, 2] \tab kilohours \tab Accumulated hours at \code{event} (in thousands) \tab \bold{Numeric}\cr
  [, 3] \tab event \tab Event observed at \code{khours} (Repair/End)  \tab \bold{Categoric}
   }}
\source{
Lee, L. (1980) 
        Testing adequacy of the Weibull and loglinear rate models for a Poisson Process,
        Technometrics, 22, 195-199

Ascher, H. and Feingold, H. (1984), Repairable Systems Reliability, New York, NY; Marcel Dekker
}
\description{
Lee presents a dataset containing the times (in thousands of operating hours) of unscheduled maintenance actions on the 
             number 4 diesel engine of the U.S.S. Grampus. The dataset contains observations for the first 16,000 hours of operation.  
             The observations in this dataset should be treated as is they were observed from a single system, since information as 
             to which component in the engine failed was not included. The unscheduled maintenance actions were caused by either system 
             failures or by events that indicated that a system failure was imminent. Such maintenance actions are inconvenient and 
             expensive.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grids1}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
