% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek.R
\name{greek}
\alias{greek}
\title{Unicode Greek letters}
\usage{
greek(ltr)
}
\arguments{
\item{ltr}{A english language representation 
of a letter from the Greek alphabet}
}
\value{
A Unicode representation of a letter
        from the Greek alphabet
}
\description{
Unicode Greek letters
}
