% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grids1}
\alias{grids1}
\title{Locomotive braking grid replacements}
\format{A \code{data.frame} with 39 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab unit \tab Unit designator code \tab \bold{Categoric}\cr
  [, 2] \tab days \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 3] \tab event \tab Event observed at \code{days} (Replacement/End)  \tab \bold{Categoric}
   }}
\source{
Doganaksoy, N., and Nelson, W. (1991) A method and computer program MCFDIFF to compare two samples of repair data,
        TIS Report 91CRD172, General Electric Company Research and Development, Schenectady, NY.
}
\description{
Age of locomotives when either their braking grids were replaced or the largest age observed for each locomotive
}
\details{
Two batches \code{grids1} & \code{grids2}
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grampus}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
