% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupi.contour.R
\name{groupi.contour}
\alias{groupi.contour}
\title{Title}
\usage{
groupi.contour(data.ld, distribution, stresses = get.x.markers(data.ld,
  group.var = group.var), group.var = 1, contour.indicators = NULL,
  xlim = c(NA, NA), ylim = c(NA, NA), loc.or.quant = "Quantile",
  the.quantile = 0.1, rel.or.conf = "Joint confidence region",
  my.title = NULL, cex = 1.5,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  do.legend = "On plot", stresses.limit = 15, size = 50,
  the.factor = 3.5, show.mle = F, original.par = T,
  col.fhat.vec = 1:length(stresses), lty = (1:(length(stresses) +
  1))[-2], ...)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{stresses}{}

\item{group.var}{}

\item{contour.indicators}{}

\item{xlim}{}

\item{ylim}{}

\item{loc.or.quant}{}

\item{the.quantile}{}

\item{rel.or.conf}{}

\item{my.title}{}

\item{cex}{}

\item{title.option}{}

\item{do.legend}{}

\item{stresses.limit}{}

\item{size}{}

\item{the.factor}{}

\item{show.mle}{}

\item{original.par}{}

\item{col.fhat.vec}{}

\item{lty}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

Snubber.ld <- frame.to.ld(snubber, 
                          response.column = "cycles", 
                          censor.column = "event",
                          time.units = "Cycles",
                          case.weight.column = "count",
                          x.columns = "design")
```

event.plot(Snubber.ld)
summary(Snubber.ld)

Snubber.groupi.nor.out <-  groupi.mleprobplot(Snubber.ld,"normal")

tmpxx <- groupi.contour(Snubber.ld,"Weibull",the.quantile=.1)
tmpxx <- groupi.contour(Snubber.ld,"lognormal",the.quantile=.1)
tmpxx <- groupi.contour(Snubber.ld,"lognormal")
tmpxx <- groupi.contour(Snubber.ld,"normal")
tmpxx <- groupi.contour(Snubber.ld,"normal",the.quantile=.1)

}
}
