% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupi.Dest.Degrad.oneplot.R
\name{groupi.Dest.Degrad.oneplot}
\alias{groupi.Dest.Degrad.oneplot}
\title{Title}
\usage{
groupi.Dest.Degrad.oneplot(data.ddd, distribution, transformation.response,
  transformation.time, ylim = c(NA, NA), xlim = c(NA, NA),
  my.title = NULL, ylab = NULL, xlab = NULL, cex = 1.2,
  cex.labs = 1.5, cex.points = 1, add = F, grids = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"), pch.point = NULL,
  response.on.yaxis = T, subset = T, do.legend = "On plot",
  fail.level = NULL, group.var = 1:ncol(xmat(data.ddd)),
  plot.lines = T, lty = NULL, ...)
}
\arguments{
\item{data.ddd}{}

\item{distribution}{}

\item{transformation.response}{}

\item{transformation.time}{}

\item{ylim}{}

\item{xlim}{}

\item{my.title}{}

\item{ylab}{}

\item{xlab}{}

\item{cex}{}

\item{cex.labs}{}

\item{cex.points}{}

\item{add}{}

\item{grids}{}

\item{title.option}{}

\item{pch.point}{}

\item{response.on.yaxis}{}

\item{subset}{}

\item{do.legend}{}

\item{fail.level}{}

\item{group.var}{}

\item{plot.lines}{}

\item{lty}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

InsulationBrkdwn.ddd <- frame.to.ddd(insulationbrkdwn,
                                     response.column = 3, 
                                     time.column = 1,
                                     x.columns = 2,
                                     data.title = "Voltage Breakdown Data",
                                     response.units = "Volts",
                                     time.units = "Weeks")

print(InsulationBrkdwn.ddd)

plot(InsulationBrkdwn.ddd,
     transformation.Response = "log",
     transformation.time = "linear")

tmp <- groupi.Dest.Degrad.indivplots(InsulationBrkdwn.ddd,
                                     transformation.Response = "log", 
                                     transformation.time = "linear",
                                     distribution = "normal")

groupi.Dest.Degrad.oneplot(InsulationBrkdwn.ddd,
                           transformation.Response = "log", 
                           transformation.time = "linear",
                           distribution="normal")
groupm.Dest.Degrad(InsulationBrkdwn.ddd, 
                   distribution = "normal",
                   transformation.Response = "log10",
                   transformation.x = "invtemp",
                   transformation.time = "linear")


groupm.Dest.Degrad(InsulationBrkdwn.ddd, 
                   distribution = "normal",
                   transformation.Response = "log",
                   transformation.x = "arrhenius",
                   transformation.time="linear")

# Do individual analyses at each level of temperature

InsulationBrkdwn.groupi.Dest.Degrad <-groupi.Dest.Degrad(InsulationBrkdwn.ddd,
                                                         distribution = "normal",
                                                         transformation.Response = "log", 
                                                         transformation.time = "sqrt")


plot(InsulationBrkdwn.groupi.Dest.Degrad,
     transformation.x = "Arrhenius")

InsulationBrkdwn.groupm.Dest.Degrad <-groupm.Dest.Degrad(InsulationBrkdwn.ddd,
                                                         distribution = "normal", 
                                                         transformation.Response = "log",
                                                         transformation.x = "arrhenius", 
                                                         transformation.time = "sqrt")

InsulationBrkdwn.groupm.Dest.Degrad<-groupm.Dest.Degrad(InsulationBrkdwn.ddd,
                                                        distribution = "normal",
                                                        transformation.Response = "log",
                                                        transformation.x = "arrhenius",
                                                        transformation.time = "sqrt",
                                                        new.data = c("150,260"))


}
}
