% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupi.mleprobplot.R
\name{groupi.mleprobplot}
\alias{groupi.mleprobplot}
\title{Title}
\usage{
groupi.mleprobplot(
  data.ld,
  distribution,
  gamthr = 0,
  stresses = get.x.markers(data.ld, group.var = group.var, do.order = T),
  group.var = 1,
  xlab = get.time.units(data.ld),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  time.range = c(NA, NA),
  dump = 1,
  grids = F,
  my.title = NULL,
  cex = 1,
  linear.axes = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  pch = (1:(length(stresses) + 1))[-2],
  lty = NULL,
  ci.list = NULL,
  lwd = rep(2, length(stresses)),
  plot.censored.ticks = F,
  trunc.correct = T,
  col.fhat.vec = (1:(length(stresses) + length(col.ci)))[-col.ci],
  col.ci = 4,
  shape = NULL,
  do.legend = "On plot",
  stresses.limit = 18,
  plotem = rep(T, length(stresses)),
  check.level = SMRDOptions("SMRD.DataCheck"),
  title.line.adj = -2,
  mar = c(4.5, 5.25, 3.5, 12.1),
  bty = if (grids) "o" else "L",
  ...
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{gamthr}{}

\item{stresses}{}

\item{group.var}{}

\item{xlab}{}

\item{ylab}{}

\item{conf.level}{}

\item{xlim}{}

\item{ylim}{}

\item{time.range}{}

\item{dump}{}

\item{grids}{}

\item{my.title}{}

\item{cex}{}

\item{linear.axes}{}

\item{title.option}{}

\item{pch}{}

\item{lty}{}

\item{ci.list}{}

\item{lwd}{}

\item{plot.censored.ticks}{}

\item{trunc.correct}{}

\item{col.fhat.vec}{}

\item{col.ci}{}

\item{shape}{}

\item{do.legend}{}

\item{stresses.limit}{}

\item{plotem}{}

\item{check.level}{}

\item{title.line.adj}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

DeviceA.ld <- frame.to.ld(devicea, 
                          data.title = "Device-A ALT Results",
                          response.column = 1,
                          time.units = "Hours",
                          censor.column = 2,
                          case.weight.column = 3,
                          x.columns = 4, 
                          xlab = "Degrees C")

print(DeviceA.ld)
summary(DeviceA.ld)

censored.data.plot(DeviceA.ld)

censored.data.plot(DeviceA.ld, 
                   y.axis ="log", 
                   x.axis = "Arrhenius")

groupi.mleprobplot(DeviceA.ld, 
                   distribution = "Weibull")

four.groupi.mleprobplot(DeviceA.ld)

DeviceA.weib.groupi <- groupi.mleprobplot(DeviceA.ld,
                                          distribution = "Weibull")

print(DeviceA.weib.groupi)
summary(DeviceA.weib.groupi)

}
}
