% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupm.mleprobplot.R
\name{groupm.mleprobplot}
\alias{groupm.mleprobplot}
\title{Title}
\usage{
groupm.mleprobplot(
  data.ld,
  distribution,
  formula = NULL,
  group.var = 1:ncol(xmat(data.ld)),
  xlab = get.time.units(data.ld),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  relationship = NULL,
  power = NULL,
  dump = 0,
  mle.intervals = F,
  cex = 1,
  grids = F,
  slope.axis = F,
  linear.axes = F,
  lty = NULL,
  plot.censored.ticks = F,
  time.range = c(NA, NA),
  shape = NULL,
  ci.list = NULL,
  col.ci = 4,
  printem = F,
  trunc.correct = T,
  include.interaction = F,
  new.data = NULL,
  plotem = T,
  do.legend = "On plot",
  stresses.limit = 18,
  number.points = 55,
  plotted = rep(T, length(stresses.plus)),
  from.six.plot = F,
  debug1 = F,
  theta.start = NULL,
  parameter.fixed = NULL,
  compute.subsets = T,
  check.level = SMRDOptions("SMRD.DataCheck"),
  title.line.adj,
  lwd = 2,
  mar = c(4.5, 5.25, 3.5, 12.1),
  bty = if (grids) "o" else "L",
  ...
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{formula}{}

\item{group.var}{}

\item{xlab}{}

\item{ylab}{}

\item{conf.level}{}

\item{xlim}{}

\item{ylim}{}

\item{relationship}{}

\item{power}{}

\item{dump}{}

\item{mle.intervals}{}

\item{cex}{}

\item{grids}{}

\item{slope.axis}{}

\item{linear.axes}{}

\item{lty}{}

\item{plot.censored.ticks}{}

\item{time.range}{}

\item{shape}{}

\item{ci.list}{}

\item{col.ci}{}

\item{printem}{}

\item{trunc.correct}{}

\item{include.interaction}{}

\item{new.data}{}

\item{plotem}{}

\item{do.legend}{}

\item{stresses.limit}{}

\item{number.points}{}

\item{plotted}{}

\item{from.six.plot}{}

\item{debug1}{}

\item{theta.start}{}

\item{parameter.fixed}{}

\item{compute.subsets}{}

\item{check.level}{}

\item{title.line.adj}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

ICDevice2.ld <- frame.to.ld(icdevice2, 
                            response.column = c(1,2),
                            censor.column = 3,
                            case.weight.column = 4,
                            x.column = 5,  
                            data.title = "New Technology Device ALT", 
                            xlabel = "Degrees C", 
                            time.units = "Hours")

groupi.mleprobplot(ICDevice2.ld, 
                   distribution = "Lognormal")

ICDevice02.groupm.lognor <- groupm.mleprobplot(ICDevice2.ld, 
                                               distribution = "Lognormal", 
                                               relationship = "Arrhenius", 
                                               ci.list = 6)
}
}
