% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heatexchanger}
\alias{heatexchanger}
\title{Nuclear power plant heat exchanger tube crack data}
\format{A \code{data.frame} with 9 rows and 5 variables:
\tabular{rlll}{
  [, 1] \tab lower \tab Start of an inspection interval (in years) \tab \bold{Numeric}\cr
  [, 2] \tab upper \tab End of an inspection interval (in years) \tab \bold{Numeric}\cr
  [, 3] \tab event \tab Event observed in the interval (left-censored/right-censored/interval-censored)\tab \bold{Categoric}\cr
  [, 4] \tab count \tab Number of events observed in the interval \tab \bold{Numeric}\cr
  [, 5] \tab plant  \tab Plant where the tube was installed \tab \bold{Categoric}
 }}
\source{
Meeker, W. Q. and Escobar, L. A. (1996), Statistical Methods for Reliability Data, New York, NY; Wiley-Interscience
}
\description{
Inspection data for heat exchanger tubes across three different nuclear power plants.  The data were recorded in 1983,
             at this point in time, Plant 1 had been in operation for 3 years, Plant 2 for 2 years, and Plant 3 for only 1 year. 
             Because all of the heat exchangers were manufactured according to the same design specifications and because the heat 
             exchangers were operated in generating plants run under similar tightly controlled conditions, the data from the 
             different plants was combined for the sake of making inferences and predictions about the time-to-crack distribution 
             of the heat exchanger tubes.
}
\details{
Nuclear power plants use heat exchangers to transfer energy from the reactor to steam turbines.  A typical heat exchanger 
         contains thousands of tubes through which steam flows continuously when the heat exchanger is in service.  With age, 
         heat exchanger tubes develop cracks, usually due to some combination of stress-corrosion and fatigue. A heat exchanger 
         can continue to operate safely when the cracks are small.  If cracks get large enough, however, leaks can develop, and 
         these could lead to serious safety problems and expensive, unplanned plant shut-down time. To protect against having leaks, 
         heat exchangers are taken out of service periodically so that its tubes (and other components) can be inspected with 
         nondestructive evaluation techniques.  At the end of each inspection period, tubes with detected cracks are plugged so that 
         water will no longer pass through them.  This reduces plant efficiency, but extends the life of the expensive heat exchangers. 
         With this in mind, heat exchangers are built with extra capacity and can remain in operation up until the point where a certain 
         percentage (e.g., 5\%) of the tubes have been plugged.
}
