% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igau.mle.R
\name{igau.mle}
\alias{igau.mle}
\title{Title}
\usage{
igau.mle(data.ld, debug1 = F, theta.start = NULL)
}
\arguments{
\item{data.ld}{}

\item{debug1}{}

\item{theta.start}{}
}
\description{
Title
}
\examples{
\dontrun{

lzbearing.ld <- frame.to.ld(lzbearing, 
                            response.column = 1)

bear.igau.gmle.out <- igau.mle(lzbearing.ld)

bkfatigue10.ld <- frame.to.ld(bkfatigue10, 
                              response.column = 1,
                              time.units = "Kilocycles")
   
bkfatigue10.igau.gmle.out <- igau.mle(bkfatigue10.ld)                           
}
}
