% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interarrival.plot.R
\name{interarrival.plot}
\alias{interarrival.plot}
\title{Title}
\usage{
interarrival.plot(
  data.rdu,
  my.title = NULL,
  xlab = paste("Age in", get.time.units(data.rdu)),
  ylab = "Time Between Failure",
  ...
)
}
\arguments{
\item{data.rdu}{}

\item{my.title}{}

\item{xlab}{}

\item{ylab}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")

interarrival.plot(halfbeak.rdu,
                  xlab = "Thousands of Hours of Operation",
                  ylab = "Thousands of Hours Between Maintenance Actions",
                  my.title="")

}
}
