% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interarrival.times.R
\name{interarrival.times}
\alias{interarrival.times}
\title{Title}
\usage{
interarrival.times(data.rdu, small.for.tie = 1e-05, replace.zero = T)
}
\arguments{
\item{data.rdu}{}

\item{small.for.tie}{}

\item{replace.zero}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")


summary(halfbeak.rdu)
event.plot(halfbeak.rdu)
print(mcf(halfbeak.rdu))
mcf.plot(halfbeak.rdu)
interarrival.times(halfbeak.rdu)

}
}
