% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-igau.R
\name{digau}
\alias{digau}
\alias{pigau}
\alias{qigau}
\title{The Inverse Gaussian Distribution}
\usage{
digau(x, shape, scale = 1)

pigau(q, shape, scale = 1)

qigau(p, shape = stop("no shape arg"), scale = 1)
}
\description{
The Inverse Gaussian Distribution
}
