% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.test.R
\name{laplace.test}
\alias{laplace.test}
\title{Title}
\usage{
laplace.test(data.rdu)
}
\arguments{
\item{data.rdu}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")

laplace.test(halfbeak.rdu)
lewis.robinson.test(halfbeak.rdu)
milhbk189.test(halfbeak.rdu)

}
}
