% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-lev.R
\name{dlev}
\alias{dlev}
\alias{plev}
\alias{qlev}
\title{The Largest Extreme Value Distribution}
\usage{
dlev(x, location = 0, scale = 1)

plev(q, location = 0, scale = 1)

qlev(p, location = 0, scale = 1)
}
\description{
The Largest Extreme Value Distribution
}
