% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.test.simulation.R
\name{life.test.simulation}
\alias{life.test.simulation}
\title{Perform a life test simulation using planning values}
\usage{
life.test.simulation(plan.values, n, censor.time = NULL,
  censor.number = NULL, my.title, number.sim = 2000,
  quantile.mark = NULL, number.points = 20, perc.low = 0.005,
  perc.high = 0.995, cex = 1, xlim = c(NA, NA), ylim = NULL,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  xlab = plan.values$time.units, print.bad.count = T,
  max.lines.to.plot = min(50, number.sim), save.data.list = NULL,
  number.detail = 5, force.from.gui = F,
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  show.histogram = TRUE,
  band.method = GetSMRDDefault("SMRD.ConfidenceBandMethod"),
  mono.tran = FALSE, ...)
}
\arguments{
\item{plan.values}{}

\item{n}{}

\item{censor.time}{}

\item{censor.number}{}

\item{my.title}{}

\item{number.sim}{}

\item{quantile.mark}{}

\item{number.points}{}

\item{perc.low}{}

\item{perc.high}{}

\item{cex}{}

\item{xlim}{}

\item{ylim}{}

\item{title.option}{}

\item{xlab}{}

\item{print.bad.count}{}

\item{max.lines.to.plot}{}

\item{save.data.list}{}

\item{number.detail}{}

\item{force.from.gui}{}

\item{conf.level}{}

\item{show.histogram}{}

\item{band.method}{}

\item{mono.tran}{}

\item{...}{}
}
\value{
A series of plots
}
\description{
Perform a life test simulation using planning values
}
\examples{
\dontrun{
plan.values1 <- get.plan.values("Weibull", 
                                 beta = 2, 
                                 prob = .1, 
                                 time = 100, 
                                 time.units = "Hours")
                                 
life.test.simulation(plan.values1, 
                     n = 50,
                     censor.time = 120, 
                     number.detail = 5, 
                     quantile.mark = 0.2) 
                     
plan.values2 <- get.plan.values("Lognormal", 
                                sigma = 0.5,
                                prob = 0.1, 
                                time = 100, 
                                time.units = "Hours")

life.test.simulation(plan.values2, 
                     n = 50,
                     censor.time = 1000, 
                     quantile.mark = .1)
}
}
