% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikeNHPPvec.R
\name{loglikeNHPPvec}
\alias{loglikeNHPPvec}
\title{Title}
\usage{
loglikeNHPPvec(data.rdu, theta.mat, form)
}
\arguments{
\item{data.rdu}{}

\item{theta.mat}{}

\item{form}{}
}
\description{
Title
}
\examples{
\dontrun{

AMSAAExactFail.rdu <- frame.to.rdu(amsaaexactfail,
                                   ID.column = "vehicle", 
                                   time.column = "miles" , 
                                   event.column = "event")

theta.mat <- matrix(c(1,1,2,2),2,2)
theta.mat <- c(1,2)
loglikeNHPPvec(AMSAAExactFail.rdu, theta.mat, "power.law")

theta.mat <- matrix(c(.01,.01,.02,.02),2,2)
loglikeNHPPvec(AMSAAExactFail.rdu, theta.mat, "log.linear")

}
}
