% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-loglogis.R
\name{dloglogis}
\alias{dloglogis}
\alias{ploglogis}
\alias{qloglogis}
\title{The Loglogistic Distribution}
\usage{
dloglogis(x, locationlog = 0, scalelog = 1)

ploglogis(q, locationlog = 0, scalelog = 1)

qloglogis(p, locationlog = 0, scalelog = 1)
}
\description{
The Loglogistic Distribution
}
