% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.mle.R
\name{ls.mle}
\alias{ls.mle}
\title{Title}
\usage{
ls.mle(data.ld, distribution = "weibull", debug1 = 0,
  theta.start = NULL, ...)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{debug1}{}

\item{theta.start}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1, 
                                censor.column = 3,
                                time.units = "Kilometers")
                                
SA.weibull.gmle <- ls.mle(ShockAbsorber.ld, 
                          distribution = "Weibull")
}
}
