% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls2.mle.R
\name{ls2.mle}
\alias{ls2.mle}
\title{Title}
\usage{
ls2.mle(data.ld, distribution = "weibull", debug1 = F,
  theta.start = NULL)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{debug1}{}

\item{theta.start}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1, 
                                censor.column = 3,
                                time.units = "Kilometers")
                                
SA.weibull.gmle <- ls2.mle(ShockAbsorber.ld, 
                          distribution = "Weibull")
                          
BearingCage.ld <- frame.to.ld(bearingcage,
                              response.column = 1, 
                              censor.column = 2, 
                              case.weight.column = 3,
                              time.units = "Hours")

summary(BearingCage.ld)

ls2.mle(BearingCage.ld, 
        distribution ="Weibull", 
        theta.start = c(9,.4))
}
}
