% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsinf.R
\name{lsinf}
\alias{lsinf}
\title{Title}
\usage{
lsinf(z, censor.type, distribution, force = F, debug1 = F)
}
\arguments{
\item{z}{}

\item{censor.type}{}

\item{distribution}{}

\item{force}{}

\item{debug1}{}
}
\value{
list of values
}
\description{
Title
}
\examples{
\dontrun{

lsinf(seq(-1,1, by = 0.1),"right","sev")

lsinf(seq(-2,2, by = 0.2),"right","normal")

}
}
