% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lzbearing}
\alias{lzbearing}
\title{Ball bearing fatigue test data}
\format{
A \code{data.frame} with 23 rows and 1 variable:
\tabular{rlll}{
  [, 1] \tab megacycles \tab Accumulated cycles at failure (in millions) \tab \bold{Numeric}
 }
}
\source{
Lawless, J. F. (1982), Statistical Models and Methods for Lifetime Data, New York, NY; Wiley & Sons
}
\description{
The ball bearings came from four different major bearing companies. 
             There was disagreement in the industry as to the appropriate parameter 
             values to use to describe the relationship between fatigue life and stress 
             loading. The main objective of the study was to estimate values of the parameters 
             in the equation relating bearing life to load.
}
