% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf.plot.R
\name{mcf.plot}
\alias{mcf.plot}
\title{Title}
\usage{
mcf.plot(
  data.rdu,
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  plot.data = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  my.title = NULL,
  xlab = paste("Time in", get.time.units(data.rdu)),
  ylab = "Mean Cumulative Function",
  add = F,
  lty = 1,
  plot.seg = T,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  band.type = "p",
  ci.opt = "solid"
)
}
\arguments{
\item{data.rdu}{}

\item{conf.level}{}

\item{plot.data}{}

\item{title.option}{}

\item{my.title}{}

\item{xlab}{}

\item{ylab}{}

\item{add}{}

\item{lty}{}

\item{plot.seg}{}

\item{xlim}{}

\item{ylim}{}

\item{band.type}{}

\item{ci.opt}{}
}
\description{
Title
}
\examples{
\dontrun{

ValveSeat.rdu <- frame.to.rdu(valveseat,
                              ID.column = "unit", 
                              time.column = "days" ,
                              event.column = "event", 
                              data.title = "Valve-Seat Replacement Data",
                              time.units = "Days")

#attr(ValveSeat.rdu,"WindowInfo")

summary(ValveSeat.rdu)

event.plot(ValveSeat.rdu)

mcf.plot(ValveSeat.rdu)

}
}
