% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcf.R
\name{mcf}
\alias{mcf}
\title{Title}
\usage{
mcf(data.rdu, debug1 = F, kprint = 0)
}
\arguments{
\item{data.rdu}{}

\item{debug1}{}

\item{kprint}{}
}
\description{
Title
}
\examples{
\dontrun{

#time.column, ID.column, cost.count.column, event.column
WorkStation.rdu <- frame.to.rdu(workstation,
                                ID.column = "station",
                                time.column = "days",
                                event.column = "event")

#attr(WorkStation.rdu,"WindowInfo")

WorkStation.mcf <- mcf(WorkStation.rdu)

#sqrt(WorkStation.mcf$Var)

event.plot(WorkStation.rdu)

plot(WorkStation.mcf)

}
}
