% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metalwear}
\alias{metalwear}
\title{Metal alloy sliding wear resistance test}
\format{A \code{data.frame} with 96 rows and 4 variables:
\tabular{rlll}{
  [, 1] \tab microns \tab Specimen thickness \tab \bold{Numeric}\cr
  [, 2] \tab unit \tab Unit designator \tab \bold{Categoric}\cr
  [, 3] \tab cycles \tab Accumulated cycles at an observation \tab \bold{Numeric}\cr
  [, 4] \tab grams \tab Weight applied to the test specimen \tab \bold{Numeric}
   }}
\source{
Meeker, W. Q. and Escobar, L. A. (1998) Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons.
}
\description{
An experiment was conducted to test the sliding wear resistance of a particular metal alloy. 
             The data show changes in specimen thickness after being subjected to metal-to-metal sliding friction.
             Adding weight can increase the sliding frictional forces and cause the specimen to wear more quickly.
             A range of different weights were applied to (1) study the relationship between applied weight and wear 
             resistance and (2) gain a better understanding of the wear mechanism.
}
