% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfm.to.ld.R
\name{mfm.to.ld}
\alias{mfm.to.ld}
\title{Title}
\usage{
mfm.to.ld(
  data.ld,
  modes = NULL,
  prefix = NULL,
  allow.poor.data = T,
  do.assign = T
)
}
\arguments{
\item{data.ld}{}

\item{modes}{}

\item{prefix}{}

\item{allow.poor.data}{}

\item{do.assign}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1,
                                failure.mode.column = 2,
                                censor.column = 3, 
                                time.units = "Kilometers")
summary(ShockAbsorber.ld)
event.plot(ShockAbsorber.ld)

# Split out by failure mode

mleprobplot(ShockAbsorber.ld, 
            distribution = "Weibull")

mfmi.mleprobplot(ShockAbsorber.ld, 
                 distribution = "Weibull")

mfmc.mleprobplot(ShockAbsorber.ld, 
                 distribution = "Weibull")
                 
ShockAbsorber.mfld <- mfm.to.ld(ShockAbsorber.ld)

multiple.mleprobplot(ShockAbsorber.mfld,
                     data.ld.name="xx",
                     xlab="yy",
                     distribution="Weibull")

mleprobplot(ShockAbsorber.Mode1.ld, 
            distribution = "Weibull")

mleprobplot(ShockAbsorber.Mode2.ld,
            distribution = "Weibull")

get.time.vector(ShockAbsorber.Mode2.ld)
}
}
