% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfmc.mleprobplot.R
\name{mfmc.mleprobplot}
\alias{mfmc.mleprobplot}
\title{Title}
\usage{
mfmc.mleprobplot(
  data.ld,
  distribution,
  xlab = get.time.units(data.ld),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  time.range = c(NA, NA),
  time.vector.length = 20,
  dump = 1,
  grids = F,
  cex = 1.2,
  linear.axes = F,
  plot.censored.ticks = F,
  ci.list = NULL,
  cex.points = 1.2,
  shape = NULL,
  do.legend = "On plot",
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  band.type = "pointwise",
  col.ci = 4,
  lwd.ci = 2,
  show.individual = T,
  slope.axis = F,
  cexlab = 1.5,
  ...
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{xlab}{}

\item{ylab}{}

\item{xlim}{}

\item{ylim}{}

\item{time.range}{}

\item{time.vector.length}{}

\item{dump}{}

\item{grids}{}

\item{cex}{}

\item{linear.axes}{}

\item{plot.censored.ticks}{}

\item{ci.list}{}

\item{cex.points}{}

\item{shape}{}

\item{do.legend}{}

\item{conf.level}{}

\item{band.type}{}

\item{col.ci}{}

\item{lwd.ci}{}

\item{show.individual}{}

\item{slope.axis}{}

\item{cexlab}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

DeviceG.ld <- frame.to.ld(deviceg,
                          response.column = 1, 
                          failure.mode.column =  2)
                          
# Plots the individual failure modes separately

par(mfrow = c(1,2))
mfmi.mleprobplot(DeviceG.ld, distribution = "weibull") 
mfmi.mleprobplot(DeviceG.ld, distribution = "lognormal") 
par(mfrow = c(1,1))

# Plots both modes in one plot

mfmc.mleprobplot(DeviceG.ld, 
                 distribution = "Weibull") 

mfmc.mleprobplot(DeviceG.ld, 
                 distribution = "Weibull", 
                 band.type = "none")
                 
# different distributions for the different failure modes

mfmc.mleprobplot(DeviceG.ld, 
                 distribution = "Weibull", 
                 distribution.vec = c("Weibull","Lognormal"))
}
}
