% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfmi.mleprobplot.R
\name{mfmi.mleprobplot}
\alias{mfmi.mleprobplot}
\title{Title}
\usage{
mfmi.mleprobplot(
  data.ld,
  distribution,
  distribution.vec = rep(distribution, length = length(data.mfld)),
  the.failure.modes = names(data.mfld),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  xlab = get.time.units(data.ld),
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  time.range = c(NA, NA),
  dump = 1,
  grids = F,
  cex = 1.2,
  linear.axes = F,
  pch = (1:(length(the.failure.modes) + 1))[-2],
  lty = rep(1, length(the.failure.modes)),
  ci.list = NULL,
  lwd = rep(1, length(the.failure.modes)),
  slope.axis = F,
  plot.censored.ticks = F,
  trunc.correct = T,
  col.fhat.vec = (1:(length(the.failure.modes) + length(col.ci)))[-col.ci],
  col.ci = 4,
  shape = NULL,
  do.legend = NULL,
  ...
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{distribution.vec}{}

\item{the.failure.modes}{}

\item{ylab}{}

\item{conf.level}{}

\item{xlab}{}

\item{xlim}{}

\item{ylim}{}

\item{time.range}{}

\item{dump}{}

\item{grids}{}

\item{cex}{}

\item{linear.axes}{}

\item{pch}{}

\item{lty}{}

\item{ci.list}{}

\item{lwd}{}

\item{slope.axis}{}

\item{plot.censored.ticks}{}

\item{trunc.correct}{}

\item{col.fhat.vec}{}

\item{col.ci}{}

\item{shape}{}

\item{do.legend}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

DeviceG.ld <- frame.to.ld(deviceg,
                          response.column = 1, 
                          failure.mode.column =  2)
                          
# Plots the individual failure modes separately

par(mfrow = c(1,2))
mfmi.mleprobplot(DeviceG.ld, distribution = "weibull") 
mfmi.mleprobplot(DeviceG.ld, distribution = "lognormal") 
par(mfrow = c(1,1))

# Plots both modes in one plot

mfmc.mleprobplot(DeviceG.ld, 
                 distribution = "Weibull") 

mfmc.mleprobplot(DeviceG.ld, 
                 distribution = "Weibull", 
                 band.type = "none")
                 
# different distributions for the different failure modes

mfmc.mleprobplot(DeviceG.ld, 
                 distribution = "Weibull", 
                 distribution.vec = c("Weibull","Lognormal"))
}
}
