% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlehazplot.R
\name{mlehazplot}
\alias{mlehazplot}
\title{Title}
\usage{
mlehazplot(
  data.ld,
  distribution,
  gamthr = 0,
  xlab = get.time.units(data.ld),
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  time.vec = NULL,
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  my.title = NULL,
  cex = 1,
  sub.title = NULL,
  yaxis.line = 8,
  x.axis = "linear",
  y.axis = "linear",
  ylab = NULL,
  fits = F,
  grids = F,
  theta.start = c(NA, NA),
  parameter.fixed = NULL,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  print.parameters = T,
  band.type = "p",
  plotem = T,
  param.loc = "bottomright",
  ...
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{gamthr}{}

\item{xlab}{}

\item{xlim}{}

\item{ylim}{}

\item{time.vec}{}

\item{conf.level}{}

\item{my.title}{}

\item{cex}{}

\item{sub.title}{}

\item{yaxis.line}{}

\item{x.axis}{}

\item{y.axis}{}

\item{ylab}{}

\item{fits}{}

\item{grids}{}

\item{theta.start}{}

\item{parameter.fixed}{}

\item{title.option}{}

\item{print.parameters}{}

\item{band.type}{}

\item{plotem}{}

\item{param.loc}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1, 
                                censor.column = 3,
                                time.units = "Kilometers")

mlehazplot(ShockAbsorber.ld, 
           distribution = "Weibull", 
           param.loc = "topleft")

mlehazplot(ShockAbsorber.ld, 
           distribution = "Frechet")


mlehazplot(ShockAbsorber.ld, 
           distribution = "Lognormal")


mlehazplot(ShockAbsorber.ld, 
           distribution = "Lognormal",
           xlim = c(5000,500000), 
           y.axis = "log")


mlehazplot(ShockAbsorber.ld, 
           distribution = "Weibull",
           time.vec = c(10000,20000,30000))

mlehazplot(ShockAbsorber.ld, 
           distribution = "Weibull",
           time.vec = c(10000,20000,30000),
           parameter.fixed = c(F,T),
           theta.start = c(9.,2))
}
}
