% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleprobplot.R
\name{mleprobplot}
\alias{mleprobplot}
\title{Title}
\usage{
mleprobplot(
  data.ld,
  distribution,
  gamthr = 0,
  xlab = NULL,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  time.range = c(NA, NA),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  my.title = NULL,
  cex = 1,
  cex.labs = 1.1,
  cex.axis = 1.1,
  sub.title = NULL,
  grids = F,
  linear.axes = F,
  slope.axis = F,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  trunc.correct = T,
  add = F,
  cex.points = 1.2,
  theta.start = c(NA, NA),
  parameter.fixed = NULL,
  plot.censored.ticks = F,
  pch = 16,
  print.parameters = T,
  band.type = "p",
  debug1 = F,
  lwd.ci = 3,
  lwd.fhat = 2,
  band.method = GetSMRDDefault("SMRD.ConfidenceBandMethod"),
  mono.tran = T,
  title.line.adj,
  param.loc = "bottomright",
  ...
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{gamthr}{}

\item{xlab}{}

\item{xlim}{}

\item{ylim}{}

\item{time.range}{}

\item{conf.level}{}

\item{my.title}{}

\item{cex}{}

\item{cex.labs}{}

\item{cex.axis}{}

\item{sub.title}{}

\item{grids}{}

\item{linear.axes}{}

\item{slope.axis}{}

\item{title.option}{}

\item{ylab}{}

\item{trunc.correct}{}

\item{add}{}

\item{cex.points}{}

\item{theta.start}{}

\item{parameter.fixed}{}

\item{plot.censored.ticks}{}

\item{pch}{}

\item{print.parameters}{}

\item{band.type}{}

\item{debug1}{}

\item{lwd.ci}{}

\item{lwd.fhat}{}

\item{band.method}{}

\item{mono.tran}{}

\item{title.line.adj}{}

\item{param.loc}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1,
                                failure.mode.column = 2,
                                censor.column = 3, 
                                time.units = "Kilometers")
summary(ShockAbsorber.ld)
event.plot(ShockAbsorber.ld)

# Split out by failure mode

mleprobplot(ShockAbsorber.ld, 
            distribution = "Weibull")

mfmi.mleprobplot(ShockAbsorber.ld, 
                 distribution = "Weibull")

mfmc.mleprobplot(ShockAbsorber.ld, 
                 distribution = "Weibull")
                 
ShockAbsorber.mfld <- mfm.to.ld(ShockAbsorber.ld)

multiple.mleprobplot(ShockAbsorber.mfld,
                     data.ld.name="xx",
                     xlab="yy",
                     distribution="Weibull")

mleprobplot(ShockAbsorber.Mode1.ld, 
            distribution = "Weibull")

mleprobplot(ShockAbsorber.Mode2.ld,
            distribution = "Weibull")

get.time.vector(ShockAbsorber.Mode2.ld)
}
}
