% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlest.R
\name{mlest}
\alias{mlest}
\title{Title}
\usage{
mlest(data.ld, distribution, relationship.parameters = 1,
  explan.vars = NULL, gamthr = 0, theta.start = NULL,
  parameter.fixed = NULL, intercept = T, kprint = 0, maxit = 500,
  debug1 = F, genforce = F, ...)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{relationship.parameters}{}

\item{explan.vars}{}

\item{gamthr}{}

\item{theta.start}{}

\item{parameter.fixed}{}

\item{intercept}{}

\item{kprint}{}

\item{maxit}{}

\item{debug1}{}

\item{genforce}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

berkson200.ld <- frame.to.ld(berkson200,
                             response.column = c(1,2),
                             censor.column = 3,
                             case.weight.column = 4,
                             time.units = "1/5000 Seconds")

summary(berkson200.ld)

plot(berkson200.ld)
plot(berkson200.ld, dist = "Exponential")

cdfest(berkson200.ld)

berkson200.mle.exp <- mlest(berkson200.ld, 
                            distribution = "Exponential")

print(berkson200.mle.exp)

doatrun.ld <- frame.to.ld(doatrun,
                          response.column = c(1,2),
                          censor.column = 3, 
                          case.weight.column = 4,
                          truncation.response.column = 5, 
                          truncation.type.column = 6, 
                          data.title = "DOA Truncated Data", 
                          time.units = "Hours")

summary(doatrun.ld)

cdfest(doatrun.ld)

doatrun.mle.weib <- mlest(doatrun.ld,"Weibull")

print(doatrun.mle.weib)
}
}
