% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mlmod1}
\alias{mlmod1}
\alias{MLMOD1}
\title{Meeker-LuValle Model 1}
\usage{
MLMOD1(times, number_times, a2_init, a2_limit, rate, rate_factor, a2)
}
\description{
Compute growth of filament according 
             to Meeker-LuValle Model 1 different 
             rate parameter at each time interval
}
