% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MLSIM2}
\alias{MLSIM2}
\title{method 2}
\usage{
MLSIM2(
  x,
  y,
  cen,
  wt,
  nrow,
  nter,
  ny,
  nty,
  ty,
  tcodes,
  kdist,
  gamthr,
  lfix,
  nparm,
  intcpt,
  escale,
  e,
  maxit,
  kprint,
  dscrat,
  iscrat,
  devian,
  thetah,
  fsder,
  vcv,
  r,
  res,
  fv,
  theta,
  iarray,
  marray,
  wtnew,
  xnew,
  ynew,
  iret,
  retmat,
  numsim,
  numret,
  tspass,
  lrand,
  iersim
)
}
\description{
nonparametric          parametric
}
\details{
method 2 simulation---nonparametric sampling/parametric inference

must send down a complete data set

   parameter meanings same as in wqm_mlboth except the following:

   can get rid of scalars like iervcv,ierfit,xlike,nobs

theta(nparm)   true value of theta

thetah(nparm)     space for thetahat

iarray(marray)    scratch  space needed to generate random weights

marray            length of scratch space. must be at least equal
                 in length to the number of rows in wt
                 a computationally more efficient method is used
                 if the number is larger than the sum of the weights

wtnew             vector for the random weights for the simulation

xnew              hold space for x

ynew             hold space for y

iret    amount of stuff to return

retmat(numret,numsim)  matrix to return the results, one result per col

numsim  number of simulations to be run

nrowr   number of rows being returned in retmat

tspass(33)  should be able to eliminate

lrand     should be able to eliminate


iersim  data space too small

  need to send down data space big enough to cover all data situations
}
