% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.mleprobplot.R
\name{multiple.mleprobplot}
\alias{multiple.mleprobplot}
\title{Title}
\usage{
multiple.mleprobplot(data.list, data.ld.name, distribution,
  distribution.vec = rep(distribution, length = length(data.list)), xlab,
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"),
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  check.level = SMRDOptions("SMRD.DataCheck"), xlim = c(NA, NA),
  ylim = c(NA, NA), time.range = c(NA, NA), dump = 1, grids = F,
  my.title = NULL, cex = 1.2, linear.axes = F, slope.axis = F,
  pch = (1:(length(do.list) + 1))[-2], lty = rep(1, length(do.list)),
  ci.list = NULL, lwd = rep(1, length(do.list)), trunc.correct = T,
  col.fhat.vec = (1:(length(do.list) + length(col.ci)))[-col.ci],
  col.ci = 4, shape = NULL, do.legend = NULL, add.title = NULL,
  do.list = names(data.list), plotem = rep(T, length(do.list)),
  plot.censored.ticks = F, ylim.data = NULL, plot.np = T,
  plot.frame = T, debug1 = F, ...)
}
\arguments{
\item{data.list}{}

\item{data.ld.name}{}

\item{distribution}{}

\item{distribution.vec}{}

\item{xlab}{}

\item{ylab}{}

\item{conf.level}{}

\item{check.level}{}

\item{xlim}{}

\item{ylim}{}

\item{time.range}{}

\item{dump}{}

\item{grids}{}

\item{my.title}{}

\item{cex}{}

\item{linear.axes}{}

\item{slope.axis}{}

\item{pch}{}

\item{lty}{}

\item{ci.list}{}

\item{lwd}{}

\item{trunc.correct}{}

\item{col.fhat.vec}{}

\item{col.ci}{}

\item{shape}{}

\item{do.legend}{}

\item{add.title}{}

\item{do.list}{}

\item{plotem}{}

\item{plot.censored.ticks}{}

\item{ylim.data}{}

\item{plot.np}{}

\item{plot.frame}{}

\item{debug1}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1,
                                failure.mode.column = 2,
                                censor.column = 3, 
                                time.units = "Kilometers")
summary(ShockAbsorber.ld)
event.plot(ShockAbsorber.ld)

# Split out by failure mode

mleprobplot(ShockAbsorber.ld, 
            distribution = "Weibull")

mfmi.mleprobplot(ShockAbsorber.ld, 
                 distribution = "Weibull")

mfmc.mleprobplot(ShockAbsorber.ld, 
                 distribution = "Weibull")
                 
ShockAbsorber.mfld <- mfm.to.ld(ShockAbsorber.ld)

multiple.mleprobplot(ShockAbsorber.mfld,
                     data.ld.name="xx",
                     xlab="yy",
                     distribution="Weibull")

mleprobplot(ShockAbsorber.Mode1.ld, 
            distribution = "Weibull")

mleprobplot(ShockAbsorber.Mode2.ld,
            distribution = "Weibull")

get.time.vector(ShockAbsorber.Mode2.ld)
}
}
