% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.probplot.sim.R
\name{multiple.probplot.sim}
\alias{multiple.probplot.sim}
\title{Title}
\usage{
multiple.probplot.sim(
  distribution = "normal",
  parameter = 1,
  sample.size.vec = c(10, 20, 40),
  number.simulations = 5,
  my.title = NULL
)
}
\arguments{
\item{distribution}{}

\item{parameter}{}

\item{sample.size.vec}{}

\item{number.simulations}{}

\item{my.title}{}
}
\description{
Title
}
\examples{
\dontrun{

multiple.probplot.sim()

multiple.probplot.sim(dist = "exponential")

}
}
