% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.profile.plot.R
\name{multiple.profile.plot}
\alias{multiple.profile.plot}
\title{Title}
\usage{
multiple.profile.plot(grid.list, which, lty = (1:(length(grid.list) +
  1))[-2], col = 1:length(grid.list), do.legend = "On plot",
  do.list = names(grid.list))
}
\arguments{
\item{grid.list}{}

\item{which}{}

\item{lty}{}

\item{col}{}

\item{do.legend}{}

\item{do.list}{}
}
\description{
Title
}
\examples{
\dontrun{

Snubber.ld <- frame.to.ld(snubber, 
                          response.column = "cycles", 
                          censor.column = "event",
                          time.units = "Cycles",
                          case.weight.column = "count",
                          x.columns = "design")

event.plot(Snubber.ld)
summary(Snubber.ld)

Snubber.groupi.nor.out <-  groupi.mleprobplot(Snubber.ld,"normal")

tmpxx <- groupi.contour(Snubber.ld,
                        "Weibull",
                        the.quantile = 0.1)

tmpxx <- groupi.contour(Snubber.ld,
                        "lognormal",
                        the.quantile = 0.1)

multiple.profile.plot(tmpxx, which = "x")
multiple.profile.plot(tmpxx, which = "y")

}
}
