% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple.simulate.type.two.R
\name{multiple.simulate.type.two}
\alias{multiple.simulate.type.two}
\title{Title}
\usage{
multiple.simulate.type.two(n, r, distribution, number.sim = 1000)
}
\arguments{
\item{n}{}

\item{r}{}

\item{distribution}{}

\item{number.sim}{}
}
\description{
Title
}
\examples{
\dontrun{

mstt1 <- multiple.simulate.type.two(n = c(20,30,40), 
                                    r = c(20,30,40),
                                    distribution = "lognormal",
                                    number.sim = 1000)

plot(mstt1, qprob = 0.9)
plot(mstt1, qprob = 0.8)
plot(mstt1, qprob = 0.5)


mstt2 <- multiple.simulate.type.two(n = c(20,30,40), 
                                    r = c(10,15,20),
                                    distribution = "lognormal",
                                    number.sim = 1000)

plot(mstt2, qprob = 0.9)
plot(mstt2, qprob = 0.8)
plot(mstt2, qprob = 0.5, grids = T)

mstt3 <- multiple.simulate.type.two(n = c(24,28,33), 
                                    r = c(24,28,33),
                                    distribution = "normal", 
                                    number.sim = 1000)

plot(mstt3, qprob = 0.9)
}
}
