% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiplefail}
\alias{multiplefail}
\title{#####}
\format{
A \code{data.frame} with 40 rows and 7 variables:
\tabular{rlll}{
  [, 1] \tab celsius \tab Temperature applied \tab \bold{Numeric}\cr
  [, 2] \tab turn_hours \tab Time of an observed turn event \tab \bold{Numeric}\cr
  [, 3] \tab turn_event \tab Type of turn event at \code{turn_hours} (failed/right-censored) \tab \bold{Categoric}\cr
  [, 4] \tab phase_hours \tab Time of an observed phase event \tab \bold{Numeric}\cr
  [, 5] \tab phase_event \tab Type of phase event at \code{phase_hours} (failed/right-censored) \tab \bold{Categoric}\cr
  [, 6] \tab ground_hours \tab Time of an observed ground event \tab \bold{Numeric}\cr
  [, 7] \tab ground_event \tab Type of ground event at \code{ground_hours} (failed/right-censored) \tab \bold{Categoric}
   }
}
\source{
Nelson (1990)
}
\description{
####
}
\seealso{
Other data-notdone: 
\code{\link{adhesivebondc}},
\code{\link{adhesivestrength}},
\code{\link{amsaaexactfail}},
\code{\link{appliancea}},
\code{\link{applianceb}},
\code{\link{aprel72}},
\code{\link{atrazinejune}},
\code{\link{bearinga}},
\code{\link{censortruncationtest}},
\code{\link{chemicalprocess}},
\code{\link{cirpack4}},
\code{\link{cirpack5}},
\code{\link{componentd}},
\code{\link{concrete}},
\code{\link{connectionstrength}},
\code{\link{customerlife}},
\code{\link{devicedconnected}},
\code{\link{deviceh}},
\code{\link{devicen}},
\code{\link{diskber}},
\code{\link{doatrun}},
\code{\link{electromech}},
\code{\link{engineemissions}},
\code{\link{epoxyweathering}},
\code{\link{filament}},
\code{\link{hcfdata}},
\code{\link{heater}},
\code{\link{hpcrepairs}},
\code{\link{icdevice1}},
\code{\link{inconel}},
\code{\link{insulation}},
\code{\link{laminatepanel}},
\code{\link{largeball}},
\code{\link{locomotivecontrol}},
\code{\link{luminosity}},
\code{\link{machineh}},
\code{\link{mechanicalswitch}},
\code{\link{newspring}},
\code{\link{parta}},
\code{\link{piccioto}},
\code{\link{pipelinethickness}},
\code{\link{pmax}()},
\code{\link{prob3_5}},
\code{\link{riverchem}},
\code{\link{shelflifea}},
\code{\link{testdadtplan}},
\code{\link{titanium2}},
\code{\link{titanium}},
\code{\link{tractorbreaks}},
\code{\link{turbinedevice}},
\code{\link{vehiclemotor}},
\code{\link{voltageendurance}},
\code{\link{workstation}}
}
\concept{data-notdone}
