% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newspring}
\alias{newspring}
\title{High-temp gas spring accelerated test}
\format{A \code{data.frame} with 80 rows and 6 variables:
\tabular{rlll}{
  [, 1] \tab kilocycles \tab Accumulated cycles at \code{event} (in thousands) \tab \bold{Numeric}\cr
  [, 2] \tab centimeters \tab Spring stroke length \tab \bold{Numeric}\cr
  [, 3] \tab fahrenheit \tab Temperature applied to the spring \tab \bold{Numeric}\cr
  [, 4] \tab method \tab Manufacturing method (New/Old) \tab \bold{Categoric}\cr
  [, 5] \tab event \tab Event observed at \code{kcycles} (Suspended/Failed) \tab \bold{Categoric}\cr
  [, 6] \tab count \tab Number of events observed at \code{kcycles} \tab \bold{Numeric}
  }}
\source{
Meeker, W. Q. (1999) A factorial experiment to compare the lifetimes of springs as a function of a processing temperature and amount of displacement in the spring test (Unpublished)
}
\description{
####
}
\seealso{
Other data-notdone: \code{\link{adhesivebondc}},
  \code{\link{adhesivestrength}},
  \code{\link{amsaaexactfail}}, \code{\link{appliancea}},
  \code{\link{applianceb}}, \code{\link{aprel72}},
  \code{\link{atrazinejune}}, \code{\link{bearinga}},
  \code{\link{censortruncationtest}},
  \code{\link{chemicalprocess}}, \code{\link{cirpack4}},
  \code{\link{cirpack5}}, \code{\link{componentd}},
  \code{\link{concrete}}, \code{\link{connectionstrength}},
  \code{\link{customerlife}},
  \code{\link{devicedconnected}}, \code{\link{deviceh}},
  \code{\link{devicen}}, \code{\link{diskber}},
  \code{\link{doatrun}}, \code{\link{electromech}},
  \code{\link{engineemissions}},
  \code{\link{epoxyweathering}}, \code{\link{filament}},
  \code{\link{hcfdata}}, \code{\link{heater}},
  \code{\link{hpcrepairs}}, \code{\link{icdevice1}},
  \code{\link{inconel}}, \code{\link{insulation}},
  \code{\link{laminatepanel}}, \code{\link{largeball}},
  \code{\link{locomotivecontrol}},
  \code{\link{luminosity}}, \code{\link{machineh}},
  \code{\link{mechanicalswitch}},
  \code{\link{multiplefail}}, \code{\link{parta}},
  \code{\link{pipelinethickness}}, \code{\link{pmax}},
  \code{\link{prob3_5}}, \code{\link{riverchem}},
  \code{\link{shelflifea}}, \code{\link{testdadtplan}},
  \code{\link{titanium2}}, \code{\link{titanium}},
  \code{\link{tractorbreaks}}, \code{\link{turbinedevice}},
  \code{\link{vehiclemotor}},
  \code{\link{voltageendurance}}, \code{\link{workstation}}
}
\concept{data-notdone}
