% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHPP.mle.R
\name{NHPP.mle}
\alias{NHPP.mle}
\title{Title}
\usage{
NHPP.mle(
  data.rdu,
  form = "power rule",
  debug1 = 0,
  theta.start = c(NA, NA),
  kprint = 0,
  ...
)
}
\arguments{
\item{data.rdu}{}

\item{form}{}

\item{debug1}{}

\item{theta.start}{}

\item{kprint}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")

NHPP.mle(halfbeak.rdu, form = "power rule")
NHPP.mle(halfbeak.rdu, form = "log linear")

}
}
