% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nicdbattery}
\alias{nicdbattery}
\title{Accelerated test of spacecraft battery cell data}
\format{A \code{data.frame} with 87 rows and 10 variables:
\tabular{rlll}{
  [, 1] \tab celsius \tab Temperature applied to the battery \tab \bold{Numeric}\cr
  [, 2] \tab discharge_depth \tab Depth of discharge (as percent) \tab \bold{Numeric}\cr
  [, 3] \tab discharge_time \tab Discharge time (in hours) \tab \bold{Numeric}\cr
  [, 4] \tab charge \tab Charge time (in hours) \tab \bold{Numeric}\cr
  [, 5] \tab recharge \tab Level of recharge \tab \bold{Numeric}\cr
  [, 6] \tab koh_percent \tab Concentration of Potassium Hydroxide (as percent) \tab \bold{Numeric}\cr
  [, 7] \tab koh_volume \tab Volume of Potassium Hydroxide (in cubic-centimeters) \tab \bold{Numeric}\cr
  [, 8] \tab precharge \tab Precharge time (in hours) \tab \bold{Numeric}\cr
  [, 9] \tab cycles \tab Accumulated cycles at \code{event} \tab \bold{Numeric}\cr
  [,10] \tab event \tab Event observed at \code{cycles} (failure/right-censored)  \tab \bold{Categoric}
   }}
\source{
Brown, H. M., and Mains D. E (1979)
        Accelerated Test Program for Sealed Nickel--Cadmium Spacecraft Batteries/Cells,
        Technical Report WQEC/C 79-145. Available from the Department of the Navy, 
        Naval Weapons Support Center, Weapons Quality Engineering Center, Crane, IN 47522.
}
\description{
Brown and Mains (1979) present the results of an extensive experiment to evaluate 
             the long-term performance of rechargable nickel-cadmium battery cells that 
             were to be used in spacecraft. The study used 8 experimental factors. The 
             first five factors are environmental or accelerating factors (set to higher 
             than usual levels to obtain failure information more quickly).  The other 
             three factors were product-design factors that could be adjusted in the product
             design to optimize performance and reliability of the batteries to be manufactured.  
             The experiment ran 82 batteries, each containing 5 individual cells.  Each battery 
             was tested at a combination of factor levels determined according to a central 
             composite experimental plan.
}
