% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnnc0.R
\name{nnnc0}
\alias{nnnc0}
\title{Flag if a character object is either NA, NULL, or has zero characters}
\usage{
nnnc0(object)
}
\arguments{
\item{object}{An one-dimensional character-class object}
}
\value{
Returns \code{TRUE} If \code{is.na(object)}, \code{is.na(object)}, or \code{nchar(object) == 0}
        otherwise returns \code{FALSE}
}
\description{
Flag if a character object is either NA, NULL, or has zero characters
}
