% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{nonparametric.bootstrap}
\alias{nonparametric.bootstrap}
\title{Title}
\usage{
nonparametric.bootstrap(data.ld, number.sim = 10, kprint = 0,
  maxit = 500, max.sim.scratch.space = 1000, maxmsd = 100,
  debug1 = F, randomize = T)
}
\arguments{
\item{data.ld}{}

\item{number.sim}{}

\item{kprint}{}

\item{maxit}{}

\item{max.sim.scratch.space}{}

\item{maxmsd}{}

\item{debug1}{}

\item{randomize}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1,
                                censor.column = 3,
                                time.units = "Kilometers")

ShockAbsorber.boot.np<- nonparametric.bootstrap(ShockAbsorber.ld,
                                                number.sim = 20)

plot(ShockAbsorber.boot.np)

summary(ShockAbsorber.boot.np, 
        compare = T)
 
}
}
