% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NPSYS}
\alias{NPSYS}
\title{Compute the nonparametric estimate of the
 mean cumulative occurence rate of failures
 and corresponding estimate of the var if the estimator}
\usage{
NPSYS(
  itime,
  isys,
  icodes,
  weight,
  nobs,
  nfailt,
  nsys,
  utime,
  iusys,
  ctime,
  dead,
  idelta,
  dsum,
  idlsum,
  dbar,
  varsum,
  xmuhat,
  varxmu,
  kprint
)
}
\description{
inputs:
        itime(nobs)       time indices   (from 1,2,..., nfailt)
                          gives index pointing to the vector of unique times
        isys(nobs)         system indices (from 1,2,..., nsys)
                          gives index pointing to the vector of unique sys ids
        icodes(nobs)       censor/failure indicator (1 fail, 2 censor)
        weight(nobs)       weight (multiplicity) or cost of failure
        nobs               number of observations in data set
        nfailt             number of unique failure times
        nsys               number of unique system ids
        utime(nfailt)     list of unique times
        iusys(nsys)        list of unique system ids
}
\details{
ctime(nsys),varsum(nsys) scratch arrays
        dead(nfailt,nsys),idelta(nfailt,nsys) scratch arrays
        dsum(nfailt),idlsum(nfailt),dbar(nfailt)   scratch arrays


  outputs:
        xmuhat(nfailt)   estimate of mu at each failure time
        varxmu(nfailt)   estimate of var(muhat) at each failure time
}
