% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.dim.profile.R
\name{one.dim.profile}
\alias{one.dim.profile}
\title{Title}
\usage{
one.dim.profile(
  gmle.out,
  profile.setup = NULL,
  profile.stable.parameters = NULL,
  profile.on.list = 1:length(theta.hat),
  special.stuff.profile = NULL,
  range.list = NULL,
  size = 40,
  interactive = F,
  save.structures = T,
  addname = NULL,
  save.parameter.vectors = F,
  debug1 = 0,
  plot.em = T,
  print.ci = TRUE
)
}
\arguments{
\item{gmle.out}{}

\item{profile.setup}{}

\item{profile.stable.parameters}{}

\item{profile.on.list}{}

\item{special.stuff.profile}{}

\item{range.list}{}

\item{size}{}

\item{interactive}{}

\item{save.structures}{}

\item{addname}{}

\item{save.parameter.vectors}{}

\item{debug1}{}

\item{plot.em}{}

\item{print.ci}{}
}
\description{
Title
}
\examples{
\dontrun{

R4490.rdu <- frame.to.rdu(r4490,
                          ID.column = "vin",
                          time.column = "days" , 
                          cost.count.column = "costcount" ,
                          event.column = "code")

attr(R4490.rdu,"WindowInfo")

event.plot(R4490.rdu)
R4490.mcf <- mcf(R4490.rdu)
plot(R4490.mcf)

R4490.nhpp.out <- PlotMCFandNHPP(R4490.rdu, form = "power rule")
one.dim.profile(R4490.nhpp.out,size = 5,save.s = T)
two.dim.profile(R4490.nhpp.out, 
                profile.on.list = NULL,
                which = c(1,2), 
                size = c(5,5))

profile.contour(R4490.nhpp.outstruct1x2,
                transformationy = "log",
                variable.namey = "sigma", 
                variable.namex = "mu", 
                v = c(0.001, 0.01, .1,0.2, 0.4, 0.7, 0.9) )

}
}
