% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.effect.R
\name{parallel.effect}
\alias{parallel.effect}
\title{Title}
\usage{
parallel.effect(
  xrange = c(0.5, 1),
  yrange = c(0.8, 1),
  ncomp = c(1, 2, 3, 4, 5)
)
}
\arguments{
\item{xrange}{}

\item{yrange}{}

\item{ncomp}{}
}
\description{
Title
}
\examples{
\dontrun{
component.effect(parallel = TRUE)

parallel.effect()

parallel.dep.effect()

series.dep.effect()

}
}
