% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{parametric.bootstrap}
\alias{parametric.bootstrap}
\title{Title}
\usage{
parametric.bootstrap(
  data.ld,
  distribution,
  number.sim,
  escale = 10000,
  e = rep(1e-04, number.parameters),
  parameter.fixed = rep(F, number.parameters),
  intercept = T,
  kprint = 0,
  maxit = 500,
  max.sim.scratch.space = 1000,
  debug1 = F,
  randomize = T
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{number.sim}{}

\item{escale}{}

\item{e}{}

\item{parameter.fixed}{}

\item{intercept}{}

\item{kprint}{}

\item{maxit}{}

\item{max.sim.scratch.space}{}

\item{debug1}{}

\item{randomize}{}
}
\description{
Title
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1,
                                censor.column = 3,
                                time.units = "Kilometers")

ShockAbsorber.boot.p <- parametric.bootstrap(ShockAbsorber.ld,
                                             distribution = "Weibull",
                                             number.sim = 2000)

plot(ShockAbsorber.boot.p)
plot(ShockAbsorber.boot.p, 
     simulate.parameters = TRUE, 
     parameter.sims = 500)

summary(ShockAbsorber.boot.p,
        inference.on = "parameter", 
        which = 1)

summary(ShockAbsorber.boot.p,
        inference.on = "parameter", 
        which = 2,
        do.compare = T)

summary(ShockAbsorber.boot.p,
        inference.on = "parameter", 
        which = 2)

summary(ShockAbsorber.boot.p,
        inference.on = "quantile",
        which = 0.1)

summary(ShockAbsorber.boot.p,
        inference.on = "probability", 
        which = 1000)

}
}
