% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{piccioto}
\alias{piccioto}
\title{Tensile Fatigue Test of Polyester/Viscose Yarn}
\format{A \code{data.frame} with 797 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab inches \tab Length of the test specimen \tab \bold{Numeric}\cr
  [, 2] \tab kilocycles \tab Accumulated cycles at \code{event} (in thousands) \tab \bold{Numeric}\cr
  [, 3] \tab event \tab Event observed at \code{kilocycles} (failed/right-censored) \tab \bold{Categoric}
   }}
\source{
Picciotto R., (1970) 
        Tensile Fatigue Characteristics of a Sized Polyester/Viscose Yarn and Their Effect on Weaving Performance. 
        A thesis submitted to the Graduate Faculty of North Carolina State University at Raleigh in partial fulfillment of the requirements for the Degree of Master of Science. 
        Department of Textile Technology.
}
\description{
Picciotto includes the \code{length} and \code{kcycles} columns.  picciotto2 is a subset of the picciotto data set, including only the units for which \code{length = 30,60, or 90}.  picciotto3 mirrors picciotto2 but assumes that the test conlcuded at 100 kilocycles
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grampus}},
  \code{\link{grids1}}, \code{\link{halfbeak}},
  \code{\link{heatexchanger}}, \code{\link{lfp1370}},
  \code{\link{lfptrun100}}, \code{\link{lzbearing}},
  \code{\link{metalwear}}, \code{\link{mylarsub}},
  \code{\link{nicdbattery}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
