% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nonparametric.estimate.R
\name{plot.nonparametric.estimate}
\alias{plot.nonparametric.estimate}
\title{Title}
\usage{
\method{plot}{nonparametric.estimate}(x, cdpoints.out, distribution,
  log.of.data, how.show.fhat, shape, point.cex, pch, point.pch, xlim,
  col.points = 1, plot.vert.line = T, cex.axis = 1.1,
  cex.lab = 1.1, ...)
}
\arguments{
\item{x}{}

\item{cdpoints.out}{}

\item{distribution}{}

\item{log.of.data}{}

\item{how.show.fhat}{}

\item{shape}{}

\item{point.cex}{}

\item{pch}{}

\item{point.pch}{}

\item{xlim}{}

\item{col.points}{}

\item{plot.vert.line}{}

\item{cex.axis}{}

\item{cex.lab}{}

\item{...}{}
}
\description{
Title
}
