% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.plan.values.R
\name{plot.plan.values}
\alias{plot.plan.values}
\title{Plot reliability life test}
\usage{
\method{plot}{plan.values}(x, censor.time = NULL, my.title, grids = 2,
  quantile.mark = NULL, number.points = 20, perc.low = 0.01,
  perc.high = 0.99, cex = 1, xlim = NULL, ylim = NULL,
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  xlab = x$time.units, digits = GetSMRDDefault("SMRD.DigitsPrinted"),
  ...)
}
\arguments{
\item{x}{}

\item{censor.time}{}

\item{my.title}{}

\item{grids}{}

\item{quantile.mark}{}

\item{number.points}{}

\item{perc.low}{}

\item{perc.high}{}

\item{cex}{}

\item{xlim}{}

\item{ylim}{}

\item{title.option}{}

\item{xlab}{}

\item{digits}{}

\item{...}{}
}
\description{
Plot reliability life test
}
