% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMCFandNHPP.R
\name{PlotMCFandNHPP}
\alias{PlotMCFandNHPP}
\title{Title}
\usage{
PlotMCFandNHPP(
  data.rdu,
  form,
  xlab = paste("Age in", get.time.units(data.rdu)),
  ylab = "Mean Cumulative Failures",
  plot.seg = T,
  number.points = 200,
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  band.type = "p",
  theta.start = c(NA, NA),
  my.title = NULL
)
}
\arguments{
\item{data.rdu}{}

\item{form}{}

\item{xlab}{}

\item{ylab}{}

\item{plot.seg}{}

\item{number.points}{}

\item{conf.level}{}

\item{band.type}{}

\item{theta.start}{}

\item{my.title}{}
}
\description{
Title
}
\examples{
\dontrun{

Cylinder.rdu <- frame.to.rdu(cylinder,
                             ID.column = "engine", 
                             time.column = "days", 
                             event.column = "event", 
                             cost.count.column = "count", 
                             data.title = "Cylinder Replacement Data",
                             time.units = "Days")

#attr(Cylinder.rdu,"WindowInfo")

PlotMCFandNHPP(Cylinder.rdu, form = "power rule")

Cylinder.mcf <- mcf(Cylinder.rdu)

plot(Cylinder.mcf)

}
}
