% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRiskSet.R
\name{plotRiskSet}
\alias{plotRiskSet}
\title{Title}
\usage{
plotRiskSet(data.rdu, xlab = paste("Time in", get.time.units(data.rdu)),
  ylab = ifelse(proportion, "Proportion at Risk", "Number in Risk Set"),
  proportion = F, plot.average = T, my.title = NULL, JustEvent = F)
}
\arguments{
\item{data.rdu}{}

\item{xlab}{}

\item{ylab}{}

\item{proportion}{}

\item{plot.average}{}

\item{my.title}{}

\item{JustEvent}{}
}
\description{
Title
}
\examples{
\dontrun{

AMSAAExactFail.rdu <- frame.to.rdu(amsaaexactfail,
                                   ID.column = "vehicle", 
                                   time.column = "miles" , 
                                   event.column = "event")

plotRiskSet(AMSAAExactFail.rdu,proportion = T)

}
}
