% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pmax}
\alias{pmax}
\title{#####}
\format{
A \code{data.frame} with 170 rows and 9 variables:
\tabular{rlll}{
  [, 1] \tab penid \tab Identification designator \tab \bold{Categoric}\cr
  [, 2] \tab batch \tab Batch indicator \tab \bold{Categoric}\cr
  [, 3] \tab celsius \tab Temperature applied \tab \bold{Numeric}\cr
  [, 4] \tab days \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 5] \tab sample \tab Type of sample tested \tab \bold{Categoric}\cr
  [, 6] \tab wafer \tab Type of wafer \tab \bold{Categoric}\cr
  [, 7] \tab pmax \tab Maximum value of p observed \tab \bold{Numeric}\cr
  [, 8] \tab locus \tab Type of locus \tab \bold{Categoric}\cr
  [, 9] \tab event \tab Event observed at \code{days} (failure/right-censored) \tab \bold{Categoric}
   }
}
\source{
####
}
\description{
####
}
\seealso{
Other data-notdone: 
\code{\link{adhesivebondc}},
\code{\link{adhesivestrength}},
\code{\link{amsaaexactfail}},
\code{\link{appliancea}},
\code{\link{applianceb}},
\code{\link{aprel72}},
\code{\link{atrazinejune}},
\code{\link{bearinga}},
\code{\link{censortruncationtest}},
\code{\link{chemicalprocess}},
\code{\link{cirpack4}},
\code{\link{cirpack5}},
\code{\link{componentd}},
\code{\link{concrete}},
\code{\link{connectionstrength}},
\code{\link{customerlife}},
\code{\link{devicedconnected}},
\code{\link{deviceh}},
\code{\link{devicen}},
\code{\link{diskber}},
\code{\link{doatrun}},
\code{\link{electromech}},
\code{\link{engineemissions}},
\code{\link{epoxyweathering}},
\code{\link{filament}},
\code{\link{hcfdata}},
\code{\link{heater}},
\code{\link{hpcrepairs}},
\code{\link{icdevice1}},
\code{\link{inconel}},
\code{\link{insulation}},
\code{\link{laminatepanel}},
\code{\link{largeball}},
\code{\link{locomotivecontrol}},
\code{\link{luminosity}},
\code{\link{machineh}},
\code{\link{mechanicalswitch}},
\code{\link{multiplefail}},
\code{\link{newspring}},
\code{\link{parta}},
\code{\link{piccioto}},
\code{\link{pipelinethickness}},
\code{\link{prob3_5}},
\code{\link{riverchem}},
\code{\link{shelflifea}},
\code{\link{testdadtplan}},
\code{\link{titanium2}},
\code{\link{titanium}},
\code{\link{tractorbreaks}},
\code{\link{turbinedevice}},
\code{\link{vehiclemotor}},
\code{\link{voltageendurance}},
\code{\link{workstation}}
}
\concept{data-notdone}
