% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{POSTKP}
\alias{POSTKP}
\title{Compute empirical the predictive distributions 
 (pdf, cdf) for the kth order statistic in a 
 future sample.
 @name POSTKP}
\usage{
POSTKP(kord, nsamsz, xltime, ntvec, xmu, sigma, nsim, kdist, pdf, cdf)
}
\arguments{
\item{kord}{The order statistic}

\item{nsamsz}{The size of the future sample}

\item{xltime}{The time vector}

\item{ntvec}{The length of the tvector (at which evaluations will be done)}

\item{xmu}{The mu* values from the posterior sample}

\item{sigma}{The sigma* values from the posterior sample}

\item{nsim}{The value M*}
}
\description{
Compute empirical the predictive distributions 
 (pdf, cdf) for the kth order statistic in a 
 future sample.
 @name POSTKP
}
