% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{postpr}
\alias{postpr}
\alias{POSTPR}
\title{Compute empirical the predictive distributions 
(pdf, cdf) of a single future observation from 
distribution kdist}
\usage{
POSTPR(xltime, ntvec, xmu, sigma, nsim, kdist, pdf, cdf)
}
\arguments{
\item{xltime}{The time vector}

\item{ntvec}{The length of the tvector (at which evaluations will be done}

\item{xmu}{The mu* values from the posterior sample}

\item{sigma}{The sigma* values from the posterior sample}

\item{nsim}{The value M*}
}
\description{
Compute empirical the predictive distributions 
(pdf, cdf) of a single future observation from 
distribution kdist
}
