% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictTable.R
\name{PredictTable}
\alias{PredictTable}
\title{Title}
\usage{
PredictTable(
  mlest.out,
  FtimeStart,
  FtimeEnd,
  warranty.time = Inf,
  print.total = TRUE
)
}
\arguments{
\item{mlest.out}{}

\item{FtimeStart}{}

\item{FtimeEnd}{}

\item{warranty.time}{}

\item{print.total}{}
}
\description{
Title
}
\examples{
\dontrun{

DeviceN.ld <- frame.to.ld(devicen,
                          response.column = "months", 
                          censor.column = "event",
                          case.weight.column = "counts")
                          
PredictTable(DeviceN.mlest.lnorm,
             FtimeStart = 0,
             FtimeEnd = 1)

PredictTable(DeviceN.mlest.lnorm,
             FtimeStart = 1,
             FtimeEnd = 2)

PredictTable(DeviceN.mlest.lnorm,
             FtimeStart = 0,
             FtimeEnd = 36)

PredictTable(DeviceN.mlest.lnorm,
             FtimeStart = 0,
             FtimeEnd = 36,
             warranty.time = 36)

DeviceN.mlest.weib <- mlest(DeviceN.ld,
                            dist = "Weibull")

PredictTable(DeviceN.mlest.weib,
             FtimeStart = 1,
             FtimeEnd = 2)

PredictTable(DeviceN.mlest.weib,
             FtimeStart = 0,
             FtimeEnd = 36)

}
}
