% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gmle.out.R
\name{print.gmle.out}
\alias{print.gmle.out}
\title{Title}
\usage{
\method{print}{gmle.out}(x,
  conf.level = GetSMRDDefault("SMRD.ConfLevel")/100,
  digits = GetSMRDDefault("SMRD.DigitsPrinted"),
  print.vcv = GetSMRDDefault("SMRD.LongPrint"), add.title = NULL,
  quote = T, prefix = "", ...)
}
\arguments{
\item{x}{}

\item{conf.level}{}

\item{digits}{}

\item{print.vcv}{}

\item{add.title}{}

\item{quote}{}

\item{prefix}{}

\item{...}{}
}
\value{
printed stuff
}
\description{
Title
}
