% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probpaper.R
\name{probpaper}
\alias{probpaper}
\title{Title}
\usage{
probpaper(distribution, xlab = "Time", xlim = c(1, 10),
  ylim = c(0.011, 0.99), original.par = F, my.title = NULL,
  cex = 1, cex.labs = 1.1, cex.tic.lab = 1, cex.title = cex.labs,
  sub.title = NULL, grids = FALSE, linear.axes = F, slope.axis = F,
  title.option = "blank", shape = NULL, draw.line = F,
  ylab = GetSMRDDefault("SMRD.LabelOnYaxis"))
}
\arguments{
\item{distribution}{}

\item{xlab}{}

\item{xlim}{}

\item{ylim}{}

\item{original.par}{}

\item{my.title}{}

\item{cex}{}

\item{cex.labs}{}

\item{cex.tic.lab}{}

\item{cex.title}{}

\item{sub.title}{}

\item{grids}{}

\item{linear.axes}{}

\item{slope.axis}{}

\item{title.option}{}

\item{shape}{}

\item{draw.line}{}

\item{ylab}{}
}
\description{
Title
}
\examples{
\dontrun{

probpaper("Weibull",
          xlim = c(1, 10), 
          grid = TRUE, 
          ylim = c(0.011,0.981))

probpaper("Weibull",
          xlim = c(1, 100),
          grid = TRUE,
          ylim = c(0.011,0.981))

probpaper("Weibull",
          xlim = c(1, 1000),
          grid = TRUE,
          ylim = c(0.011,0.981))

probpaper("Weibull",
          xlim = c(1, 1000),
          grid = TRUE, 
          ylim = c(0.0011,0.9981))

probpaper("Lognormal",
          xlim = c(1, 10),
          grid = TRUE,
          ylim = c(0.011,0.981))

probpaper("Lognormal",
          xlim = c(1, 100),
          grid = TRUE,
          ylim = c(0.011,0.981))

probpaper("Lognormal",
          xlim = c(1, 1000),
          grid = TRUE,
          ylim = c(0.011,0.981))
}
}
