% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.contour.R
\name{profile.contour}
\alias{profile.contour}
\title{Title}
\usage{
\method{profile}{contour}(
  fitted,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  profile.title = paste(fitted$subtitle, "\\n", lplot.type, "Likelihood for",
    variable.namex, "and", variable.namey, "\\n", model.dist.str),
  variable.namex = fitted$xlab,
  variable.namey = fitted$ylab,
  transformationx = "linear",
  transformationy = "linear",
  original.par = F,
  levels = c(0.001, 0.01, 0.2, 0.4, 0.6, 0.8, 0.9),
  pretty.x = NULL,
  pretty.y = NULL,
  add = F,
  lty = 1,
  col = 1,
  lwd = 1,
  static = static,
  ...
)
}
\arguments{
\item{fitted}{}

\item{xlim}{}

\item{ylim}{}

\item{profile.title}{}

\item{variable.namex}{}

\item{variable.namey}{}

\item{transformationx}{}

\item{transformationy}{}

\item{original.par}{}

\item{levels}{}

\item{pretty.x}{}

\item{pretty.y}{}

\item{add}{}

\item{lty}{}

\item{col}{}

\item{lwd}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

R4490.rdu <- frame.to.rdu(r4490,
                          ID.column = "vin",
                          time.column = "days" , 
                          cost.count.column = "costcount" ,
                          event.column = "code")

attr(R4490.rdu,"WindowInfo")

event.plot(R4490.rdu)
R4490.mcf <- mcf(R4490.rdu)
plot(R4490.mcf)

R4490.nhpp.out <- PlotMCFandNHPP(R4490.rdu, form = "power rule")
one.dim.profile(R4490.nhpp.out,size = 5,save.s = T)
two.dim.profile(R4490.nhpp.out, 
                profile.on.list = NULL,
                which = c(1,2), 
                size = c(5,5))

profile.contour(R4490.nhpp.outstruct1x2,
                transformationy = "log",
                variable.namey = "sigma", 
                variable.namex = "mu", 
                v = c(0.001, 0.01, .1,0.2, 0.4, 0.7, 0.9) )

}
}
