% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r4490}
\alias{r4490}
\title{Car door lock transmitter replacement data}
\format{
A \code{data.frame} with 47 rows and 4 variables:
\tabular{rlll}{
  [, 1] \tab vin \tab Vehicle identifcation number \tab \bold{Categoric}\cr
  [, 2] \tab days \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 3] \tab event \tab Event observed at \code{days} (R4490/End/MEnd) \tab \bold{Categoric}\cr
  [, 4] \tab costcount \tab What is cost count? \tab \bold{Numeric}
   }
}
\source{
####
}
\description{
R4490 is the code used by General Motors denoting Remote Control Door Lock Transmitter Replacement
}
\seealso{
Other data-almost: 
\code{\link{repairtimes}},
\code{\link{rocketmotor}},
\code{\link{shockabsorber}},
\code{\link{snubber}},
\code{\link{tree25years}},
\code{\link{v805tube}}
}
\concept{data-almost}
