% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renewal.plots.R
\name{renewal.plots}
\alias{renewal.plots}
\title{Title}
\usage{
renewal.plots(data.rdu, my.title = NULL)
}
\arguments{
\item{data.rdu}{}

\item{my.title}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")

renewal.plots(halfbeak.rdu)

}
}
