% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.tsplot.R
\name{repair.tsplot}
\alias{repair.tsplot}
\title{Title}
\usage{
repair.tsplot(data.rdu, my.title = NULL, xlab = "Failure Number",
  ylab = "Time Between Failure", ...)
}
\arguments{
\item{data.rdu}{}

\item{my.title}{}

\item{xlab}{}

\item{ylab}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")

repair.tsplot(halfbeak.rdu)

}
}
