% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.vs.explan.R
\name{resid.vs.explan}
\alias{resid.vs.explan}
\title{Title}
\usage{
\method{resid}{vs.explan}(groupm.list, original.par = T, cex = 1,
  my.title = NULL, grids = F, xlim = c(NA, NA), ylim = c(NA, NA),
  x.axis = "linear", x.to.plot = NULL, cex.points = 1, ...)
}
\arguments{
\item{groupm.list}{}

\item{original.par}{}

\item{cex}{}

\item{my.title}{}

\item{grids}{}

\item{xlim}{}

\item{ylim}{}

\item{x.axis}{}

\item{x.to.plot}{}

\item{cex.points}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

ZelenCap.ld <- frame.to.ld(zelencap,
                           response.column = 1,
                           censor.column = 2,
                           case.weight.column = 3,
                           x.columns = c(4, 5),
                           time.units = "Hours")
ZelenCap.groupm.out1 <-
  groupm.mleprobplot(ZelenCap.ld,
                     distribution = "Weibull")

resid.vs.order(ZelenCap.groupm.out1)

resid.vs.fit(ZelenCap.groupm.out1)

resid.vs.explan(ZelenCap.groupm.out1,
                x.to.plot = 1)

resid.vs.explan(ZelenCap.groupm.out1,
                x.to.plot = 2)

}
}
