% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.plots.R
\name{residual.plots}
\alias{residual.plots}
\title{Title}
\usage{
residual.plots(
  results.out,
  which.plot = "All",
  x.to.plot = NULL,
  my.title = NULL,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  fvquant = 0.5
)
}
\arguments{
\item{results.out}{}

\item{which.plot}{}

\item{x.to.plot}{}

\item{my.title}{}

\item{xlim}{}

\item{ylim}{}

\item{fvquant}{}
}
\description{
Title
}
\examples{
\dontrun{

ZelenCap.ld <- frame.to.ld(zelencap,
                           response.column = 1,
                           censor.column = 2,
                           case.weight.column = 3,
                           x.columns = c(4, 5),
                           time.units = "Hours")

ZelenCap.groupm.out3 <- 
  groupm.mleprobplot(ZelenCap.ld, 
                     distribution = "Lognormal", 
                     relationship = c("linear", "linear"), 
                     formula = Location ~ + g(volts) + g(celsius) + g(volts):g(celsius))

residual.plots(ZelenCap.groupm.out3)

}
}
