% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resistor}
\alias{resistor}
\title{Carbon-Film Resistor Accelerated Degradation Test}
\format{A \code{data.frame} with 116 rows and 4 variables:
\tabular{rlll}{
  [, 1] \tab percent \tab Observed increase in resistance \tab \bold{Numeric}\cr
  [, 2] \tab resistor \tab Resistor type \tab \bold{Categoric}\cr
  [, 3] \tab celsius \tab Temperature applied to \code{resistor} \tab \bold{Numeric}\cr
  [, 4] \tab hours \tab Accumulated hours (in thousands) when \code{percent} was measured\tab \bold{Numeric}
   }}
\source{
Shiomi, H., and Yanagisawa, T. (1979), 
        On distribution parameter during accelerated life test for a carbon film resistor, 
        Bulletin of the Electrotechnical Laboratory, \bold{43}, 330-345 (in Japanese).

Suzuki, K., Maki, K., and Yokogawa, S. (1993), 
        An analysis of degradation data of a carbon film and properties of the estimators, 
        in Statistical Sciences and Data Analysis, 501-511. K. Matusita, M. Puri, and T. Hayakawa, Editors.
        Utrecht, Netherlands: VSP.
}
\description{
Samples of carbon-film resistors were tested at each of three levels of temperature. At the standard 
             operating temperature of 50^{o} C, carbon-film resistors will slowly degrade. Changes in resistance 
             can cause reduced product performance or even cause system failures.  The test was run at high levels 
             of temperature to accelerate the chemical degradation process and obtain degradation data more quickly.
             
             The \code{resistor} dataset presents the percent change in resistance measured throughout the test, 
             while the \code{resistor2} dataset presents the absolute value of resistance measure during the test.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grampus}},
  \code{\link{grids1}}, \code{\link{halfbeak}},
  \code{\link{heatexchanger}}, \code{\link{lfp1370}},
  \code{\link{lfptrun100}}, \code{\link{lzbearing}},
  \code{\link{metalwear}}, \code{\link{mylarsub}},
  \code{\link{nicdbattery}}, \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{superalloy}},
  \code{\link{tantalum}}, \code{\link{turbine}},
  \code{\link{v7tube}}, \code{\link{valveseat}},
  \code{\link{zelencap}}
}
\concept{data-done}
