% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflm.mle.R
\name{rflm.mle}
\alias{rflm.mle}
\title{Title}
\usage{
rflm.mle(data.ld, cond.dist, fl.dist, theta.start = NULL,
  tran.theta.start = T, debug1 = F, ...)
}
\arguments{
\item{data.ld}{}

\item{cond.dist}{}

\item{fl.dist}{}

\item{theta.start}{}

\item{tran.theta.start}{}

\item{debug1}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

LaminatePanel.ld <- frame.to.ld(laminatepanel,
                                response.column = "kilocycles",
                                time.units = "kilocycles",
                                data.title = "Laminate Panel Fatigue Data",
                                censor.column = "event", 
                                x.column = "mpa")

censored.data.plot(LaminatePanel.ld, 
                   x.axis = "log", 
                   y.axis = "log",
                   response.on.yaxis = F)

censored.data.plot(LaminatePanel.ld)

LaminatePanel.fit11.out <- rflm.mle(LaminatePanel.ld,
                                    cond.dist = "sev", 
                                    fl.dist = "sev")

plot(LaminatePanel.fit11.out,response.on.yaxis = F)

LaminatePanel.fit11f.out <- rflm.mle(LaminatePanel.ld,
                                     cond.dist = "sev", 
                                     fl.dist = "sev",
                                     debug = 1,
                                     fixed.param.list =   
                                     list(fixed.parameters = 3,
                                          fixed.parameter.values = 0.20))

plot(LaminatePanel.fit11f.out,
     response.on.yaxis = F)


}
}
