% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RISKSET}
\alias{RISKSET}
\title{Compute the risk set for a recurrence process within 
observation windows.}
\usage{
RISKSET(muniqrecurr, tuniq, nwindows, twindowsl, twindowsu, wcounts, iordl,
  iordu, delta, kdebug, iscrat)
}
\arguments{
\item{muniqrecurr}{number of unique recurrence times}

\item{tuniq}{length(muniqrecurr) unique recurrence 
times (in increasing order);}

\item{nwindows}{length of twindowsl and twindowsu;}

\item{twindowsl}{vector giving observation window 
start points}

\item{twindowsu}{window end points corresponding 
to twindowsl}

\item{wcounts}{counts for the windows;}

\item{iordl}{scratch space ordering vector for twindowsl;}

\item{iordu}{scratch ordering vector for twindowsu;}
}
\value{
delta length(muniqrecurr) size of the risk set 
        just before time tuniq(j)
}
\description{
We want the number of units at risk 
             just before each event time it is 
             also possible to send down a longer 
             list of times for purposes of making 
             a riskset plot.
}
\details{
Wrapper for wqm_riskset
}
