% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskSet.R
\name{RiskSet}
\alias{RiskSet}
\title{Title}
\usage{
RiskSet(data.rdu, kdebug1 = F, JustEvent = T)
}
\arguments{
\item{data.rdu}{}

\item{kdebug1}{}

\item{JustEvent}{}
}
\description{
Title
}
\examples{
\dontrun{

halfbeak.rdu <- frame.to.rdu(halfbeak,
                             ID.column = "unit", 
                             time.column = "hours" ,
                             event.column = "event", 
                             data.title = "Halfbeak Data", 
                             time.units = "Thousands of Hours of Operation")

RiskSet(halfbeak.rdu)

}
}
