% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.to.ld.R
\name{rmd.to.ld}
\alias{rmd.to.ld}
\title{Title}
\usage{
rmd.to.ld(
  data.rmd,
  fail.level,
  x.axis = "Linear",
  y.axis = "Linear",
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  title.option = GetSMRDDefault("SMRD.TitleOption"),
  censor.time = NULL,
  xlab = get.time.units(data.rmd),
  time.units = xlab,
  ylab = get.response.units(data.rmd),
  subset = T,
  big.one = 1e+36,
  my.title = NULL,
  group.var = 1:length(get.x.columns(data.rmd)),
  stresses = get.x.markers(data.rmd, group.var = group.var),
  extrapolation.control = "infer",
  print.estimates = T
)
}
\arguments{
\item{data.rmd}{}

\item{fail.level}{}

\item{x.axis}{}

\item{y.axis}{}

\item{xlim}{}

\item{ylim}{}

\item{title.option}{}

\item{censor.time}{}

\item{xlab}{}

\item{time.units}{}

\item{ylab}{}

\item{subset}{}

\item{big.one}{}

\item{my.title}{}

\item{group.var}{}

\item{stresses}{}

\item{extrapolation.control}{}

\item{print.estimates}{}
}
\description{
Title
}
\examples{
\dontrun{

GaAsLaser.rmd <- 
  frame.to.rmd(gaaslaser, 
               response.column = 1, 
               unit.column = 2, 
               time.column = 3, 
               response.units = "Increase in Operating Current (\%)")

# Convert to a \code{life.data} object

GaAsLaser.ld <- rmd.to.ld(GaAsLaser.rmd, 
                          fail.level = 10, 
                          x.axis = "sqrt")
summary(GaAsLaser.ld)
plot(GaAsLaser.ld)

}
}
