% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SBVN}
\alias{SBVN}
\title{Compute the probability that a bivariate normal with mean vectorc}
\usage{
SBVN(ah, ak, xmu1, xmu2, v1, v2, c12, prob, n, kprint)
}
\description{
Compute the probability that a bivariate normal with mean vectorc
}
