% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.effect.R
\name{series.dep.effect}
\alias{series.dep.effect}
\title{Title}
\usage{
series.dep.effect(xrange = c(0.8, 0.999), yrange = c(0.8, 0.999),
  mu = logb(10), sd = 1, rho = c(0, 0.4, 0.7, 0.9, 1))
}
\arguments{
\item{xrange}{}

\item{yrange}{}

\item{mu}{}

\item{sd}{}

\item{rho}{}
}
\description{
Title
}
\examples{
\dontrun{
component.effect(parallel = TRUE)

parallel.effect()

parallel.dep.effect()

series.dep.effect()

}
}
