% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.effect.R
\name{series.effect}
\alias{series.effect}
\title{Title}
\usage{
series.effect(
  xrange = c(0.98, 1),
  yrange = c(0, 1),
  ncomp = c(1, 10, 25, 50, 100, 200, 400)
)
}
\arguments{
\item{xrange}{}

\item{yrange}{}

\item{ncomp}{}
}
\description{
Title
}
\examples{
\dontrun{
component.effect(parallel = TRUE)

parallel.effect()

parallel.dep.effect()

series.dep.effect()

}
}
