% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sfteval}
\alias{sfteval}
\alias{SFTEVAL}
\title{Does good stuff}
\usage{
SFTEVAL(
  kdmod,
  xmu2,
  sig2,
  xmu3,
  sig3,
  rho,
  df,
  d0,
  sfact,
  tf,
  number,
  answer,
  ier,
  kprint
)
}
\description{
Does good stuff
}
