% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shockabsorber}
\alias{shockabsorber}
\title{Vehicle shock absorber failure data}
\format{
A \code{data.frame} with 38 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab kilometers \tab Accumulated distance at \code{event} \tab \bold{Numeric}\cr
  [, 2] \tab mode       \tab Failure mode observed at \code{miles} \tab \bold{Categoric}\cr
  [, 3] \tab event      \tab Event observed at \code{miles} (failure/right-censored) \tab \bold{Categoric}
   }
}
\source{
O'Connor, P. D. T. (1985), 
        Practical Reliability Engineering (Second Edition), New York, NY; John Wiley & Sons.
}
\description{
O'Connor gives the failure times (in number of kilometers of use) of vehicle shock absorbers. 
             The data shows two different failure modes occurring, denoted by M1 and M2. Engineers responsible 
             for shock absorber manufacturing and reliability were interested in the distribution of kilometers 
             to failure for the individual failure modes. Engineers responsible for higher-level automobile system 
             reliability and choosing among alternative vendors were interested in the overall failure distribution 
             for the shock absorbers.
}
\seealso{
Other data-almost: 
\code{\link{r4490}},
\code{\link{repairtimes}},
\code{\link{rocketmotor}},
\code{\link{snubber}},
\code{\link{tree25years}},
\code{\link{v805tube}}
}
\concept{data-almost}
