% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.contour.R
\name{simple.contour}
\alias{simple.contour}
\title{Create a contour plots of a likelihood function}
\usage{
simple.contour(
  data.ld,
  distribution,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  size = 100,
  static = F,
  quantile = NA,
  show.confidence = TRUE,
  contour.indicators = NULL,
  zoom.level = 3.5,
  show.mle = T,
  my.title = NULL,
  threeD = F,
  direction = NULL,
  profile = NULL,
  lwd = 1,
  elevation = NULL,
  distance = NULL,
  original.par = T,
  log.quantile = F,
  log.axis = T,
  add = F,
  ...
)
}
\arguments{
\item{data.ld}{}

\item{distribution}{}

\item{xlim}{}

\item{ylim}{}

\item{size}{}

\item{quantile}{}

\item{show.confidence}{}

\item{contour.indicators}{}

\item{zoom.level}{}

\item{show.mle}{}

\item{my.title}{}

\item{threeD}{}

\item{direction}{}

\item{profile}{}

\item{lwd}{}

\item{elevation}{}

\item{distance}{}

\item{original.par}{}

\item{log.quantile}{}

\item{log.axis}{}

\item{add}{}

\item{...}{}
}
\description{
Create a contour plots of a likelihood function
}
\examples{
\dontrun{

ShockAbsorber.ld <- frame.to.ld(shockabsorber,
                                response.column = 1, 
                                censor.column = 3,
                                time.units = "Kilometers")

simple.contour(ShockAbsorber.ld,
               distribution = "lognormal",
               show.confidence = T)


simple.contour(ShockAbsorber.ld,
               distribution = "lognormal",
               show.confidence = T,
               threeD = T)
               
bulb.ld <- frame.to.ld(bulb,
                       response.column = 1,
                       data.title = "Bulb Data",
                       time.units = "Hours")

mlest(bulb.ld,"lognormal")

simple.contour(bulb.ld,"lognormal", quantile = .5)

# Bearing Cage Data Set

BearingCage.ld <- frame.to.ld(bearingcage,
                              response.column = 1, 
                              censor.column = 2, 
                              case.weight.column = 3,
                              time.units = "Hours")

simple.contour(BearingCage.ld,
               distribution = "Weibull",
               zoom.level = 4,
               quantile = .1,
               threeD = TRUE)

simple.contour(BearingCage.ld,
               distribution = "Weibull",
               zoom.level = 4,
               profile = "x")

simple.contour(BearingCage.ld,
               distribution = "Weibull",
               zoom.level = 4,
               profile = "y")


simple.contour(BearingCage.ld,
               distribution = "Weibull",
               threeD = T,
               zoom.level = 3,
               size = 75)
}
}
