% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleDistSim.R
\name{SingleDistSim}
\alias{SingleDistSim}
\title{Title}
\usage{
SingleDistSim(number.sim, distribution, sample.size,
  censor.type = "None", fail.number = NULL, fail.fraction = NULL,
  theta = c(0, 1), debug1 = F, kprint = 0, randomize = T,
  remove.bad = F, var = T)
}
\arguments{
\item{number.sim}{}

\item{distribution}{}

\item{sample.size}{}

\item{censor.type}{}

\item{fail.number}{}

\item{fail.fraction}{}

\item{theta}{}

\item{debug1}{}

\item{kprint}{}

\item{randomize}{}

\item{remove.bad}{}

\item{var}{}
}
\description{
Title
}
\examples{
\dontrun{

SingleDistSim(number.sim = 10,
              distribution = "Weibull",
              theta = c(mu = 0.0, sigma = 1.0),
              sample.size = 10,
              censor.type = "None")

SingleDistSim(number.sim = 10,
              distribution = "Weibull",
              theta = c(mu = 0.0, sigma = 1.0),
              sample.size = 10,
              censor.type = "Type 1", 
              fail.fraction = 0.5)

SingleDistSim(number.sim = 10,
              distribution = "Weibull",
              theta = c(mu = 0.0, sigma = 1.0),
              sample.size = 10,
              censor.type = "Type 2",
              fail.number = 5)

}
}
