% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sloglikenhpp}
\alias{sloglikenhpp}
\alias{SLOGLIKENHPP}
\title{Compute a vector of NHPP log likelihood values 
corresponding to a thetav matrix}
\usage{
SLOGLIKENHPP(
  time,
  ntimes,
  recurrcosts,
  timel,
  timeu,
  kwcount,
  nwindows,
  kform,
  thetav,
  nparm,
  ntheta,
  answer
)
}
\description{
Compute a vector of NHPP log likelihood values 
corresponding to a thetav matrix
}
