% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small.interval.R
\name{small.interval}
\alias{small.interval}
\title{Title}
\usage{
small.interval(data.ld, delta = 1e-05, tlog = T)
}
\arguments{
\item{data.ld}{}

\item{delta}{}

\item{tlog}{}
}
\description{
Title
}
\examples{
\dontrun{

Fan.ld <- frame.to.ld(fan,
                      response.column = 1, 
                      censor.column = 2, 
                      case.weight.column = 3,
                      time.units = "Hours")
                    
gets.mle(small.interval(Fan.ld, delta = 0.01),
         distribution = "sev")

gets.mle(small.interval(Fan.ld, delta = 0.01),
         distribution = "normal")

}
}
