% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-sev.R
\name{dsev}
\alias{dsev}
\alias{psev}
\alias{qsev}
\alias{rsev}
\title{The Smallest Extreme Value Distribution}
\usage{
dsev(x, location = 0, scale = 1)

psev(q, location = 0, scale = 1)

qsev(p, location = 0, scale = 1)

rsev(n, location = 0, scale = 1)
}
\description{
The Smallest Extreme Value Distribution
}
