% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRD.R
\docType{package}
\name{SMRD-package}
\alias{SMRD}
\alias{SMRD-package}
\title{SMRD: Statistical Methods For Reliability Data}
\description{
Implements the reliability estimation functions and 
    data sets for use with the text "Statistical Methods for 
    Reliability Data" Copyright 1998 by Meeker & Escobar.
}
\author{
\strong{Maintainer}: Jason Freels \email{auburngrads@live.com}

Authors:
\itemize{
  \item William Meeker \email{wqmeeker@iastate.edu}
  \item Luis Escobar \email{luis@lsu.edu}
}

}
\keyword{internal}
