% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SSFT2GR1}
\alias{SSFT2GR1}
\title{Vector version of urlike for R calling}
\usage{
SSFT2GR1(t, nt, r1log, mut2, sigmat2, mur1, sigmar1, mur2, sigmar2, rho,
  answer, kprint, ier)
}
\description{
Vector version of urlike for R calling
}
