% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.posterior.or.prior.R
\name{summarize.posterior.or.prior}
\alias{summarize.posterior.or.prior}
\title{Title}
\usage{
summarize.posterior.or.prior(
  posterior.object,
  task,
  marginal.on.pos,
  marginal.on.sigma,
  type.position,
  newdata,
  include.likelihood,
  post.or.prior,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = parameter.contour.axes.labels.out$xlab,
  ylab = parameter.contour.axes.labels.out$ylab,
  axes.range.default.post = T,
  factor = 4.5,
  size = GetSMRDDefault("SMRD.DefaultGridPoints"),
  number.points.plot = 500
)
}
\arguments{
\item{posterior.object}{}

\item{task}{}

\item{marginal.on.pos}{}

\item{marginal.on.sigma}{}

\item{type.position}{}

\item{newdata}{}

\item{include.likelihood}{}

\item{post.or.prior}{}

\item{xlim}{}

\item{ylim}{}

\item{xlab}{}

\item{ylab}{}

\item{axes.range.default.post}{}

\item{factor}{}

\item{size}{}

\item{number.points.plot}{}
}
\description{
Title
}
\examples{
\dontrun{
# Specify prior distribution characteristics for the 
# \code{bearingcage} data using a non-informative quantile 
# and a noninformative sigma

prior.spec1 <-
  specify.simple.prior(p = .01,
                       qdist = "loguniform",
                       qlower = 100,
                       qupper = 5000,
                       sigma.dist =  "lognormal",
                       sigma.lower = 0.2,
                       sigma.upper =  0.5,
                       distribution = "Weibull")
                       
# Specify prior distribution characteristics for the 
# \code{bearingcage} data using a noninformative quantile 
# and an informative sigma

prior.spec2 <-
  specify.simple.prior(p = .01,
                       qdist = "loguniform",
                       qlower = 1000,
                       qupper = 1400,
                       sigma.dist = "lognormal",
                       sigma.lower = 1.5,
                       sigma.upper = 2.5, 
                       distribution  = "Weibull")
                       
# Specify prior distribution characteristics for the 
# \code{bearingcage} data using an informative quantile 
# and an informative sigma

prior.spec3 <-
  specify.simple.prior(p = .01,
                       qdist = "lognormal",
                       qlower = 1000,
                       qupper = 1400,
                       sigma.dist = "lognormal",
                       sigma.lower = 1.5,
                       sigma.upper = 2.5,
                       distribution  = "Weibull")
 
# Create the prior distributions                      
prior3.bcage <- 
  make.prior(spec = prior.spec3, 
             number.in.prior = 3000)


prior.and.post3.bcage <-
  get.big.posterior(prior.spec3,
                    BearingCage.ld)

prior.and.post3.bcage$post[1:10,] 

prior.and.post3.bcage <- 
  make.small.posterior.object(prior.and.post3.bcage)
  
summarize.posterior.or.prior(prior.and.post3.bcage,
                             post.or.prior = "post",
                             task = "Marginals only",
                             marginal.on.sigma = T,
                             marginal.on.pos = F,
                             type.position = "Parameter",
                             newdata = "mu",
                             include.likelihood = T)
}
}
