% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.plan.values.R
\name{summary.plan.values}
\alias{summary.plan.values}
\title{Return a summmary of reliability test plan values}
\usage{
\method{summary}{plan.values}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\value{
A summary
}
\description{
Return a summmary of reliability test plan values
}
