% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SURLIKE}
\alias{SURLIKE}
\title{Vector version of urlike for Splus calling}
\usage{
SURLIKE(
  t,
  nt,
  mut1,
  sigmat1,
  mut2,
  sigmat2,
  mur1,
  sigmar1,
  mur2,
  sigmar2,
  rho,
  answer,
  kprint
)
}
\description{
Vector version of urlike for Splus calling
}
