% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.lines.R
\name{table.lines}
\alias{table.lines}
\title{Title}
\usage{
table.lines(z, distribution)
}
\arguments{
\item{z}{}

\item{distribution}{}
}
\value{
a list of values
}
\description{
Title
}
\examples{
\dontrun{

table.lines(seq(-1,1,by=.1),"sev")

table.lines(seq(-1,1,by=.1),"normal")

}
}
\seealso{
\code{\link{lsinf}} \code{\link{variance.factor}}
}
