% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanium2}
\alias{titanium2}
\title{Titanium fatigue crack growth data}
\format{
A \code{data.frame} with 10 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab kilocycles \tab Accumulated cycles at \code{event} \tab \bold{Numeric}\cr
  [, 2] \tab event      \tab Event observed at \code{cycles} (failure/right-censored) \tab \bold{Categoric}\cr 
  [, 3] \tab count      \tab Number of events observed at \code{cycles} \tab \bold{Numeric}
  }
}
\source{
Meeker, W. Q. and Escobar, L. A. (1998) Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons.

Hudak, S.J., Saxena, A., Bucci, R. J., and Malcom, R .C. (1978), 
        Development of standard methods of testing and analyzing fatigure crack growth rate data, 
        Technical Report AFML-TR-78-40, Westinghouse R & D Center, Westinghouse Electric Corporation, Pittsburgh, PA.
}
\description{
A sample of 100 specimens of a titanium alloy were subjected to a fatigue test to determine time 
             to crack initiation. The test was run up to a limit of 100,000 cycles.  The observed number of cycles to crack 
             initiation (in units of 1,000 of cycles) were: 18, 32, 39, 53, 59, 68, 77, 78, 93. No crack had initiated in 
             any of the other 91 other specimens prior to reaching 100,000 cycles.
}
\seealso{
Other data-notdone: 
\code{\link{adhesivebondc}},
\code{\link{adhesivestrength}},
\code{\link{amsaaexactfail}},
\code{\link{appliancea}},
\code{\link{applianceb}},
\code{\link{aprel72}},
\code{\link{atrazinejune}},
\code{\link{bearinga}},
\code{\link{censortruncationtest}},
\code{\link{chemicalprocess}},
\code{\link{cirpack4}},
\code{\link{cirpack5}},
\code{\link{componentd}},
\code{\link{concrete}},
\code{\link{connectionstrength}},
\code{\link{customerlife}},
\code{\link{devicedconnected}},
\code{\link{deviceh}},
\code{\link{devicen}},
\code{\link{diskber}},
\code{\link{doatrun}},
\code{\link{electromech}},
\code{\link{engineemissions}},
\code{\link{epoxyweathering}},
\code{\link{filament}},
\code{\link{hcfdata}},
\code{\link{heater}},
\code{\link{hpcrepairs}},
\code{\link{icdevice1}},
\code{\link{inconel}},
\code{\link{insulation}},
\code{\link{laminatepanel}},
\code{\link{largeball}},
\code{\link{locomotivecontrol}},
\code{\link{luminosity}},
\code{\link{machineh}},
\code{\link{mechanicalswitch}},
\code{\link{multiplefail}},
\code{\link{newspring}},
\code{\link{parta}},
\code{\link{piccioto}},
\code{\link{pipelinethickness}},
\code{\link{pmax}()},
\code{\link{prob3_5}},
\code{\link{riverchem}},
\code{\link{shelflifea}},
\code{\link{testdadtplan}},
\code{\link{titanium}},
\code{\link{tractorbreaks}},
\code{\link{turbinedevice}},
\code{\link{vehiclemotor}},
\code{\link{voltageendurance}},
\code{\link{workstation}}
}
\concept{data-notdone}
