% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turbine}
\alias{turbine}
\title{Turbine wheel crack initiation data}
\format{A \code{data.frame} with 21 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab hours \tab Accumulated hours (in hundreds) at \code{event} \tab \bold{Numeric}\cr
  [, 2] \tab event \tab Event observed at \code{hours} (right-censored/left-censored) \tab \bold{Categoric}\cr
  [, 3] \tab count \tab Number of events observed at \code{hours} \tab \bold{Numeric}
   }}
\source{
Nelson, W.	(1982) Applied Life Data Analysis, New York, NY: John Wiley \& Sons.
}
\description{
Nelson describes a study to estimate the 
             distribution of time to crack initiation for turbine 
             wheels.  Each of 432 wheels were inspected once to 
             determine if it had started to crack or not.  At the 
             time of the inspections, the wheels had different 
             amounts of service time (age). A unit found to be 
             cracked at its inspection was labelled as left-censored 
             at its age (because the crack had initiated at some 
             unknown point before its inspection age).  A unit found 
             to be uncracked at its inspection was labelled as 
             right-censored at its age (because a crack would be 
             initiated at some unknown point after that age). 
             The data show the number of cracked and uncracked 
             wheels in different age categories, showing the midpoint
             of the time interval given by Nelson.  The data were 
             put into intervals to facilitate simpler analyses.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grampus}},
  \code{\link{grids1}}, \code{\link{halfbeak}},
  \code{\link{heatexchanger}}, \code{\link{lfp1370}},
  \code{\link{lfptrun100}}, \code{\link{lzbearing}},
  \code{\link{metalwear}}, \code{\link{mylarsub}},
  \code{\link{nicdbattery}}, \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{v7tube}}, \code{\link{valveseat}},
  \code{\link{zelencap}}
}
\concept{data-done}
