% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.factor.R
\name{variance.factor}
\alias{variance.factor}
\title{Title}
\usage{
variance.factor(distribution = NULL, type = "quantile",
  quantile.of.interest = NULL, proportion.failing = NULL,
  number.points = 200, ...)
}
\arguments{
\item{distribution}{}

\item{type}{}

\item{quantile.of.interest}{}

\item{proportion.failing}{}

\item{number.points}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{

# For the normal distribution
variance.factor(distribution = 'normal', 
                type = 'quantile', 
                quantile.of.interest = 0.02,
                proportion.failing = 0.2)


# For the smallest extreme value distribution
variance.factor(distribution = 'sev',
                type = 'quantile', 
                quantile.of.interest = 0.02,
                proportion.failing = 0.2)
}
}
