% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varone.grid.R
\name{varone.grid}
\alias{varone.grid}
\title{Title}
\usage{
varone.grid(pi.vec = seq(0.001, 0.999, length = 100),
  xi.vec = seq(xiL.lower, 0.999, length = 100),
  distribution = "weibull", a = NULL, b1 = NULL, b2 = 0,
  pd = NULL, ph = NULL, theta = 1, theta.known = T,
  quad.model = F, perc = 0.01, zd = 0, kprint = 0,
  levels = c(0.1, 0.2, 0.5, 1, 2), cex2 = 2, xlab = "X",
  ylab = "Y", scale.var = T, old.a = NULL, old.b1 = NULL,
  xiL.lower = 0.001, dump = F)
}
\arguments{
\item{pi.vec}{}

\item{xi.vec}{}

\item{distribution}{}

\item{a}{}

\item{b1}{}

\item{b2}{}

\item{pd}{}

\item{ph}{}

\item{theta}{}

\item{theta.known}{}

\item{quad.model}{}

\item{perc}{}

\item{zd}{}

\item{kprint}{}

\item{levels}{}

\item{cex2}{}

\item{xlab}{}

\item{ylab}{}

\item{scale.var}{}

\item{old.a}{}

\item{old.b1}{}

\item{xiL.lower}{}

\item{dump}{}
}
\description{
Title
}
\examples{
\dontrun{

varone.grid(a = -0.9, 
            b1 = -7.9, 
            perc = 0.01, 
            xlab = "xiL", 
            ylab = "piL",
            xiL.lower = -1, 
            levels = c(.001,.01,0.10000000000000001, 0.20000000000000001, 0.5, 1, 2),
            dump = T)

}
}
