% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VVAR1}
\alias{VVAR1}
\title{Vectorized version ov svar1
Calculate variance of percentile at some stress level
zivar is the variable defining that level
(usually set to zero for design stress)
this a wrap for a call from splus}
\usage{
VVAR1(param, z, pi, zivar, npar, nplan, nlev, perc, idist, knownt, fret,
  varret, iprint)
}
\description{
Vectorized version ov svar1
Calculate variance of percentile at some stress level
zivar is the variable defining that level
(usually set to zero for design stress)
this a wrap for a call from splus
}
