% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wqm_cdfest}
\alias{wqm_cdfest}
\alias{WQMCDFEST}
\title{Compute Nonparametric CDF Estimates}
\usage{
WQMCDFEST(
  y,
  ny,
  codes,
  weight,
  ty,
  nty,
  tcodes,
  n,
  nstart,
  dscrat,
  scrat,
  iscrat,
  kprint,
  maxit,
  tol,
  maxmsd,
  p,
  q,
  prob,
  sd,
  m,
  pchmax,
  lsd,
  ier
)
}
\arguments{
\item{y}{Numeric matrix \code{[n x ny]} containing the lower and upper limits for group
censored observations.}

\item{ny}{Number of columns in \code{y} (either 1 or 2)}

\item{codes}{Integer vector of censor codes (see Details)}

\item{weight}{Vector of observation weights or multiplicities}

\item{ty}{Numeric matrix containing lower and upper truncation limits}

\item{nty}{Number of columns in ty}

\item{tcodes}{Integer vector of truncation codes (see details)}

\item{n}{Number of rows in y}

\item{nstart}{If \code{nstart = 0} set automatic start values, otherwise send down \code{nstart}
values in \code{prob} for restart}

\item{dscrat}{Numeric scratch vector (length = 3 * n + 2)}

\item{scrat}{Numeric scratch array (length = max((7 * n), maxmsd * (maxmsd - 1) / 2))}

\item{iscrat}{Integer scratch array (length = 6 * n + 4)}

\item{maxit}{Maximum number of iterations for s-c algorithm}

\item{tol}{Desired estimation accuracy (0 < \code{tol} < 0.1)}

\item{maxmsd}{Maximum \code{m} for which the full information matrix estimates of the
standard errors can be computed}

\item{iprint}{Print level for debug dump}
}
\value{
A \code{list} of length 6
\itemize{
 \item{p}{Numeric vector of lower limits of intervals for the cdf estimate (length = n + 1)}
 \item{q}{Numeric vector of upper limits of intervals for the cdf estimate (length = n + 1)}
 \item{prob}{Numeric vector of cdf estimates corresponding to intervals \code{p} and \code{q}}
 \item{m}{Actual length of \code{p}, \code{q}, and \code{prob} (depends on data)}
 \item{pchmax}{Max change in last iteration (if greater than tol, 0.0 if Kaplan-Meier was used)}
 \item{ier}{Error code (see details)}
}
             0     no error
             1     n <= 0 on input
             2     ny not equal to 1 or 2
             3     nty not between 0 and 2
             4     tol outside range (0, 0.1)
             6     censor code out of range 0 to 4
             7     y(i,1) != y(i,2) in a type 1, 2, or 3 observation
             8     y(i,1) <  y(i,2) in type 4 obs
             9     ny=1 but type 4 obs found
            10     tcode outside range 1 to 4
            11     tcode = 1, 2, or 3 but tyl != tyu
            12     backwards truncation interval
            12     ty(i,1) < ty(i,2) in type 4 obs
            13     only 1 col of truncation values, code 4 found
            14     not enough data to estimate distribution
                      (e.g., all right censored observations)
            15     nstart>0 does not agree with computed m
            16     observation not within the truncation interval
           *21     product-limit estimate could not be computed
                   directly and maxmsd was too small to allow the
                   full information matrix to be computed
                   an approximation was computed under the assumption
                   that the individual hazard
                   estimates are uncorrelated
           *22     information matrix not positive definite
           *23     only one non-zero s probability

        * signifies warning message only
}
\description{
Computes nonparametric estimates of the cumulative distribution function
             using maximum likelihood. If possible, estimation this is done with a
             generalized version of the Kaplan-Meier estimate, and otherwise, by using
             Turnbull's e-m algorithm
}
\details{
\itemize{
 \item{0}{dummy observation}
 \item{1}{exact failure time}
 \item{2}{right censored observation}
 \item{3}{left censored observation}
 \item{4}{interval or group censored observation}
 \item{5}{exact failure time recoded as a small interval}
 }
\itemize{
 \item{1}{no truncation}
 \item{2}{right truncated observation}
 \item{3}{left truncated observation}
 \item{4}{interval truncated observation}
 }

 \code{iprint = 0} for no debug output, if > 0 dump setup and every iprint iteration,
 \code{iprint = 1} gives the maximum amount of output, \code{iprint = 10} is a good choice to debug
}
