% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarray.R
\name{xarray}
\alias{xarray}
\title{Convert a LaTex-style array for use in rmarkdown}
\usage{
xarray(table, digits = NULL, matrix = FALSE, ...)
}
\arguments{
\item{table}{A \code{data.frame} or \code{matrix} to be converted}

\item{digits}{Sets the number of significant digits for numeric values}

\item{matrix}{A \code{logical} indicating if \code{table} is a 
matrix-class object.}

\item{...}{Additional options used in \code{xtable:::print.xtable}}
}
\value{
An printed LaTeX array class object
}
\description{
This function utilizes \code{xtable::xtable} to 
             return an array class object. This is useful for 
             rendering matrices in rmarkdown documents and
             presentations
}
