% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmat.R
\name{xmat}
\alias{xmat}
\title{Title}
\usage{
xmat(data.ld, allow = T)
}
\arguments{
\item{data.ld}{}

\item{allow}{}
}
\description{
Title
}
\examples{
\dontrun{

ZelenCap.ld <- frame.to.ld(zelencap,
                           response.column = 1,
                           censor.column = 2,
                           case.weight.column = 3,
                           x.columns = c(4, 5),
                           time.units = "Hours")

xmat(ZelenCap.ld)

}
}
