% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XXMCF}
\alias{XXMCF}
\title{Computes the sample MCF and its robust sample 
variance. Created by Huaiqing Wu & William Q. 
Meeker, Iowa State University}
\usage{
XXMCF(numrecurr, timeofrecurr, krecurrid, dcost, muniqrecurr, tuniq,
  apoint, lnumrecurr, delta, nunitsgroups, wpoint, nwindows, twindowsl,
  twindowsu, wcounts, inwindowj, muhat, varhat, dbar, iordl, iordu, iorder,
  iscrat, kdebug)
}
\description{
Computes the sample MCF and its robust sample 
variance. Created by Huaiqing Wu & William Q. 
Meeker, Iowa State University
}
