% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zelencap}
\alias{zelencap}
\title{Glass capacitor accelerated life test data}
\format{A \code{data.frame} with 40 rows and 5 variables:
\tabular{rlll}{
  [, 1] \tab hours \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 2] \tab event \tab Event observed at \code{hours} (failure/right-censored) \tab \bold{Categoric}\cr
  [, 3] \tab count \tab Number of events observed at \code{hours} \tab \bold{Numeric}\cr
  [, 4] \tab celsius \tab Temperature applied  \tab \bold{Numeric}\cr
  [, 5] \tab volts \tab Electrical power applied \tab \bold{Numeric}
  }}
\source{
Zelen, M. (1959) Factorial experiments in life testing, Technometrics, \bold{1}3, 269-288
}
\description{
Zelen describes an accelerated life test performed on glass capacitors at higher than usual levels of 
             temperature and voltage. The data were obtained from a factorial experiment in which eight capacitors were tested 
             at each combination of temperature and voltage. For each combination, the test was terminated after the fourth failure
             was observed, yielding failure (Type II) censored data.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grampus}},
  \code{\link{grids1}}, \code{\link{halfbeak}},
  \code{\link{heatexchanger}}, \code{\link{lfp1370}},
  \code{\link{lfptrun100}}, \code{\link{lzbearing}},
  \code{\link{metalwear}}, \code{\link{mylarsub}},
  \code{\link{nicdbattery}}, \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}
}
\concept{data-done}
