% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.failure.R
\name{zero.failure.k}
\alias{zero.failure.k}
\title{Determine test length for a zero-failure test as a multiple of spec}
\usage{
zero.failure.k(beta = 2, quantile = 0.01, conlev = 0.9, n = 10,
  printem = T)
}
\arguments{
\item{beta}{}

\item{quantile}{}

\item{conlev}{}

\item{n}{}

\item{printem}{}
}
\description{
Determine test length for a zero-failure test as a multiple of spec
}
\examples{
\dontrun{

zero.failure.k(beta = 2, quantile = 0.1, conlev = 0.99,	n = 5)

zero.failure.k(beta = 1, quantile = 0.01, conlev = 0.95, n = 5)

zero.failure.k(beta = 2, quantile = 0.01, conlev = 0.95, n = 5)
}
}
