% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.failure.R
\name{zero.failure.n}
\alias{zero.failure.n}
\title{Determine sample size for a zero-failure test}
\usage{
zero.failure.n(conlev, quantile, k, beta)
}
\arguments{
\item{conlev}{}

\item{quantile}{}

\item{k}{}

\item{beta}{}
}
\description{
Determine sample size for a zero-failure test
}
\examples{
\dontrun{

zero.failure.n(conlev = 0.95, quantile = 0.01, k = 14, beta = 1)

zero.failure.n(conlev = 0.95, quantile = 0.01, k = 3.369, beta = 2)

}
}
