% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.failure.R
\name{zero.failure.plan}
\alias{zero.failure.plan}
\title{Number of test time units needed for minimum-sized demonstration}
\usage{
zero.failure.plan(
  betavec = c(0.8, 1, 1.5, 2, 3),
  quantile = 0.1,
  conlev = 0.99,
  krange = c(1.5, 4),
  xlim = NULL,
  ylim = NULL,
  grid = F,
  bw = FALSE,
  my.title = NULL
)
}
\arguments{
\item{betavec}{}

\item{quantile}{}

\item{conlev}{}

\item{krange}{}

\item{xlim}{}

\item{ylim}{}

\item{grid}{}

\item{bw}{}

\item{my.title}{}
}
\description{
Number of test time units needed for minimum-sized demonstration
}
\examples{
\dontrun{

zero.failure.plan(xlim = c(1.51,3.99), 
                  ylim = c(.1,29), 
                  krange = c(1.5,3.83))

zero.failure.plan(betavec = c( 1., 2.), 
                  quantile = 0.01, 
                  conlev = 0.95, 
                  xlim = c(1.51,10), 
                  ylim = c(.1,199), 
                  krange = c(1.5,10),
                  grid = T,
                  bw = FALSE)
}
}
