% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero.failure.R
\name{zero.failure.prsd}
\alias{zero.failure.prsd}
\title{Title}
\usage{
zero.failure.prsd(alpha.vec, quantile, pfactor = 3,
  number.points = 100, grids = TRUE, bw = FALSE, my.title = NULL)
}
\arguments{
\item{alpha.vec}{}

\item{quantile}{}

\item{pfactor}{}

\item{number.points}{}

\item{grids}{}

\item{bw}{}

\item{my.title}{}
}
\description{
Title
}
\examples{
\dontrun{

zero.failure.prsd(alpha.vec = c(0.05,0.1), quantile = 0.01, pfactor = 3)

}
}
