% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfest.R
\name{cdfest}
\alias{cdfest}
\title{Generate a CDF estimate}
\usage{
cdfest(data.ld, gamthr = 0, kprint = 0, maxit = 5e+05, tol = 0.001,
  maxmsd = 200, start.values = NULL, debug1 = F)
}
\arguments{
\item{data.ld}{}

\item{gamthr}{}

\item{kprint}{}

\item{maxit}{}

\item{tol}{}

\item{maxmsd}{}

\item{start.values}{}

\item{debug1}{}
}
\value{
A formal S4 object of class CDFEST with slot \code{list}
}
\description{
Generate a CDF estimate
}
\examples{
\dontrun{
lz.ld <- life_data(lzbearing, response.column = 1)
lz.cd <- cdfest(lz.ld, distribution = 'weibull')
lz.cd
}
}
