% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comptime}
\alias{comptime}
\title{Computer program execution time data}
\format{
A \code{data.frame} with 17 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab users \tab Number of users \tab \bold{Numeric}\cr
  [, 2] \tab load \tab System load \tab \bold{Numeric}\cr
  [, 3] \tab seconds \tab Execution time observed at \code{load} \tab \bold{Numeric}
   }
}
\source{
Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons.
}
\description{
Meeker and Escobar report the amount of time required for a Unix computer
             to execute a particular computer program on a multiuser computer system.
             The response (execution) times are a function of the total
             system load, which was obtained using the Unix \code{uptime}
             command.
}
