% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cylinder}
\alias{cylinder}
\title{Diesel engine cylinder replacements}
\format{A \code{data.frame} with 276 rows and 4 variables:
\tabular{rlll}{
  [, 1] \tab engine \tab Type of engine used \tab \bold{Categoric}\cr
  [, 2] \tab days   \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 3] \tab count  \tab Number of events observed at \code{days} \tab \bold{Numeric}\cr
  [, 4] \tab event  \tab Event observed at \code{days} (Replacement/End) \tab \bold{Categoric}
  }}
\source{
Nelson and Doganaksoy (1989),
        A computer program for an estimate and confidence limits for the mean cumulative function for cost or number of repairs of repairable products,
        TIS report 89CRD239, General Electric Company Research and Development, Schenectady, NY
}
\description{
Nelson and Doganaksoy present data on cylinder replacement times for 120 diesel engines.  
             We take these engines to be a sample from a larger population of engines. Each engine has 16 cylinders.
}
\details{
Cylinders in a type of diesel engine can develop leaks or have low compression. 
         Cylinders are inspected at times of convenience, along with other usual engine maintenance operations. 
         Faulty cylinders are replaced by a rebuilt cylinder.  More than one cylinder could be replaced at an inspection.  
         Management needed to know if the company should perform preventive replacement of cylinders before they develop 
         low compression.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{devicea}}, \code{\link{deviceb}},
  \code{\link{devicec}}, \code{\link{deviceg}},
  \code{\link{fan}}, \code{\link{gaaslaser}},
  \code{\link{grampus}}, \code{\link{grids1}},
  \code{\link{halfbeak}}, \code{\link{heatexchanger}},
  \code{\link{lfp1370}}, \code{\link{lfptrun100}},
  \code{\link{lzbearing}}, \code{\link{metalwear}},
  \code{\link{mylarsub}}, \code{\link{nicdbattery}},
  \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
