% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deviceg}
\alias{deviceg}
\title{Field tracking study with multiple failure modes}
\format{A \code{data.frame} with 30 rows and 2 variables:
\tabular{rlll}{
  [, 1] \tab kilocycles \tab Accumulated cycles at failure (in thousands) \tab \bold{Numeric}\cr
  [, 2] \tab mode \tab Failure mode observed at \code{kcycles} (Wearout/Surge/Suspension) \tab \bold{Categoric}
   }}
\source{
Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons.
}
\description{
Thirty (30) samples of a device that was part of a power generation system were tested as part of a field tracking study 
             for up to 300 kilocycles. The devices under test were installed in typical service environments and at the occurence 
             of a failure, the observed number of cycles and the type of failure were recorded. 
             Surge failures, which predominated early in the device life-cycle, were caused by an accumulation of randomly occurring 
             damage from power-line voltage spikes during electric storms resulting in failure of a particular unprotected electronic 
             component. Wearout failures resulted from normal product wear and began to appear after 100 thousand cycles of use.
}
