% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaaslaser}
\alias{gaaslaser}
\title{Gallium-Arsenic Laser Degradation Test}
\format{A \code{data.frame} with 255 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab increase \tab Percent increase in observed operating current \tab \bold{Numeric}\cr
  [, 2] \tab unit \tab Unit designator code \tab \bold{Categoric}\cr
  [, 3] \tab hours \tab Accumulated time when \code{increase} was measured \tab \bold{Numeric}
   }}
\source{
Meeker W.Q. and Escobar L.A. (1998) Statistical Methods for Reliability Data, New York, NY; John Wiley & Sons.
}
\description{
Over the life of some lasers devices, degradation causes
             a decrease in light output.  In this test, a feedback mechanism was used 
             to maintain a constant light output by increasing the operating current.
             Fifteen GaAs laser devices were tested at an elevated temperature of 
             80^{o} C to accelerate the degradation process.  The operating current
             was measured in 250-hour intervals and the data were recorded as the
             percent increase in current compared to the initial operating current
             measured when the test began.  A device was considered to have failed
             if the percent increase reached 10\%.
}
