% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{printedcircuitboard}
\alias{printedcircuitboard}
\title{Printed circuit board accelerated life test}
\format{A \code{data.frame} with 140 rows and 5 variables:
\tabular{rlll}{
  [, 1] \tab lower \tab Start of an inspection interval (in hours) \tab \bold{Numeric}\cr
  [, 2] \tab upper \tab End of an inspection interval (in hours) \tab \bold{Numeric}\cr
  [, 3] \tab count \tab Number of events observed in the interval \tab \bold{Numeric}\cr
  [, 4] \tab event \tab Event observed in the interval (right-censored/interval-censored) \tab \bold{Categoric}\cr
  [, 5] \tab rh    \tab Relative humidity applied \tab \bold{Numeric}
   }}
\source{
Meeker, W. Q., and LuValle, M. J. (1995), 
        An accelerated life test model based on reliability kinetics, 
        Technometrics, \bold{37}, 133-146.
}
\description{
Meeker and LuValle give data from an accelerated life test on failure of 
             printed circuit boards. The purpose of the experiment was to study the 
             effect of the stresses on the failure-time distribution and to predict 
             reliability under normal operating conditions. More specifically, the 
             experiment was designed to study a particular failure mode - the formation 
             and growth of conductive anodic filaments between copper-plated through-holes 
             in the printed circuit boards.  Actual growth of the filaments could not be 
             monitored, only failure time (defined as a short circuit) could be observed 
             directly.  Special test boards were constructed for the experiment. 
             
             The data provided here are the number of failures observed in each of a series of 
             4-hour and 12-hour long intervals over the life-test period.  This experiment resulted 
             in interval-censored data because only the interval in which each failure occurred was 
             known.  Further, the data are part of the results of a much larger experiment aimed at 
             determining the effects of temperature, relative humidity, and electric field 
             on the reliability of printed circuit boards.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grampus}},
  \code{\link{grids1}}, \code{\link{halfbeak}},
  \code{\link{heatexchanger}}, \code{\link{lfp1370}},
  \code{\link{lfptrun100}}, \code{\link{lzbearing}},
  \code{\link{metalwear}}, \code{\link{mylarsub}},
  \code{\link{nicdbattery}}, \code{\link{piccioto}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{valveseat}}, \code{\link{zelencap}}
}
\concept{data-done}
