\name{SMRD2-package}
\alias{SMRD2-package}
\alias{SMRD2}
\docType{package}
\title{\packageTitle{SMRD2}}
\description{\packageDescription{SMRD2}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{SMRD2}
  \packageIndices{SMRD2}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{SMRD2}
  
  Maintainer: \packageMaintainer{SMRD2}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
