% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRD.R
\docType{package}
\name{SMRD2-package}
\alias{SMRD2}
\alias{SMRD2-package}
\title{SMRD2: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the
      package does as one or more full sentences.
}
\keyword{internal}
