% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tantalum}
\alias{tantalum}
\title{Electrolytic capacitor accelerated life test data}
\format{A \code{data.frame} with 48 rows and 5 variables:
\tabular{rlll}{
  [, 1] \tab hours \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 2] \tab event \tab Event observed at \code{hours} (failure/right-censored) \tab \bold{Categoric}\cr
  [, 3] \tab count \tab Number of events observed at \code{hours} \tab \bold{Numeric}\cr
  [, 4] \tab volts \tab Voltage applied to the unit \tab \bold{Numeric}\cr
  [, 5] \tab celsius \tab Temperature applied to the unit \tab \bold{Numeric}
   }}
\source{
Singpurwalla, N. D., Castellino, V. C., and Goldschen, D. Y. (1975),
        Inference from accelerated life tests using Eyring type re-parameterizations,  
        Naval Research Logistics Quarterly, \bold{22}, 289-296.
}
\description{
Singpurwalla, Castellino, and Goldschen present temperature/voltage accelerated life test data on tantalum 
             electrolytic capacitors. The tests were conducted at temperature/voltage combinations that were nonrectangular 
             and with unequal allocations of units.
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grampus}},
  \code{\link{grids1}}, \code{\link{halfbeak}},
  \code{\link{heatexchanger}}, \code{\link{lfp1370}},
  \code{\link{lfptrun100}}, \code{\link{lzbearing}},
  \code{\link{metalwear}}, \code{\link{mylarsub}},
  \code{\link{nicdbattery}}, \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{turbine}},
  \code{\link{v7tube}}, \code{\link{valveseat}},
  \code{\link{zelencap}}
}
\concept{data-done}
