% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valveseat}
\alias{valveseat}
\title{Diesel engine valve seat data}
\format{A \code{data.frame} with 89 rows and 3 variables:
\tabular{rlll}{
  [, 1] \tab engine \tab Designator code specifying each engine under test \tab \bold{Categoric}\cr
  [, 2] \tab days   \tab Accumulated time at \code{event} \tab \bold{Numeric}\cr
  [, 3] \tab event  \tab Event observed at \code{days} (replacement/end) \tab \bold{Categoric}\cr
  }}
\source{
Nelson, W. and Doganaksoy, N. (1989)  
        A computer program for an estimate and confidence limits for the mean cumulative function for cost or number of repairs of
        repairable products, 
        TIS report 89CRD239, General Electric Company Research and Development, Schenectady, NY.

Nelson, W. (1995) Confidence limits for recurrence data - applied to cost or number of product repairs, 
        Technometrics, \bold{37}, 147-157.
}
\description{
Nelson and Doganaksoy report the age when a valve seat replacement occurred across a fleet of 41 diesel 
             engines. For each engine, the number of days when a replacement event occured were recorded along with the number of 
             days when the observational period ended. The data were recorded to answer the questions (1) Does the replacement rate increased with age?
             (2) How many replacement valves will be needed in some future period of time? and (3) Can valve life in these 
             systems be modeled as a superimposed renewal process?
}
\seealso{
Other data-done: \code{\link{alloya}},
  \code{\link{at7987}}, \code{\link{bearingcage}},
  \code{\link{berkson20}}, \code{\link{bleed}},
  \code{\link{ceramicbearing}}, \code{\link{cirpack6}},
  \code{\link{comptime}}, \code{\link{computerlab}},
  \code{\link{cylinder}}, \code{\link{devicea}},
  \code{\link{deviceb}}, \code{\link{devicec}},
  \code{\link{deviceg}}, \code{\link{fan}},
  \code{\link{gaaslaser}}, \code{\link{grampus}},
  \code{\link{grids1}}, \code{\link{halfbeak}},
  \code{\link{heatexchanger}}, \code{\link{lfp1370}},
  \code{\link{lfptrun100}}, \code{\link{lzbearing}},
  \code{\link{metalwear}}, \code{\link{mylarsub}},
  \code{\link{nicdbattery}}, \code{\link{piccioto}},
  \code{\link{printedcircuitboard}},
  \code{\link{resistor2}}, \code{\link{resistor}},
  \code{\link{superalloy}}, \code{\link{tantalum}},
  \code{\link{turbine}}, \code{\link{v7tube}},
  \code{\link{zelencap}}
}
\concept{data-done}
