% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WQMCPOINTS}
\alias{WQMCPOINTS}
\title{Wrapper for wqm_cpoints}
\usage{
WQMCPOINTS(
  y,
  ny,
  codes,
  codes2,
  weight,
  weight2,
  ty,
  nty,
  tcodes,
  n,
  nstart,
  dscrat,
  scrat,
  iscrat,
  iprint,
  maxit,
  tol,
  maxmsd,
  p,
  q,
  prob,
  sd,
  m,
  pchmax,
  lsd,
  ier,
  ilcv,
  iucv,
  iltv,
  iutv,
  iorder,
  xlcen,
  xrcen,
  fail,
  xltru,
  xrtru,
  ys,
  pgrad,
  s,
  probd
)
}
\description{
Wrapper for wqm_cpoints
}
\details{
Need wrapper to be able to call from 
         debugging mail program
}
